% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adiv_boxplot.r
\name{adiv_boxplot}
\alias{adiv_boxplot}
\title{Visualize alpha diversity with boxplots.}
\usage{
adiv_boxplot(
  biom,
  x = NULL,
  metric = "Shannon",
  layers = "rls",
  color.by = NULL,
  pattern.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  xvals = NULL,
  colors = NULL,
  patterns = NULL,
  shapes = NULL,
  facets = NULL,
  p.adj = "fdr",
  p.label = 0.05,
  ci = 95,
  xlab.angle = "auto",
  safe = FALSE,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{x}{A categorical metadata column name to use for the x-axis. The 
default, \code{NULL} lumps all samples into a single column.}

\item{metric}{Alpha diversity metric(s) to use. Options are: \code{"OTUs"}, 
\code{"Shannon"}, \code{"Chao1"}, \code{"Simpson"}, and/or 
\code{"InvSimpson"}. Default: \code{"Shannon"}.}

\item{layers}{\code{"box"}, \code{"bar" ("r")}, \code{"violin"}, 
\code{"dot"}, \code{"strip"}, \code{"crossbar"}, \code{"errorbar"}, 
\code{"linerange"}, and \code{"pointrange"}. Single letter 
abbreviations are also accepted. For instance, \code{c("box", "dot")} 
is equivalent to \code{c("b", "d")} and \code{"bd"}.
Default: \code{"rls"}.}

\item{color.by, pattern.by, shape.by, facet.by}{Metadata 
column to color, pattern, shape, and/or facet by. If 
that column is a \code{factor}, the ordering of levels will be 
maintained in the plot.}

\item{colors, patterns, shapes, facets, xvals}{Names of the colors, patterns,
shapes, facets, and/or x values to use in the plot. Available values 
for colors, patterns, and shapes are given by \code{colors()}, 
\code{gridpattern::names_magick}, and \code{0:25},
respectively. Use a named character vector to map them to specific
factor levels in the metadata. \code{facets} and \code{xvals} are
coerced to unnamed character vectors. If the length of these vectors
is less than the values present in their corresponding metadata 
column, then the data set will be subseted accordingly.}

\item{p.adj}{Method to use for multiple comparisons adjustment of p-values.
Run \code{p.adjust.methods} for a list of available options.
(Default: \code{fdr})}

\item{p.label}{Minimum adjusted p-value to display on the plot with a bracket.
Set to \code{Inf} to display all p-values, or \code{-Inf} for no brackets.
If a numeric vector with more than one value is provided, they will be
used as breaks for asterisk notation. For ordinations, \code{p.label} 
applies to biplot taxa. (Default: \code{0.05})}

\item{ci}{How to calculate min/max of the \bold{crossbar}, 
\bold{errorbar}, \bold{linerange}, and \bold{pointrange} layers.
Provide a number between 75 and 100 to define a confidence interval's
confidence level, commonly 95 or 97.5. Other options are: 
\bold{range}, 
\bold{sd} (standard deviation), 
\bold{se} (standard error), and 
\bold{mad} (median absolute deviation). 
The center mark of \code{crossbar} and \code{pointrange} represents
the mean, except for \bold{mad} in which case it represents
the median. Trendlines require a confidence interval value. 
Set to \code{NULL} to disable. Default: \code{95}}

\item{xlab.angle}{How to rotate the tick labels on the x-axis. 
\bold{'auto'} (the default), automatically selects a rotation value. 
\bold{0}, \bold{30}, and \bold{90} sets the angle to horizontal, 
angled, and vertical, respectively.}

\item{safe}{If \code{FALSE}, data.frame column names such as 
\code{".metric"} will be auto-converted to \code{"Metric"} to improve
human-readability. Conversion if aborted if a conflict is found with
a metadata column name. (Default: \code{FALSE})}

\item{...}{Parameters are matched to formal arguments of ggplot2
functions. Prefixing parameter names with a layer name ensures that
a particular parameter is passed to, and only to, that layer. For
instance, \code{dot.size = 2} or \code{d.size = 2} ensures only the 
dotplot layer has its size set to \code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
        
Shapes can also be given as their string values, defined in pch_table here:
https://github.com/tidyverse/ggplot2/blob/master/R/geom-point.r . Note that
some shapes have a colored outline given by `color`, some are filled with 
`color`, and some are outlined in `color` and filled with `fill`. See
https://blog.albertkuo.me/post/point-shape-options-in-ggplot/ for details.
}
\description{
Visualize alpha diversity with boxplots.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    
    adiv_boxplot(biom, x = "Body Site", metric = "Shannon")
    adiv_boxplot(biom, x = "Sex", metric = c("OTUs", "Shannon"), layers="b", color.by="Body Site", scales="free")
    adiv_boxplot(biom, x = "Body Site", metric = "Simpson", layers="p", color.by="Sex", xlab.angle=30)
    
}
\seealso{
\code{\link{stats_table}}

Other plotting: 
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()}
}
\concept{plotting}
