% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adiv_table.r
\name{adiv_matrix}
\alias{adiv_matrix}
\title{Create a matrix of samples x alpha diversity metrics.}
\usage{
adiv_matrix(biom, rarefy = FALSE)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{rarefy}{Control how/whether rarefactions are done prior to alpha
diversity computations. Options are:
\itemize{
\item{\code{FALSE} - }{
Use each sample's current set of observations without applying
any rarefaction.}
\item{\code{TRUE} - }{
Automatically select and apply a single rarefaction. }
\item{\emph{integer} - }{
Rarefy to the specified depth. }
}
Default: \code{FALSE}}
}
\value{
A numeric matrix with samples as rows and columns named
\bold{Depth}, \bold{OTUs}, \bold{Shannon}, \bold{Chao1},
\bold{Simpson}, and \bold{InvSimpson}.
}
\description{
Create a matrix of samples x alpha diversity metrics.
}
\examples{
    library(rbiom) 
    
    biom <- sample_select(hmp50, 1:6)
    adiv_matrix(biom)
}
