% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adiv_stats.r
\name{adiv_stats}
\alias{adiv_stats}
\title{Test alpha diversity differences for significance.}
\usage{
adiv_stats(
  biom,
  stat.by,
  adiv = "Shannon",
  split.by = NULL,
  trans = "none",
  test = "means",
  p.adj = "fdr"
)
}
\arguments{
\item{biom}{An rbiom-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{stat.by}{The categorical metadata field defining the statistical
groups. Required.}

\item{adiv}{Alpha diversity metric(s) to use. Options are: \code{"OTUs"},
\code{"Shannon"}, \code{"Chao1"}, \code{"Simpson"}, and/or
\code{"InvSimpson"}. Set \code{adiv=".all"} to use all metrics.
Default: \code{"Shannon"} \cr\cr
Multiple values allowed. Non-ambiguous abbreviations are allowed.}

\item{split.by}{Name(s) of metadata columns that the data should be split
by prior to any calculations. Default: \code{NULL}}

\item{trans}{Transformation to apply. Options are:
\code{c("none", "rank", "log", "log1p", "sqrt")}. \code{"rank"} is useful for
correcting for non-normally distributions before applying regression
statistics. Default: \code{"none"}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}}
}
\value{
A tibble data frame with summary statistics. \cr
The R code or generating these statistics is in \verb{$code}.
}
\description{
Test alpha diversity differences for significance.
}
\examples{
    library(rbiom) 
    
    biom <- rarefy(hmp50)
      
    adiv_stats(biom, stat.by = "Body Site")
      
    adiv_stats(biom, stat.by = "Sex", test = "pw_means")
    
    adiv_stats(biom, stat.by = "Body Site", split.by = "Sex")
      
}
\seealso{
Other alpha_diversity: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{adiv_table}()}

Other stats_tables: 
\code{\link{bdiv_stats}()},
\code{\link{distmat_stats}()},
\code{\link{taxa_stats}()}
}
\concept{alpha_diversity}
\concept{stats_tables}
