% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.div.r
\name{alpha.div}
\alias{alpha.div}
\title{Estimate the diversity of each sample.}
\usage{
alpha.div(biom, progressbar = FALSE)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM} 
object, as returned from \link{read.biom}. For matrices, the rows and 
columns are assumed to be the taxa and samples, respectively.}

\item{progressbar}{Whether to display a progress bar and status messages
(TRUE/FALSE). Will automatically tie in with \pkg{shiny} if run within
a \pkg{shiny} session.}
}
\value{
A numeric matrix of four diversity values for each sample in 
    \code{biom}. Row names are the sample names and column names are the 
    diversity metrics: \bold{OTUs}, \bold{Shannon}, \bold{Simpson}, and 
    \bold{Chao1}.
}
\description{
Estimate the diversity of each sample.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    
    ad <- alpha.div(biom)
    head(ad)

}
