% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apcoa.r
\name{apcoa}
\alias{apcoa}
\title{PCoA adjusted for corvariates.}
\usage{
apcoa(distmat, covariates)
}
\arguments{
\item{distmat}{A distance matrix (\code{dist} class object) that you want
to run the aPCoA on.}

\item{covariates}{A data.frame with the confounding covariate(s). The row 
names of this data frame should match the labels of the distance matrix.}
}
\value{
A numeric matrix with the same row names as \code{covariates}, and
     one column for each of the computed adjusted principal coordinates.
}
\description{
Please cite: Shi Y, Zhang L, Do KA, Peterson CB, Jenq RR. aPCoA: covariate
adjusted principal coordinates analysis. Bioinformatics. 2020 Jul 
1;36(13):4099-4101. doi: 10.1093/bioinformatics/btaa276.
}
\examples{
    library(rbiom)
    library(ggplot2)
    
    dm <- unifrac(hmp50)
    reg_pcoa <- ape::pcoa(dm)[['vectors']]
    adj_pcoa <- apcoa(dm, metadata(biom)[,'Sex',drop=FALSE])
    
    ids   <- sample.names(hmp50)
    color <- metadata(hmp50, 'Sex')
    ggplot(mapping=aes(x=reg_pcoa[ids, 1], y=reg_pcoa[ids, 2], color=color)) + geom_point()
    ggplot(mapping=aes(x=adj_pcoa[ids, 1], y=adj_pcoa[ids, 2], color=color)) + geom_point()
    
}
