% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.r
\name{as.matrix.rbiom}
\alias{as.matrix.rbiom}
\title{Convert an rbiom object to a simple count matrix.}
\usage{
\method{as.matrix}{rbiom}(x, ...)
}
\arguments{
\item{x}{An \link[=rbiom_objects]{rbiom object}, such as from \code{\link[=as_rbiom]{as_rbiom()}}.}

\item{...}{Not used.}
}
\value{
A base R matrix with OTUs as rows and samples as columns.
}
\description{
Identical to running \code{as.matrix(biom$counts)}.
}
\examples{
    library(rbiom)
    
    as.matrix(hmp50)[1:5,1:5]

}
\seealso{
Other conversion: 
\code{\link{as.list.rbiom}()}
}
\concept{conversion}
