% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.percent.r
\name{as.percent}
\alias{as.percent}
\title{Convert absolute counts to relative abundances.}
\usage{
as.percent(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
A \code{BIOM} object.
}
\description{
Convert absolute counts to relative abundances.
}
\examples{
    library(rbiom)
    
    counts(hmp50)[1:4,1:5]
    
    biom <- as.percent(hmp50)
    counts(biom)[1:4,1:5]

}
