% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_percent.r
\name{as_percent}
\alias{as_percent}
\title{Convert absolute counts to relative abundances.}
\usage{
as_percent(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}
}
\value{
A \code{BIOM} object.
}
\description{
Convert absolute counts to relative abundances.
}
\examples{
    library(rbiom)
    
    otu_matrix(hmp50)[1:4,1:5]
    
    biom <- as_percent(hmp50)
    otu_matrix(biom)[1:4,1:5]

}
