% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{babies}
\alias{babies}
\title{Longitudinal Stool Samples from Infants (n = 2,684)}
\format{
An rbiom object with 2,684 samples.
Includes metadata and taxonomy.
\describe{
\item{Subject ID - }{ID1, ID2, ..., ID12}
\item{Sex - }{Male or Female}
\item{Age (days) - }{1 - 266}
\item{Child's diet - }{"Breast milk", "Breast milk and formula", or "Formula"}
\item{Sample collection - }{"Frozen upon collection" or "Stored in alcohol"}
\item{Antibiotic exposure - }{Yes or No}
\item{Antifungal exposure - }{Yes or No}
\item{Delivery mode - }{Cesarean or Vaginal}
\item{Solid food introduced (Age) - }{116 - 247}
}
}
\source{
\url{https://www.nature.com/articles/s41467-018-04641-7} and \doi{10.1038/s41467-017-01973-8}
}
\usage{
babies
}
\description{
Longitudinal Stool Samples from Infants (n = 2,684)
}
\keyword{datasets}
