% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_biplot.r
\name{bdiv_biplot}
\alias{bdiv_biplot}
\title{Display distances between samples as a 2-D scatter plot.}
\usage{
bdiv_biplot(
  biom,
  dist = "Bray-Curtis",
  ord = "PCoA",
  layers = NULL,
  color.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  colors = NULL,
  shapes = NULL,
  facets = NULL,
  weighted = TRUE,
  rank = NULL,
  taxa = 5,
  p.top = Inf,
  p.adj = "fdr",
  perms = 1000,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{dist}{Beta diversity metric to use for calculating inter-sample
distances. Options are: \code{"Bray-Curtis"}, \code{"Manhattan"},
\code{"Euclidean"}, \code{"Jaccard"}, or \code{"UniFrac"}.
Default: \code{"Bray-Curtis"}.}

\item{ord}{Ordination method to use for condensing n-dimensional data to
two dimensions. Options are: \code{"UMAP"}, \code{"tSNE"},
\code{"NMDS"}, or \code{"PCoA"}. Default: \code{"PCoA"}.}

\item{layers}{What graphical elements to use for drawing. Options are:
\bold{point}, \bold{spider}, \bold{ellipse}, and \bold{name} for
samples, and \bold{mean}, \bold{taxon}, and \bold{arrow} for biplots.
Single letter abbreviations are also accepted. For instance,
\code{c("point", "ellipse")} is equivalent to \code{c("p", "e")} and 
\code{"pe"}. Default: \code{"pce"}/\code{"p"} for ordinations 
with/without a \code{color.by} argument.}

\item{color.by, shape.by, facet.by}{Metadata column to color, shape, and/or
facet by. If that column is a \code{factor}, the ordering of levels
will be maintained in the plot.}

\item{colors, shapes, facets}{Names of the colors, shapes, and/or facets to 
use in the plot. Available values for colors and shapes are given by 
\code{colors()} and \code{0:25}, respectively. Use a named character 
vector to map them  to specific factor levels in the metadata.}

\item{weighted}{When employing a beta diversity metric, use the weighted
version. Default: \code{TRUE}.}

\item{rank}{What rank of taxa to display, for example \code{"Phylum"} or 
\code{"Genus"}. Run \code{taxa_ranks()} to see all options for a 
given BIOM object. The default, \code{NULL}, selects the lowest
level.}

\item{taxa}{Which taxa to display. An integer value will show the top n
most abundant taxa. A value 0 <= n < 1 will show any taxa with that 
mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. Default: \code{5}.}

\item{p.top}{Only display taxa with the most significant differences in
abundance. If \code{p.top} is >= 1, then the \code{p.top} most 
significant taxa are displayed. If \code{p.top} is less than one, all
taxa with an adjusted p-value <= \code{p.top} are displayed.
Recommended to be used in combination with the \code{taxa} parameter
to set a lower bound on the mean abundance of considered taxa.
Default: \code{Inf}.}

\item{p.adj}{Method to use for multiple comparisons adjustment of p-values.
Run \code{p.adjust.methods} for a list of available options.
Default: \code{"fdr"}.}

\item{perms}{Number of random permutations to use for estimating statistical
significance. Default: \code{1000}.}

\item{...}{Parameters for underlying functions. Prefixing parameter names with a layer name ensures that
a particular parameter is passed to, and only to, that layer. For
instance, \code{dot.size = 2} or \code{d.size = 2} ensures only the 
dotplot layer has its size set to \code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
}
\description{
Display distances between samples as a 2-D scatter plot.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    bdiv_biplot(biom, color.by="Body Site", rank="Genus")
    
}
\seealso{
\code{\link{stats_table}}

Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()}
}
\concept{plotting}
