% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_corrplot.r
\name{bdiv_corrplot}
\alias{bdiv_corrplot}
\title{Visualize beta diversity with scatterplots and trendlines.}
\usage{
bdiv_corrplot(
  biom,
  x,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  layers = "tc",
  color.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  within = NULL,
  between = NULL,
  tree = NULL,
  formula = y ~ x,
  engine = "lm",
  level = 0.95,
  trans = "none",
  caption = TRUE,
  ...
)
}
\arguments{
\item{biom}{An rbiom-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{x}{A numeric metadata field to use for the x-axis. Required.}

\item{bdiv}{Beta diversity distance algorithm(s) to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. For \code{"UniFrac"}, a
phylogenetic tree must be present in \code{biom} or explicitly
provided via \verb{tree=}. Default: \code{"Bray-Curtis"} \cr\cr
Multiple/abbreviated values allowed.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Default: \code{TRUE} \cr\cr
Multiple values allowed.}

\item{layers}{One or more of
\code{c("trend", "confidence", "scatter", "name")}. Single
letter abbreviations are also accepted. For instance,
\code{c("trend", "scatter")} is equivalent to \code{c("t", "s")} and \code{"ts"}.
See \link[=plots]{plot types} for examples of each. Default: \code{"tc"}}

\item{color.by, facet.by, limit.by}{Metadata columns to use for aesthetics
and partitioning. See below for details. Default: \code{NULL}}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Only required when
computing UniFrac distances. Default: \code{biom$tree}}

\item{formula}{Relationship between variables. Default: \code{y ~ x}}

\item{engine}{What type of trend model to fit to the data. Options are:
\code{"lm"} (linear), \code{"local"} (uses loess or gam). Default: \code{"lm"}}

\item{level}{The confidence level for calculating a confidence interval.
Default: \code{0.95}}

\item{trans}{Transformation to apply. Options are:
\code{c("none", "rank", "log", "log1p", "sqrt")}. \code{"rank"} is useful for
correcting for non-normally distributions before applying regression
statistics. Default: \code{"none"}}

\item{caption}{Add methodology caption beneath the plot.
Default: \code{TRUE}}

\item{...}{Additional parameters to pass along to ggplot2
functions. Prefix a parameter name with either \code{t.} or
\code{s.}/\code{pt.} to ensure it gets passed to (and only to)
\link[ggplot2]{geom_smooth} or \link[ggplot2]{geom_point},
respectively. For instance, \code{s.size = 2} ensures only the
scatterplot points have their size set to \code{2}.}
}
\value{
A \code{ggplot2} plot. \cr The computed data points and ggplot command
are available as \verb{$data} and \verb{$code}, respectively.
}
\description{
Visualize beta diversity with scatterplots and trendlines.
}
\section{Aesthetics and Partitions}{


Metadata can be used to flexibly subset, partition, and apply aesthetics
when creating a plot. Common use cases are provided below. More thorough
documentation is available at \url{https://cmmr.github.io/rbiom}.

\preformatted{  ## Colors ----------------------------
  color.by = "Body Site"
  color.by = list('Body Site' = "bright")
  color.by = list('Body Site' = c("Stool", "Saliva"))
  color.by = list('Body Site' = list('values' = c("Stool", "Saliva"), 'colors' = "bright"))
  color.by = list('Body Site' = c('Stool' = "blue", 'Saliva' = "green"))
  
  ## Facets ----------------------------
  facet.by = "Body Site"
  facet.by = c("Body Site", "Sex")
  facet.by = list('Body Site' = c("Stool", "Saliva"), "Sex")
  
  ## Limits ----------------------------
  limit.by = list('Sex' = "Male", 'Age' = c(20,40))
  limit.by = list('Body Site' = c("Saliva", "Anterior nares"), 'Age' = c(NA,35))
}

\itemize{
\item{\code{color.by} - }{Any metadata column. (Max 1)}
\item{\code{facet.by} - }{Only categorical metadata column(s).}
\item{\code{limit.by} - }{Any metadata column(s).}
}

All built-in color palettes are colorblind-friendly.

The available categorical palette names are: \code{"okabe"}, \code{"carto"},
\code{"r4"}, \code{"polychrome"}, \code{"tol"}, \code{"bright"},
\code{"light"}, \code{"muted"}, \code{"vibrant"}, \code{"tableau"},
\code{"classic"}, \code{"alphabet"}, \code{"tableau20"}, \code{"kelly"},
and \code{"fishy"}.

The available numeric palette names are: \code{"reds"}, \code{"oranges"},
\code{"greens"}, \code{"purples"}, \code{"grays"}, \code{"acton"},
\code{"bamako"}, \code{"batlow"}, \code{"bilbao"}, \code{"buda"},
\code{"davos"}, \code{"devon"}, \code{"grayC"}, \code{"hawaii"},
\code{"imola"}, \code{"lajolla"}, \code{"lapaz"}, \code{"nuuk"},
\code{"oslo"}, \code{"tokyo"}, \code{"turku"}, \code{"bam"},
\code{"berlin"}, \code{"broc"}, \code{"cork"}, \code{"lisbon"},
\code{"roma"}, \code{"tofino"}, \code{"vanimo"}, \code{"vik"}
}

\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    bdiv_corrplot(biom, "Age", color.by = "==Body Site", facet.by = "==Sex")
    bdiv_corrplot(biom, "Age", color.by = list('==Body Site' = c("a", "b")), layers = "tcs")
    
}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_stats}()},
\code{\link{bdiv_table}()},
\code{\link{distmat_stats}()}

Other visualization: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{rare_stacked}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()}
}
\concept{beta_diversity}
\concept{visualization}
