% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_dist.r
\name{bdiv_distmat}
\alias{bdiv_distmat}
\title{Make a distance matrix of samples vs samples.}
\usage{
bdiv_distmat(
  biom,
  method = "bray-curtis",
  weighted = TRUE,
  tree = NULL,
  stat.by = NULL,
  seed = 0,
  perms = 999
)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM} 
object, as returned from \link{read_biom}. For matrices, the rows and 
columns are assumed to be the taxa and samples, respectively.}

\item{method}{The distance algorithm to use. Options are:
\code{"bray-curtis"}, \code{"manhattan"}, \code{"euclidean"}, 
\code{"jaccard"}, and \code{"unifrac"}. Non-ambiguous abbreviations 
of the method names are also accepted. A phylogenetic tree must be 
present in \code{biom} or explicitly provided via \code{tree=} to use
the UniFrac methods. (Default: \code{"bray-curtis"})}

\item{weighted}{Take relative abundances into account. When 
\code{weighted=FALSE}, only presence/absence is considered.
 (Default: \code{TRUE})}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Will be taken from the tree
embedded in the \code{biom} object if not explicitly specified. Only
required for computing UniFrac distance matrices.
 (Default: \code{NULL})}

\item{stat.by}{Specify a categorical metadata column name to compute adonis 
statistics and return results as attributes named 'stats_raw' and 
'stats_tbl'. (Default: \code{NULL})}

\item{seed}{Random seed for adonis permutations. (Default: \code{0})}

\item{perms}{Number of random permutations to use in adonis calcuation.
(Default: \code{999})}
}
\value{
A distance matrix.
}
\description{
Make a distance matrix of samples vs samples.
}
\examples{
    library(rbiom)
    
    biom <- select(hmp50, 1:10)
    dm <- bdiv_distmat(biom, 'unifrac')
    as.matrix(dm)[1:4,1:4]
    plot(hclust(dm))

}
