% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_ord_plot.r
\name{bdiv_ord_plot}
\alias{bdiv_ord_plot}
\title{Ordinate samples and taxa on a 2D plane based on beta diversity distances.}
\usage{
bdiv_ord_plot(
  biom,
  bdiv = "Bray-Curtis",
  ord = "UMAP",
  weighted = TRUE,
  layers = "pe",
  color.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  tree = NULL,
  test = "adonis2",
  seed = 0,
  permutations = 999,
  rank = NULL,
  taxa = 5,
  p.top = Inf,
  p.adj = "fdr",
  unc = "singly",
  ...
)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.
Alternatively, a {dist} class distance matrix can be given, in which
case only the parameters \code{ord}, \code{k}, and \code{...} are
allowed.}

\item{bdiv}{The beta diversity distance algorithm to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. A phylogenetic tree must be
present in \code{biom} or explicitly provided via \code{tree=} to use
the UniFrac methods.
Multiple values allowed. Default: \code{"Bray-Curtis"}.}

\item{ord}{Method for reducing dimensionality. Options are:
\itemize{
\item{\code{"UMAP"} - }{ Uniform manifold approximation and projection; \code{\link[uwot:umap]{uwot::umap()}}. }
\item{\code{"PCoA"} - }{ Principal coordinate analysis; \code{\link[ape:pcoa]{ape::pcoa()}}. }
\item{\code{"NMDS"} - }{ Nonmetric multidimensional scaling; \code{\link[vegan:metaMDS]{vegan::metaMDS()}}. }
\item{\code{"tSNE"} - }{ t-distributed stochastic neighbor embedding; \code{\link[tsne:tsne]{tsne::tsne()}}. }
}
Default: \code{"UMAP"} \cr\cr
Multiple values allowed. Non-ambiguous abbreviations are allowed.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Multiple values allowed. Default: \code{TRUE}}

\item{layers}{What graphical elements to use for drawing. Options are:
\bold{point}, \bold{spider}, \bold{ellipse}, and \bold{name} for
samples, and \bold{mean}, \bold{taxon}, and \bold{arrow} for taxa.
Single letter abbreviations are also accepted. For instance,
\code{c("point", "ellipse")} is equivalent to \code{c("p", "e")} and
\code{"pe"}.
See the \code{vignette("ordination")} vignette for examples of each.
Default: \code{"pe"}}

\item{color.by, shape.by, facet.by, limit.by}{Metadata columns to
use for \code{vignette("data partitioning")}. Default:
\code{color.by=NULL, shape.by=NULL, facet.by=NULL, limit.by=NULL}}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Will be taken from the tree
defined in \code{biom} if not explicitly specified. Only required
when computing UniFrac distance matrices. Default: \code{NULL}}

\item{test, seed, permutations}{Passthrough parameters to \code{\link[=distmat_stats]{distmat_stats()}}
for assessing significance of \code{stat.by} and \code{rank} groups.
Default: \code{test="adonis2", seed=0, permutations=999}}

\item{rank}{What rank of taxa to display, for example \code{"Phylum"} or
\code{"Genus"}. Use \code{taxa_ranks()} to see all options for a
given BIOM object. The default, \code{NULL}, selects the lowest
level.}

\item{taxa}{\link{Biplot} Which taxa to include. An integer value will return
the top n most abundant taxa. A value 0 <= n < 1 will return any taxa
with that mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. Default: \code{5}}

\item{p.top}{\link{Biplot} Only return taxa with the most significant
differences in abundance. If \code{p.top} is >= 1, then the
\code{p.top} most significant taxa are displayed. If \code{p.top}
is less than one, all taxa with an adjusted p-value <= \code{p.top}
are displayed. Recommended to be used in combination with the
\code{taxa} parameter to set a lower bound on the mean abundance of
considered taxa. Default: \code{Inf}}

\item{p.adj}{\link{Biplot} Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available options.
Default: \code{fdr}}

\item{unc}{\link{Biplot} How to handle unclassified, uncultured, and similarly
ambiguous taxa names. Options are:
\itemize{
\item{\code{"singly"} - }{ Replaces them with the OTU name. }
\item{\code{"grouped"} - }{ Replaces them with a higher rank's name. }
\item{\code{"drop"} - }{ Excludes them from the result. }
\item{\code{"asis"} - }{ To not check/modify any taxa names. }
}
Default: \code{"singly"} \cr\cr
Non-ambiguous abbreviations are allowed.}

\item{...}{Parameters for layer geoms (e.g. \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}).
Prefixing parameter names with a layer name ensures that a particular
parameter is passed to, and only to, that layer. For instance,
\code{point.size = 2} or \code{p.size = 2} ensures only the points
have their size set to \code{2}. Points can also be controlled with
the \code{pt.} prefix.}
}
\value{
A \code{ggplot} object.

\if{html}{\out{<div class="sourceCode">}}\preformatted{    The computed sample coordinates are in \code{p$data)}.
    
    If \code{color.by} is given, then \code{attr(p, 'stats')} and 
    \code{attr(p, 'stats_cmds')} are set.
    
    If \code{rank} is given, then \code{attr(p$data, 'taxa_coords')}, 
    \code{attr(p, 'taxa_stats')}, and \code{attr(p, 'taxa_stats_cmds')} 
    are set.
}\if{html}{\out{</div>}}
}
\description{
Ordinate samples and taxa on a 2D plane based on beta diversity distances.
}
\examples{
    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    bdiv_ord_plot(biom, layers="pemt", color.by="Body Site", rank="g")
    
}
\seealso{
Other ordination: 
\code{\link{bdiv_ord_table}()}

Other beta_diversity: 
\code{\link{bdiv_distmat}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_table}()},
\code{\link{distmat_stats}()}

Other visualization: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_barplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()}
}
\concept{beta_diversity}
\concept{ordination}
\concept{visualization}
