% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_stats.r
\name{bdiv_stats}
\alias{bdiv_stats}
\title{Test beta diversity vs categorical or numeric metadata.}
\usage{
bdiv_stats(
  biom,
  stat.by = NULL,
  regr = NULL,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  within = NULL,
  between = NULL,
  test = ifelse(is.null(regr), "means", "trends"),
  model = "lm",
  trans = ifelse(is.null(regr), "none", "rank"),
  split.by = NULL,
  level = 0.95,
  p.adj = "fdr"
)
}
\arguments{
\item{biom}{An rbiom-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{stat.by}{The categorical or numeric metadata field over which
statistics should be calculated. Required.}

\item{regr}{To run a regression analysis, set \code{regr} to the numeric
metadata field with the "x-axis" values. Leaving \code{regr=NULL}
will generate boxplot-like statistics; when non-NULL, corrplot-like
statistics will be returned. Default: \code{NULL}}

\item{bdiv}{Beta diversity distance algorithm(s) to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. For \code{"UniFrac"}, a
phylogenetic tree must be present in \code{biom} or explicitly
provided via \verb{tree=}. Default: \code{"Bray-Curtis"} \cr\cr
Multiple/abbreviated values allowed.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Default: \code{TRUE} \cr\cr
Multiple values allowed.}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Only required when
computing UniFrac distances. Default: \code{biom$tree}}

\item{within, between}{Metadata field(s) for intra- or inter- sample
comparisons. Default: \code{NULL}}

\item{test}{Permutational test for accessing significance. Options are:
\itemize{
\item{\code{"adonis2"} - }{ Permutational MANOVA; \code{\link[vegan:adonis]{vegan::adonis2()}}. }
\item{\code{"mrpp"} - }{ Multiple response permutation procedure; \code{\link[vegan:mrpp]{vegan::mrpp()}}. }
\item{\code{"none"} - }{ Don't run any statistics. }
}
Default: \code{"adonis2"} \cr\cr
Abbreviations are allowed.}

\item{trans}{Transformation to apply. Options are:
\code{c("none", "rank", "log", "log1p", "sqrt")}. \code{"rank"} is useful for
correcting for non-normally distributions before applying regression
statistics. Default: \code{"none"}}

\item{split.by}{Name(s) of metadata columns that the data should be split
by prior to any calculations. Default: \code{NULL}}

\item{level}{The confidence level for calculating a confidence interval.
Default: \code{0.95}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}}
}
\value{
A tibble data frame with summary statistics. \cr
The R code or generating these statistics is in \verb{$code}.
}
\description{
Test beta diversity vs categorical or numeric metadata.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    
    bdiv_stats(biom, stat.by = "Body Site", split.by = "Sex")
      
    bdiv_stats(biom, stat.by = "Sex", bdiv = c("bray", "unifrac"))
    
    bdiv_stats(biom, regr = "BMI", stat.by = "==Body Site")
    
    # The R code used to compute the stats is in $code.
    tbl <- bdiv_stats(biom, stat.by = "Sex")
    tbl$code
}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_corrplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_table}()},
\code{\link{distmat_stats}()}

Other stats_tables: 
\code{\link{adiv_stats}()},
\code{\link{distmat_stats}()},
\code{\link{stats_table}()},
\code{\link{taxa_stats}()}
}
\concept{beta_diversity}
\concept{stats_tables}
