% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_table.r
\name{bdiv_table}
\alias{bdiv_table}
\title{Make a data.frame of distances between samples.}
\usage{
bdiv_table(
  biom,
  method = "bray-curtis",
  weighted = TRUE,
  tree = NULL,
  md = FALSE,
  stat.by = NULL,
  seed = 0,
  perms = 999
)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM} 
object, as returned from \link{read_biom}. For matrices, the rows and 
columns are assumed to be the taxa and samples, respectively.}

\item{method}{The distance algorithm to use. Options are:
\code{"bray-curtis"}, \code{"manhattan"}, \code{"euclidean"}, 
\code{"jaccard"}, and \code{"unifrac"}. Non-ambiguous abbreviations 
of the method names are also accepted. A phylogenetic tree must be 
present in \code{biom} or explicitly provided via \code{tree=} to use
the UniFrac methods. (Default: \code{"bray-curtis"})}

\item{weighted}{Take relative abundances into account. When 
\code{weighted=FALSE}, only presence/absence is considered.
 (Default: \code{TRUE})}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Will be taken from the tree
embedded in the \code{biom} object if not explicitly specified. Only
required for computing UniFrac distance matrices.
 (Default: \code{NULL})}

\item{md}{Include metadata in the output data frame? Options are: 
\describe{
  \item{\code{FALSE}}{ Don't include metadata. (Default) }
  \item{\code{TRUE}}{ Include all metadata. }
  \item{\emph{character vector}}{ Include only the specified metadata 
  columns. Column names can be prefixed with \bold{==} or \bold{!=} to 
  indicate that only within or between groupings, respectively, are to 
  be kept. See examples below. }
}}

\item{stat.by}{Specify a categorical metadata column name to compute adonis 
statistics and return results as attributes named 'stats_raw' and 
'stats_tbl'. (Default: \code{NULL})}

\item{seed}{Random seed for adonis permutations. (Default: \code{0})}

\item{perms}{Number of random permutations to use in adonis calcuation.
(Default: \code{999})}
}
\value{
A data.frame with first three columns named ".sample1", ".sample2", 
        and ".distance".
}
\description{
Make a data.frame of distances between samples.
}
\examples{
    library(rbiom)
    
    # Return in long format with metadata
    biom <- select(hmp50, 18:21)
    bdiv_table(biom, 'unifrac', md = c("Body Site", "Sex"))
    
    # Only look at distances among the stool sample
    bdiv_table(biom, 'unifrac', md = c("==Body Site", "Sex"))
    
    # Or between males and females
    bdiv_table(biom, 'unifrac', md = c("Body Site", "!=Sex"))

}
