% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.div.r
\name{beta.div}
\alias{beta.div}
\title{Make a distance matrix of samples vs samples.}
\usage{
beta.div(
  biom,
  method,
  weighted = TRUE,
  tree = NULL,
  long = FALSE,
  md = FALSE,
  safe = FALSE,
  stat.by = NULL,
  seed = 0,
  perms = 999
)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM} 
object, as returned from \link{read.biom}. For matrices, the rows and 
columns are assumed to be the taxa and samples, respectively.}

\item{method}{The distance algorithm to use. Options are:
\bold{manhattan}, \bold{euclidean}, 
\bold{bray-curtis}, \bold{jaccard}, and
\bold{unifrac}. Non-ambiguous abbreviations of the method 
names are also accepted. A phylogenetic tree must be present in 
\code{biom} or explicitly provided via \code{tree=} to use the UniFrac 
methods.}

\item{weighted}{Take relative abundances into account. When 
\code{weighted=FALSE}, only presence/absence is considered.}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Will be taken from the tree
embedded in the \code{biom} object if not explicitly specified. Only
required for computing UniFrac distance matrices.}

\item{long}{Pivot the returned data to long format?
\describe{
  \item{\bold{FALSE}}{ Return a \bold{dist} object. (Default) }
  \item{\bold{TRUE}}{  A \bold{data.frame} with columns "Sample1", 
                       "Sample2" and "Distance" is returned. }
}}

\item{md}{Include metadata in the output data frame? Options are: 
\describe{
  \item{\code{FALSE}}{ Don't include metadata. (Default) }
  \item{\code{TRUE}}{ Include all metadata. }
  \item{\emph{character vector}}{ Include only the specified metadata 
  columns. Column names can be prefixed with \bold{==} or \bold{!=} to 
  indicate that only within or between groupings, respectively, are to 
  be kept. See examples below. }
}}

\item{safe}{Should autogenerated columns be prefixed with a "." to avoid 
conflicting with metadata column names? (Default: \code{FALSE})}

\item{stat.by}{Compute adonis statistics and return results as an attribute 
named 'stats'. (Default: \code{NULL})}

\item{seed}{Random seed for adonis permutations. (Default: 0)}

\item{perms}{Number of random permutations to use in adonis calcuation (Default: 999)}
}
\value{
If both \code{long} and \code{md} are \bold{FALSE}, returns a distance 
        matrix, otherwise a data.frame.
}
\description{
Make a distance matrix of samples vs samples.
}
\examples{
    library(rbiom)
    
    biom <- select(hmp50, 1:10)
    dm <- beta.div(biom, 'unifrac')
    as.matrix(dm)[1:4,1:4]
    plot(hclust(dm))
    
    # Return in long format with metadata
    biom <- select(hmp50, 18:21)
    beta.div(biom, 'unifrac', md = c("Body Site", "Sex"))
    
    # Only look at distances among the stool sample
    beta.div(biom, 'unifrac', md = c("==Body Site", "Sex"))
    
    # Or between males and females
    beta.div(biom, 'unifrac', md = c("Body Site", "!=Sex"))

}
