% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{biom_id}
\alias{biom_id}
\alias{biom_id<-}
\alias{biom_comment}
\alias{biom_comment<-}
\title{Get or set a \code{BIOM} object's id or comment.}
\usage{
biom_id(biom)

biom_id(x) <- value

biom_comment(biom)

biom_comment(x) <- value
}
\arguments{
\item{biom, x}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{value}{The identifier to add (character vector of length 1).}
}
\value{
A length 1 character vector.
}
\description{
The BIOM specification includes \code{id} and \code{comment} fields
for free-form text.
}
\examples{
    library(rbiom) 
    biom <- hmp50
    
    biom_id(biom)
    biom_id(biom) <- "My new title/id"
    biom_id(biom)
    
    biom_comment(biom)
    biom_comment(biom) <- "A description of this study"
    biom_comment(biom)

}
\seealso{
Other setters: 
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}

Other setters: 
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
