% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{biom_info}
\alias{biom_info}
\title{Get \code{BIOM} object's miscellaneous information.}
\usage{
biom_info(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}
}
\value{
A list of the top-level metadata in \code{biom}.
}
\description{
Get \code{BIOM} object's miscellaneous information.
}
\examples{
    library(rbiom)
    biom_info(hmp50)

}
\seealso{
Other accessors: 
\code{\link{n_samples}()},
\code{\link{sample_sums}()}
}
\concept{accessors}
