% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biom_build.r
\name{biom_merge}
\alias{biom_merge}
\title{Combine several rbiom objects into one.}
\usage{
biom_merge(...)
}
\arguments{
\item{...}{Any number of rbiom objects (e.g. from \code{\link[=read_biom]{read_biom()}}), lists of
rbiom objects, or valid arguments to the \code{src} parameter of
\code{\link[=read_biom]{read_biom()}} (for instance file names).}
}
\value{
An \code{rbiom} class object containing the parsed data. This object
can be treated as a list with the following named elements:
\itemize{
\item{\code{$counts} - }{
A numeric \code{\link[slam:matrix]{slam::simple_triplet_matrix()}} (sparse matrix) of
observation counts. Taxa (OTUs) as rows and samples as columns.
Access or modify using \code{\link[=otu_matrix]{otu_matrix()}}. }
\item{\code{$metadata} - }{
A \code{\link[tibble:tibble]{tibble::tibble()}} (data frame) containing any embedded
metadata. Sample IDs are in the \code{.sample} column.
Access or modify using \code{\link[=sample_metadata]{sample_metadata()}}. }
\item{\code{$taxonomy} - }{
A \code{\link[tibble:tibble]{tibble::tibble()}} (data frame) mapping OTU IDs to taxonomic
clades. Columns are named .otu, Kingdom, Phylum, Class, Order,
Family, Genus, Species, and Strain, or TaxLvl.1, TaxLvl.2, ... ,
TaxLvl.N when more than 8 levels of taxonomy are encoded in the
biom file. Access or modify using \code{\link[=otu_taxonomy]{otu_taxonomy()}}. }
\item{\code{$phylogeny} - }{
An object of class \code{phylo} defining the phylogenetic
relationships between the OTUs. Although the official
specification for BIOM only includes phylogenetic trees in BIOM
version 2.1, if an rbiom version 1.0 file includes a
\code{phylogeny} entry with newick data, then it will be loaded
here as well. The \pkg{ape} package has additional functions for
working with \code{phylo} objects. Access or modify using
\code{\link[=otu_tree]{otu_tree()}}. }
\item{\code{$sequences} - }{
A named character vector, where the names are OTU IDs and the
values are the nucleic acid sequences they represent.
These values are not part of the official BIOM specification, but
will be read and written when defined. Access or modify using
\code{\link[=otu_sequences]{otu_sequences()}}. }
\item{\code{$info} - }{
A list of other attributes defined in the BIOM file, such as
\code{id}, \code{type}, \code{format}, \code{format_url},
\code{generated_by}, \code{date}, \code{matrix_type},
\code{matrix_element_type}, \code{comment}, and \code{shape}.
Access using \code{\link[=biom_info]{biom_info()}}. Access/set with \code{\link[=biom_id]{biom_id()}} or
\code{\link[=biom_comment]{biom_comment()}}. }
}
\code{metadata}, \code{taxonomy}, and \code{phylogeny} are optional
components of the BIOM file specification and therefore will be NULL
or simple placeholders in the returned object when they are not provided
by the BIOM file.
}
\description{
WARNING: It is generally ill-advised to merge BIOM datasets, as OTUs
mappings are dependent on upstream clustering and are not equivalent
between BIOM files.
}
\examples{
    library(rbiom)
    
    b1 <- sample_select(hmp50, 1:4)
    b2 <- sample_select(hmp50, 5:8)
    
    biom <- biom_merge(b1, b2)
    print(biom)
    
    otu_tree(biom) <- otu_tree(hmp50)
    print(biom)
}
\seealso{
Other biom: 
\code{\link{as_percent}()},
\code{\link{as_rbiom}()},
\code{\link{bdply}()},
\code{\link{biom_build}()},
\code{\link{biom_id}()},
\code{\link{biom_info}()},
\code{\link{biom_repair}()},
\code{\link{blply}()},
\code{\link{read_biom}()},
\code{\link{write_biom}()},
\code{\link{write_xlsx}()}
}
\concept{biom}
