% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_biom.r
\name{build_biom}
\alias{build_biom}
\title{Create a BIOM object.}
\usage{
build_biom(
  counts,
  metadata = NULL,
  taxonomy = NULL,
  phylogeny = NULL,
  sequences = NULL,
  id = NULL,
  comments = NULL
)
}
\arguments{
\item{counts}{The count data as a numeric matrix, where column names are
the sample names and row names are the taxa IDs. A BIOM object or
a filename/URL compatible with \code{read_biom()} is also acceptable.}

\item{metadata}{A data.frame with sample IDs as the row names and metadata 
names as the column names. Or a filename/URL with comma- or 
tab-separated data.}

\item{taxonomy}{A character matrix with taxa IDs as the row names. Or a 
filename/URL with comma- or tab-separated data.}

\item{phylogeny}{A \code{phylo} object with tip labels matching the taxa 
IDs. (E.g. from \code{rbiom::read_tree()}). Or a filename/URL with 
newick formatted data.}

\item{sequences}{A named character vector of DNA sequences. Or a 
filename/URL with fasta format data.
(Currently not used by \code{rbiom}.)}

\item{id, comments}{A character vector of length one with text of the user's
choosing.}
}
\description{
Create a BIOM object.
}
\examples{
    library(rbiom)
    
    biom <- select(hmp50, 1:5)
    
    # Save components to separate files.
    ct  <- write_counts(biom, tempfile())
    md  <- write_metadata(biom, tempfile())
    tax <- write_taxonomy(biom, tempfile())
    tre <- write_tree(biom, tempfile())
    fas <- write_fasta(biom, tempfile())
    
    # Peek at the file structures.
    cat(readLines(ct, n = 2L),  '', sep="\n")
    cat(readLines(md, n = 2L),  '', sep="\n")
    cat(readLines(tax, n = 2L), '', sep="\n")
    cat(readChar(tre, nchars = 50L), "\n\n")
    cat(readChar(fas, nchars = 50L), "\n\n")
    
    # Re-assemble the BIOM object.
    biom <- build_biom(ct, md, tax, tre, id = "New BIOM")
    print(biom)
    
    # Remove temporary files.
    invisible(file.remove(c(ct, md, tax, fas, tre)))
    

}
