% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloning.r
\name{cloning}
\alias{cloning}
\title{Cloning rbiom Objects.}
\description{
Cloning rbiom Objects.
}
\examples{
    library(rbiom)
    
    biom <- hmp50$clone()
    
    new_obj  <- rarefy(biom)
    same_obj <- rarefy(biom, clone = FALSE)

}
\keyword{internal}
