% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{comments}
\alias{comments}
\title{Get \code{BIOM} object's comments.}
\usage{
comments(x)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
A length 1 character vector.
}
\description{
Note that the function is named \code{comments} (plural), while the internal
property is names \code{comment} (singular) to match the biom specification.
The plural form is taken here to avoid conflicting with the existing
\code{base::comment} function.
}
\examples{
    library(rbiom)
    comments(hmp50)

}
\seealso{
Other accessors: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
