% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert_from_phyloseq}
\alias{convert_from_phyloseq}
\title{Convert a phyloseq object to an rbiom BIOM object.}
\usage{
convert_from_phyloseq(phy)
}
\arguments{
\item{phy}{A phyloseq object.}
}
\value{
A BIOM object.
}
\description{
Convert a phyloseq object to an rbiom BIOM object.
}
\examples{
  \dontrun{
    library(rbiom)
    
    fp   <- system.file("extdata", "rich_sparse_otu_table.biom", package="phyloseq")
    phy  <- phyloseq::import_biom(fp)
    biom <- convert_from_phyloseq(phy)
 }
}
