% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert_to_SE}
\alias{convert_to_SE}
\title{Convert an rbiom object to a SummarizedExperiment object.}
\usage{
convert_to_SE(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
A SummarizedExperiment object.
}
\description{
Requires the Bioconductor R package 'SummarizedExperiment' to be installed.
}
\examples{
    
    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    print(biom)
      
    # Requires 'SummarizedExperiment' Bioconductor R package
    if (nzchar(system.file(package = "SummarizedExperiment"))) {
      se <- convert_to_SE(biom)
      print(se)
    }
}
