% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert_to_SummarizedExperiment}
\alias{convert_to_SummarizedExperiment}
\title{Convert a BIOM object to a SummarizedExperiment object.}
\usage{
convert_to_SummarizedExperiment(biom)
}
\arguments{
\item{biom}{A BIOM object, as returned from \code{\link[=read_biom]{read_biom()}}.}
}
\value{
A SummarizedExperiment object.
}
\description{
Requires the Bioconductor R package 'SummarizedExperiment' to be installed.
}
\examples{
  \dontrun{
    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    se   <- convert_to_SummarizedExperiment(biom)
 }
}
