% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert_to_TreeSummarizedExperiment}
\alias{convert_to_TreeSummarizedExperiment}
\title{Convert a BIOM object to a TreeSummarizedExperiment object.}
\usage{
convert_to_TreeSummarizedExperiment(biom)
}
\arguments{
\item{biom}{A BIOM object, as returned from \code{\link[=read_biom]{read_biom()}}.}
}
\value{
A TreeSummarizedExperiment object.
}
\description{
Requires the Bioconductor R package 'TreeSummarizedExperiment' to be installed.
}
\examples{
  \dontrun{
    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    tse  <- convert_to_TreeSummarizedExperiment(biom)
 }
}
