% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert_to_TSE}
\alias{convert_to_TSE}
\title{Convert an rbiom object to a TreeSummarizedExperiment object.}
\usage{
convert_to_TSE(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
A TreeSummarizedExperiment object.
}
\description{
Requires the Bioconductor R package 'TreeSummarizedExperiment' to be installed.
}
\examples{
      
    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    print(biom)
      
    # Requires 'TreeSummarizedExperiment' Bioconductor R package
    if (nzchar(system.file(package = "TreeSummarizedExperiment"))) {
      tse <- convert_to_TSE(biom)
      print(tse)
    }
}
