% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{counts<-}
\alias{counts<-}
\title{Set the abundance counts.}
\usage{
counts(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read_biom}.}

\item{value}{A numeric matrix. Rownames and colnames must be identical to
the current \code{counts()} value.}
}
\description{
Set the abundance counts.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    taxa <- c('Unc53100', 'Unc00c7g', 'Unc25731')
    
    counts(biom)[taxa,1:5]
    
    biom <- rarefy(biom, 100)
    counts(biom)[taxa,1:5]
    
    counts(biom) <- counts(biom) / 100
    counts(biom)[taxa,1:5]

}
\seealso{
Other setters: 
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sample_names<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa_names<-}()},
\code{\link{taxa_ranks<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
