% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{counts}
\alias{counts}
\title{Get the abundance counts.}
\usage{
counts(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
A numeric matrix of the sample abundance counts in \code{biom}.
}
\description{
Get the abundance counts.
}
\examples{
    library(rbiom)
    
    counts(hmp50)[1:4,1:5]

}
\seealso{
Other accessor functions: 
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is.rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
