% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.r
\name{default_rarefaction_depth}
\alias{default_rarefaction_depth}
\title{Pick a 'good' rarefaction depth.}
\usage{
default_rarefaction_depth(biom)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM} 
object, as returned from \link{read_biom}. For matrices, the rows and
columns are assumed to be the taxa and samples, respectively.}
}
\value{
An integer.
}
\description{
Pick a 'good' rarefaction depth.
}
