% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{depth}
\alias{depth}
\title{The rarefaction depth of a BIOM object.}
\usage{
depth(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
The rarefaction depth. If the BIOM object is not rarefied, this will be 
        a sorted vector of all the unique depths.
}
\description{
The rarefaction depth of a BIOM object.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom <- read.biom(infile)
    
    depth(biom) \%>\% head()
    
    rarefy(biom) \%>\% depth()

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{info}()},
\code{\link{is.rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
