% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depths_barplot.r
\name{depths_barplot}
\alias{depths_barplot}
\title{Visualize the number of observations per sample.}
\usage{
depths_barplot(biom, rline = TRUE, labels = TRUE, trans = "log10", ...)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM}
object, as returned from \link{read_biom}. For matrices, the rows and
columns are assumed to be the taxa and samples, respectively.}

\item{rline}{Where to draw a horizontal line on the plot, intended to show
a particular rarefaction depth. Set to \code{TRUE} to show an 
auto-selected rarefaction depth or \code{NULL} to not show a line.
Default: \code{TRUE}.}

\item{labels}{Show sample names under each bar. Default: \code{TRUE}.}

\item{trans}{Y-axis transformation. Options are \code{"log10"} or 
\code{NULL}.  Default: \code{"log10"}.}

\item{...}{Additional parameters to pass along to ggplot2 
functions. Prefix a parameter name with either \code{p.}, 
\code{r.}/\code{h.}, or \code{s.} to ensure it gets passed to (and
only to)  \link[ggplot2]{geom_point}, \link[ggplot2]{hline}, or 
\link[ggplot2]{geom_smooth}, respectively. For instance, 
\code{p.size = 2} ensures only the points have their size set to 
\code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points will be attached as 
        \code{attr(p, 'data')}.
}
\description{
Visualize the number of observations per sample.
}
\examples{
    library(rbiom)
    
    depths_barplot(hmp50, rline=TRUE) 
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()}
}
\concept{plotting}
