% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.r
\name{distill}
\alias{distill}
\title{Convert to a data.frame with metadata and diversity/abundance values.}
\usage{
distill(
  biom,
  metric,
  weighted = TRUE,
  rarefy = TRUE,
  long = TRUE,
  md = TRUE,
  safe = FALSE
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read.biom}.}

\item{metric}{The diversity/abundance values of interest. Options are:
\describe{
    \item{Alpha Diversity Metrics}{
      \bold{OTUs}, \bold{Shannon}, \bold{Chao1}, \bold{Simpson}, and/or \bold{InvSimpson}.
    }
    \item{Beta Diversity Metrics}{
      \bold{manhattan}, \bold{euclidean}, \bold{bray-curtis}, \bold{jaccard}, 
      or \bold{unifrac}. Use in combination with the \code{weighted} parameter.
      Metadata column names can be prefixed with \bold{==} or \bold{!=} to limit distance
      calculations to \emph{within} or \emph{between}, respectively, those categories. See
      examples below.
    }
    \item{Taxa Abundances}{
      \bold{Kingdom}, \bold{Phylum}, \bold{Class}, \bold{Order}, \bold{Family}, \bold{Genus}, 
      \bold{Species}, \bold{Strain}, or \bold{OTU}. Supported ranks will vary by biom. Run
      \code{taxa.ranks(biom)} to see the available options.
    }
   }}

\item{weighted}{When \bold{metric} is a beta diversity metric, should it be run in in weighted
mode? (Default: TRUE)}

\item{rarefy}{Should the dataset be rarefied first? When \bold{metric} is an alpha diversity 
metric, this 'rarefy' parameter is passed on directly to \code{alpha.div()}. (Default: TRUE)}

\item{long}{Pivot the returned data to long format?
\describe{
  \item{\bold{FALSE}}{ Each metric has its own column. }
  \item{\bold{TRUE}}{  "Sample", "Metric" and "Diversity" are the columns 
                       returned. Rows are added to attain all combinations of 
                       samples x metrics. (Default) }
}}

\item{md}{Include metadata in the output data frame? Options are: 
\describe{
  \item{\code{FALSE}}{ Don't include metadata. (Default) }
  \item{\code{TRUE}}{ Include all metadata. }
  \item{\emph{character vector}}{ Include only the specified metadata columns. }
}}

\item{safe}{Should autogenerated columns be prefixed with a "." to avoid conflicting
with metadata column names? (Default: FALSE)}
}
\value{
A \code{data.frame} object. The first column will be named \bold{Sample}, 
        and when possible, the rownames will hold the sample name as well. For
        beta diversity metrics the first two columns will be named \bold{Sample1}
        and \bold{Sample2}. The remaining columns will be largely dependent on
        \code{metric}, \code{long}, and \code{md}. See examples below.
}
\description{
Convert to a data.frame with metadata and diversity/abundance values.
}
\examples{
    library(rbiom)
    
    distill(hmp50, "UniFrac", md=c("==Body Site", "!=Sex")) \%>\% head()
    
    distill(hmp50, c("Shannon", "OTUs"), md=c("BMI", "Sex")) \%>\% head()
    
    distill(hmp50, "Phylum", long=FALSE, md=FALSE)[1:4,1:4]
    
    distill(hmp50, "Phylum", long=FALSE, md=c("Age", "Body Site"))[1:4,1:6]

}
\seealso{
\code{\link{stats.table}}
}
