% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_distmat.r
\name{distmat_stats}
\alias{distmat_stats}
\title{Run statistics on a distance matrix vs a categorical or numeric variable.}
\usage{
distmat_stats(
  dm,
  groups,
  test = "adonis2",
  seed = 0,
  permutations = 999,
  p.adj = "fdr"
)
}
\arguments{
\item{dm}{The distance matrix as a \code{dist} object, or a list of
\code{dist} objects.}

\item{groups}{A named vector of grouping values. The names should
correspond to \code{attr(dm, 'Labels')}.}

\item{test}{The name of the function from the \code{vegan} R package to
use. Options are: \code{"adonis2"} and \code{"mrpp"}. Non-ambiguous
abbreviations of the function names are also accepted.
Default: \code{"adonis2"}}

\item{seed}{Random seed for permutations. Default: \code{0}}

\item{permutations}{Number of random permutations to use.
Default: \code{999}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}.}
}
\value{
A data.frame with summary statistics from \code{\link[vegan:permustats]{vegan::permustats()}}. The columns are:
\itemize{
\item{\emph{stat} - }{ The observed statistic. For mrpp, this is the overall weighted mean of group mean distances. }
\item{\emph{z} - }{ The difference of observed statistic and mean of permutations divided by the standard deviation of permutations (also known as z-values). Evaluated from permuted values without observed statistic. }
\item{\emph{p.val} - }{ Include only the specified metadata columns. }
\item{\emph{adj.p} - }{ P-value after correcting for the false discovery rate with the \code{p.adj} method. }
\item{\emph{test} - }{ The \code{test} and \code{p.adj} methods used. }
}
}
\description{
Run statistics on a distance matrix vs a categorical or numeric variable.
}
\examples{
    library(rbiom)
    
    biom <- sample_select(hmp50, 1:10)
    dm   <- bdiv_distmat(biom, 'unifrac')
    distmat_stats(dm, groups = sample_metadata(biom, 'Body Site'))
    distmat_stats(dm, groups = sample_metadata(biom, 'Age'))

}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_distmat}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_table}()}
}
\concept{beta_diversity}
