% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_boxplot}
\alias{documentation_boxplot}
\title{documentation_boxplot}
\arguments{
\item{x}{A categorical metadata column name to use for the x-axis. The
default, \code{NULL}, groups all samples into a single category.}

\item{layers}{One or more of
\code{c("bar", "box" ("x"), "violin", "dot", "strip", "crossbar", "errorbar", "linerange", "pointrange")}.
Single letter abbreviations are also accepted. For instance,
\code{c("box", "dot")} is equivalent to \code{c("x", "d")} and \code{"xd"}.
See \link[=plots]{plot types} for examples of each. Default: \code{"bld"}}

\item{color.by, pattern.by, shape.by, facet.by, limit.by}{Metadata columns to
use for aesthetics and partitioning. Default: \code{NULL}}

\item{ci}{How to calculate min/max of the \bold{crossbar},
\bold{errorbar}, \bold{linerange}, and \bold{pointrange} layers.
Options are: \code{"ci"} (confidence interval), \code{"range"},
\code{"sd"} (standard deviation), \code{"se"} (standard error), and
\code{"mad"} (median absolute deviation).
The center mark of \bold{crossbar} and \bold{pointrange} represents
the mean, except for code{"mad"} in which case it represents
the median. Default: \code{"ci"}}

\item{...}{Parameters are matched to formal arguments of ggplot2
functions. Prefixing parameter names with a layer name ensures that
a particular parameter is passed to, and only to, that layer. For
instance, \code{dot.size = 2} or \code{d.size = 2} ensures only the
dotplot layer has its size set to \code{2}. The special prefix
\code{pt.} will control both the dot and strip layers.}
}
\value{
A \code{ggplot2} plot. \cr The computed data points, statistics,
and ggplot command are available as \verb{$data}, \verb{$stats}, and
\verb{$code}, respectively.
}
\description{
documentation_boxplot
}
\keyword{internal}
