% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_boxplot_aes_section}
\alias{documentation_boxplot_aes_section}
\title{documentation_boxplot_aes_section}
\description{
documentation_boxplot_aes_section
}
\section{Aesthetics and Partitions}{


Metadata can be used to flexibly subset, partition, and apply aesthetics
when creating a plot. Common use cases are provided below. More thorough
documentation is available at \url{https://cmmr.github.io/rbiom}.

\preformatted{  ## Colors ----------------------------
  color.by = "Body Site"
  color.by = list('Body Site' = "bright")
  color.by = list('Body Site' = c("Stool", "Saliva"))
  color.by = list('Body Site' = list('values' = c("Stool", "Saliva"), 'colors' = "bright"))
  color.by = list('Body Site' = c('Stool' = "blue", 'Saliva' = "green"))
  
  ## Patterns --------------------------
  pattern.by = "Body Site"
  pattern.by = list('Body Site' = c("Stool", "Saliva"))
  pattern.by = list('Body Site' = c('Stool' = "left45", 'Saliva' = "hs_cross"))
  
  ## Shapes ----------------------------
  shape.by = "Body Site"
  shape.by = list('Body Site' = c("Stool", "Saliva"))
  shape.by = list('Body Site' = c('Stool' = 7, 'Saliva' = 8))
  
  ## Facets ----------------------------
  facet.by = "Body Site"
  facet.by = c("Body Site", "Sex")
  facet.by = list('Body Site' = c("Stool", "Saliva"), "Sex")
  
  ## Limits ----------------------------
  limit.by = list('Sex' = "Male", 'Age' = c(20,40))
  limit.by = list('Body Site' = c("Saliva", "Anterior nares"), 'Age' = c(NA,35))
}

\itemize{
\item{\code{color.by} - }{A categorical metadata column. (Max 1)}
\item{\code{pattern.by} - }{A categorical metadata column. (Max 1)}
\item{\code{shape.by} - }{A categorical metadata column. (Max 1)}
\item{\code{facet.by} - }{Categorical metadata column(s).}
\item{\code{limit.by} - }{Any metadata column(s).}
}

All built-in color palettes are colorblind-friendly. The available
categorical palette names are: \code{"okabe"}, \code{"carto"}, \code{"r4"},
\code{"polychrome"}, \code{"tol"}, \code{"bright"}, \code{"light"},
\code{"muted"}, \code{"vibrant"}, \code{"tableau"}, \code{"classic"},
\code{"alphabet"}, \code{"tableau20"}, \code{"kelly"}, and \code{"fishy"}.

Patterns are sourced from the magick R package. Pattern names are:
\code{"bricks"}, \code{"hexagons"}, \code{"horizontalsaw"},
\code{"hs_fdiagonal"}, \code{"fishscales"}, \code{"verticalsaw"},
\code{"checkerboard"}, \code{"octagons"}, \code{"right45"},
\code{"hs_cross"}, \code{"hs_bdiagonal"}, \code{"hs_diagcross"},
\code{"hs_horizontal"}, \code{"hs_vertical"}, \code{"left45"},
\code{"leftshingle"}, \code{"rightshingle"}, \code{"verticalbricks"},
\code{"verticalleftshingle"}, and \code{"verticalrightshingle"}.

Shapes can be given as per base R - numbers 0 through 17 for various shapes,
or the decimal value of an ascii character, e.g. a-z = 65:90; A-Z = 97:122 to use
letters instead of shapes on the plot. Character strings may used as well.
}

\keyword{internal}
