% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_clusters}
\alias{documentation_clusters}
\title{documentation_clusters}
\arguments{
\item{k}{Number of clusters. Default: \code{5L}}

\item{rank}{Which taxa rank to use. E.g. \code{"Phylum"},
\code{"Genus"}, \code{".otu"}, etc. An integer can also be
given, where \code{1} is the highest rank, \code{2} is the second
highest, \code{-1} is the lowest rank, \code{-2} is the second
lowest, and \code{0} is the OTU "rank". Run \code{biom$ranks} to
see all options for a given rbiom object. Default: \code{.otu}.}
}
\value{
A numeric factor assigning samples to clusters.
}
\description{
documentation_clusters
}
\keyword{internal}
