% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_cmp}
\alias{documentation_cmp}
\title{documentation_cmp}
\arguments{
\item{x}{A categorical metadata column name. Prefix the column name with
\code{==} or \code{!=} to limit comparisons to within or between
groups, respectively. The default, \code{NULL} groups all distances
into a single column.}

\item{color.by, pattern.by, shape.by, facet.by, limit.by}{Metadata columns to
use for data partitioning. Prefix the column name with
\code{==} or \code{!=} to limit comparisons to within or between
groups, respectively. Default: \code{NULL}}

\item{within, between}{Metadata field(s) for intra- or inter- sample
comparisons. Default: \code{NULL}}
}
\description{
documentation_cmp
}
\keyword{internal}
