% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_corrplot}
\alias{documentation_corrplot}
\title{documentation_corrplot}
\arguments{
\item{x}{A numeric metadata field to use for the x-axis. Required.}

\item{layers}{One or more of
\code{c("trend", "confidence", "scatter", "name")}. Single
letter abbreviations are also accepted. For instance,
\code{c("trend", "scatter")} is equivalent to \code{c("t", "s")} and \code{"ts"}.
See \link[=plots]{plot types} for examples of each. Default: \code{"tc"}}

\item{color.by, facet.by, limit.by}{Metadata columns to use for aesthetics
and partitioning. See below for details. Default: \code{NULL}}

\item{formula}{Relationship between variables. Default: \code{y ~ x}}

\item{engine}{What type of trend model to fit to the data. Options are:
\code{"lm"} (linear), \code{"local"} (uses loess or gam). Default: \code{"lm"}}

\item{...}{Additional parameters to pass along to ggplot2
functions. Prefix a parameter name with either \code{t.} or
\code{s.}/\code{pt.} to ensure it gets passed to (and only to)
\link[ggplot2]{geom_smooth} or \link[ggplot2]{geom_point},
respectively. For instance, \code{s.size = 2} ensures only the
scatterplot points have their size set to \code{2}.}
}
\value{
A \code{ggplot2} plot. \cr The computed data points and ggplot command
are available as \verb{$data} and \verb{$code}, respectively.
}
\description{
documentation_corrplot
}
\section{Aesthetics and Partitions}{


Metadata can be used to flexibly subset, partition, and apply aesthetics
when creating a plot. Common use cases are provided below. More thorough
documentation is available at \url{https://cmmr.github.io/rbiom}.

\preformatted{  ## Colors ----------------------------
  color.by = "Body Site"
  color.by = list('Body Site' = "bright")
  color.by = list('Body Site' = c("Stool", "Saliva"))
  color.by = list('Body Site' = list('values' = c("Stool", "Saliva"), 'colors' = "bright"))
  color.by = list('Body Site' = c('Stool' = "blue", 'Saliva' = "green"))
  
  ## Facets ----------------------------
  facet.by = "Body Site"
  facet.by = c("Body Site", "Sex")
  facet.by = list('Body Site' = c("Stool", "Saliva"), "Sex")
  
  ## Limits ----------------------------
  limit.by = list('Sex' = "Male", 'Age' = c(20,40))
  limit.by = list('Body Site' = c("Saliva", "Anterior nares"), 'Age' = c(NA,35))
}

\itemize{
\item{\code{color.by} - }{Any metadata column. (Max 1)}
\item{\code{facet.by} - }{Only categorical metadata column(s).}
\item{\code{limit.by} - }{Any metadata column(s).}
}

All built-in color palettes are colorblind-friendly.

The available categorical palette names are: \code{"okabe"}, \code{"carto"},
\code{"r4"}, \code{"polychrome"}, \code{"tol"}, \code{"bright"},
\code{"light"}, \code{"muted"}, \code{"vibrant"}, \code{"tableau"},
\code{"classic"}, \code{"alphabet"}, \code{"tableau20"}, \code{"kelly"},
and \code{"fishy"}.

The available numeric palette names are: \code{"reds"}, \code{"oranges"},
\code{"greens"}, \code{"purples"}, \code{"grays"}, \code{"acton"},
\code{"bamako"}, \code{"batlow"}, \code{"bilbao"}, \code{"buda"},
\code{"davos"}, \code{"devon"}, \code{"grayC"}, \code{"hawaii"},
\code{"imola"}, \code{"lajolla"}, \code{"lapaz"}, \code{"nuuk"},
\code{"oslo"}, \code{"tokyo"}, \code{"turku"}, \code{"bam"},
\code{"berlin"}, \code{"broc"}, \code{"cork"}, \code{"lisbon"},
\code{"roma"}, \code{"tofino"}, \code{"vanimo"}, \code{"vik"}
}

\keyword{internal}
