% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_dist_test}
\alias{documentation_dist_test}
\title{documentation_dist_test}
\arguments{
\item{stat.by}{The categorical or numeric metadata field over which
statistics should be calculated. Required.}

\item{test}{Permutational test for accessing significance. Options are:
\describe{
\item{\code{"adonis2"} - }{ Permutational MANOVA; \code{\link[vegan:adonis]{vegan::adonis2()}}. }
\item{\code{"mrpp"} - }{ Multiple response permutation procedure; \code{\link[vegan:mrpp]{vegan::mrpp()}}. }
\item{\code{"none"} - }{ Don't run any statistics. }
}
Abbreviations are allowed. Default: \code{"adonis2"}}
}
\description{
documentation_dist_test
}
\keyword{internal}
