% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_model_section}
\alias{documentation_model_section}
\title{documentation_model_section}
\description{
documentation_model_section
}
\section{Model Options}{


The predefined options are:
\itemize{
\item{\code{"lm"} - }{  Linear model: \code{stats::lm(formula = y ~ x)}.) }
\item{\code{"log"} - }{ Logarithmic model: \code{stats::lm(formula = y ~ log(x))}. }
\item{\code{"gam"} - }{ Generalized additive model: \code{mgcv::gam(formula = y ~ s(x, bs = "cs"), method = "REML")}. }
}

You can alternatively provide a list of length two where the first
element is a character vector of length 1 naming a function, and the
second element is a list of arguments to pass to that function. One
of the function's arguments must be named 'formula'.
For example, \code{model = list("stats::lm", list(formula = y ~ x))}.

By default, the same arguments are used for both plotting and calculating
statistics. Arguments just for statistics can be set as a third list
element:
\code{model = list("stats::lm", list(formula = y ~ x), list(formula = rank(y) ~ x))}.
}

\keyword{internal}
