% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_plot_stats_return}
\alias{documentation_plot_stats_return}
\title{documentation_plot_stats_return}
\value{
A \code{ggplot2} plot. \cr The computed data points, statistics,
and ggplot command are available as \code{$data}, \code{$stats}, and
\code{$code}, respectively.
}
\description{
documentation_plot_stats_return
}
\keyword{internal}
