% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_rank.NULL}
\alias{documentation_rank.NULL}
\title{documentation_rank.NULL}
\arguments{
\item{rank}{What rank(s) of taxa to compute biplot coordinates and
statistics for, or \code{NULL} to disable. E.g. \code{"Phylum"},
\code{"Genus"}, \code{".otu"}, etc. An integer vector can also be
given, where \code{1} is the highest rank, \code{2} is the second
highest, \code{-1} is the lowest rank, \code{-2} is the second
lowest, and \code{0} is the OTU "rank". Run \code{biom$ranks} to
see all options for a given rbiom object. Default: \code{NULL}.}
}
\description{
documentation_rank.NULL
}
\keyword{internal}
