% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_return.biom}
\alias{documentation_return.biom}
\title{documentation_return.biom}
\value{
An \link[=rbiom_objects]{rbiom object}.
}
\description{
documentation_return.biom
}
\keyword{internal}
