% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_stats_return}
\alias{documentation_stats_return}
\title{documentation_stats_return}
\value{
A tibble data frame with summary statistics. \cr
The R code or generating these statistics is in \code{$code}.
}
\description{
documentation_stats_return
}
\keyword{internal}
