% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_taxa.4}
\alias{documentation_taxa.4}
\title{documentation_taxa.4}
\arguments{
\item{taxa}{Which taxa to display. An integer value will show the top n
most abundant taxa. A value 0 <= n < 1 will show any taxa with that
mean abundance or greater (e.g. \code{0.1} implies >= 10\%). A
character vector of taxa names will show only those named taxa.
Default: \code{4}.}
}
\description{
documentation_taxa.4
}
\keyword{internal}
