% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_test.pw_means}
\alias{documentation_test.pw_means}
\title{documentation_test.pw_means}
\arguments{
\item{test}{The statistic of interest. An overview of options can be
found in the "Statistical Tests" section below. One of
\code{"predict"}, \code{"terms"}, \code{"fit"}, \code{"means"},
\code{"trends"}, \code{"es_means"}, \code{"es_trends"},
\code{"pw_means"}, or \code{"pw_trends"}.
Default: \code{"pw_means"}}
}
\description{
documentation_test.pw_means
}
\section{Statistical Tests}{


\itemize{
\item{\code{"predict"} - }{
Augments original data with fitted information. See
\code{\link[broom:augment.lm]{broom::augment.lm()}}. }
\item{\code{"terms"} - }{ Summary information about the model's
terms, including p-value, r-squared, AIC, BIC, and confidence
interval at \code{level}. See \code{\link[broom:tidy.lm]{broom::tidy.lm()}}. }
\item{\code{"fit"} - }{ Goodness of fit measures, p-values, and
more for the overall model. See \code{\link[broom:glance.lm]{broom::glance.lm()}}. }
\item{\code{"means"} - }{ The estimated marginal mean (EMM) for
each \code{stat.by} group, along with confidence intervals (CI),
standard errors (SE), t-ratios, and p-values testing for mean = zero.
See \code{\link[emmeans:emmeans]{emmeans::emmeans()}} and \code{\link[emmeans:summary.emmGrid]{emmeans::summary.emmGrid()}}. }
\item{\code{"trends"} - }{ The trendline slope for each
\code{stat.by} group, along with CI and SE and p-value testing for
slope = zero. See \code{\link[emmeans:emtrends]{emmeans::emtrends()}} and
\code{\link[emmeans:summary.emmGrid]{emmeans::summary.emmGrid()}}. }
\item{\code{"pw_means"} - }{ Pairwise means. All \code{stat.by} groups are
compared to each other and the difference in means is estimated
along with SE, t-ratios, and p-values testing if the two means
are the same. See \code{\link[emmeans:contrast]{emmeans::pairs.emmGrid()}}. }
\item{\code{"pw_trends"} - }{ Pairwise trends. As above, but
comparing trendline slopes instead of means. }
\item{\code{"es_means"} - }{ Effect sizes for pairwise differences
of means, including SE and CI. See \code{\link[emmeans:eff_size]{emmeans::eff_size()}}. }
\item{\code{"es_trends"} - }{ Effect sizes for pairwise differences
of slopes, including SE and CI. See \code{\link[emmeans:eff_size]{emmeans::eff_size()}}. }
}
}

\keyword{internal}
