% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_tranform.y.ifelse}
\alias{documentation_tranform.y.ifelse}
\title{documentation_tranform.y.ifelse}
\arguments{
\item{transform.y}{How to transform y-axis values. Useful for correcting
for non-normally distributions before applying regression statistics.
Options are \code{"none"} or \code{"rank"}.
Default: \code{ifelse(is.null(regr), "rank", "none")}}
}
\description{
documentation_tranform.y.ifelse
}
\keyword{internal}
