% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.r
\name{embed_code}
\alias{embed_code}
\title{Wraps R code with Markdown syntax highlighting.}
\usage{
embed_code(x)
}
\arguments{
\item{x}{Some R code.}
}
\value{
Markdown-compatible syntax highlighted code block.
}
\description{
Wraps R code with Markdown syntax highlighting.
}
\examples{
    library(rbiom)
    p <- adiv_boxplot(hmp50, color.by = "Body Site")
    attr(p, 'cmd') \%>\% embed_code() \%>\% cat("\n\n")
}
\seealso{
Other embed: 
\code{\link{embed_csv}()}
}
\concept{embed}
