% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.r
\name{embed_csv}
\alias{embed_csv}
\title{Converts a data.frame to a download link.}
\usage{
embed_csv(x, label = "Download Data (CSV)", filename = "data.csv", ...)
}
\arguments{
\item{x}{An object coercable to a \code{data.frame}.}

\item{filename}{Default filename to download as. Default: \code{"data.csv"}.}

\item{...}{Additional arguments to pass to \code{write.csv()}.}

\item{title}{Text/HTML label for the link. Default: \code{"Download Data (CSV)"}.}
}
\value{
An HTML string in which the object is encoded in base64.
}
\description{
Converts a data.frame to a download link.
}
\examples{
    library(rbiom)
    p <- plot(hmp50, Bray ~ Sex)
    attr(p, 'stats') \%>\% embed_csv(row.names=FALSE) \%>\% cat("\n\n")
}
\seealso{
Other embed: 
\code{\link{embed_code}()}
}
\concept{embed}
