% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.r
\name{glimpse.rbiom}
\alias{glimpse.rbiom}
\title{Get a glimpse of your metadata.}
\usage{
\method{glimpse}{rbiom}(biom, width = NULL, ...)
}
\arguments{
\item{biom}{An rbiom-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{width}{Width of output. See \code{\link[pillar:glimpse]{pillar::glimpse()}} documentation.
Default: \code{NULL}}

\item{...}{Unused, for extensibility.}
}
\value{
The original \code{biom}, invisibly.
}
\description{
Get a glimpse of your metadata.
}
\examples{
    library(rbiom)
    
    glimpse(hmp50)
    
}
\seealso{
Other metadata: 
\code{\link{bdply}()}
}
\concept{metadata}
