% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{has.sequences}
\alias{has.sequences}
\title{Checks if DNA sequences are present.}
\usage{
has.sequences(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
\code{TRUE} if DNA sequences are present, \code{FALSE} otherwise.
}
\description{
Checks if DNA sequences are present.
}
\examples{
    library(rbiom)
    has.sequences(hmp50)

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is.rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
