% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{has_phylogeny}
\alias{has_phylogeny}
\title{Checks if a phylogenetic tree is present.}
\usage{
has_phylogeny(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
\code{TRUE} if a phylogenetic tree is present, \code{FALSE} otherwise.
}
\description{
Checks if a phylogenetic tree is present.
}
\examples{
    library(rbiom)
    
    has_phylogeny(hmp50)

}
\seealso{
Other accessors: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
