% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{has_sequences}
\alias{has_sequences}
\title{Check if sequences are included in the rbiom object.}
\usage{
has_sequences(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
\code{TRUE} if \code{biom} has sequences, \code{FALSE} otherwise.
}
\description{
Check if sequences are included in the rbiom object.
}
\examples{
    library(rbiom)
    
    has_sequences(hmp50)

}
\seealso{
Other sequences: 
\code{\link{otu_sequences}()},
\code{\link{write_fasta}()}
}
\concept{sequences}
