% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{has_tree}
\alias{has_tree}
\alias{has_sequences}
\alias{is_rarefied}
\title{Check specific properties of an rbiom object.}
\usage{
has_tree(biom)

has_sequences(biom)

is_rarefied(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
\code{TRUE} if it has the property, \code{FALSE} otherwise.
}
\description{
Check specific properties of an rbiom object.
}
\examples{
    library(rbiom)
    
    has_tree(hmp50)
    has_sequences(hmp50)
    is_rarefied(hmp50)
    
}
