% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{hmp50}
\alias{hmp50}
\title{Human Microbiome Project - 50 sample demo dataset.}
\format{
A BIOM object with 50 samples. 
        Includes metadata, taxonomy, phylogeny, and sequences.
\describe{
  \item{Sex}{Male or Female}
  \item{Body Site}{Buccal mucosa, Saliva, Anterior nares, Stool, or Mid vagina}
  \item{Age}{21 - 40}
  \item{BMI}{19 - 32}
}
}
\source{
\url{https://portal.hmpdacc.org}
}
\usage{
hmp50
}
\description{
Human Microbiome Project - 50 sample demo dataset.
}
\keyword{datasets}
