% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_table.r
\name{import_table}
\alias{import_table}
\title{Read a data.frame from a filename/URL. Auto-detects separator.}
\usage{
import_table(filename, matrix = FALSE, ...)
}
\arguments{
\item{filename}{A file name or URL. Must be tab- or comma-separated 
content. Either all or no fields should be quoted.}

\item{matrix}{Set to \code{TRUE} to convert data.frame to a matrix. Set to
\code{"character"}, \code{"integer"}, etc to coerce to a matrix of
that type. The default, \code{FALSE}, does no conversion.}

\item{...}{Additional arguments for \code{read.table()}.}
}
\value{
A data.frame or matrix.
}
\description{
Read a data.frame from a filename/URL. Auto-detects separator.
}
