% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{info}
\alias{info}
\title{Get biom's misc information.}
\usage{
info(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
A data frame of the metadata in \code{biom}.
}
\description{
Get biom's misc information.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    
    info(biom)

}
\seealso{
\code{\link{sample.names}}, \code{\link{taxa.names}}, 
    \code{\link{taxa.ranks}}, \code{\link{counts}}, 
    \code{\link{taxonomy}}, \code{\link{phylogeny}},
    \code{\link{metadata}}, \code{\link{info}}

Other accessor functions: \code{\link{counts}},
  \code{\link{metadata}}, \code{\link{phylogeny}},
  \code{\link{sample.names}}, \code{\link{taxa.names}},
  \code{\link{taxa.ranks}}, \code{\link{taxonomy}}
}
