% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_cache.r
\name{init_cache}
\alias{init_cache}
\title{Initialize caching for expensive biom operations.}
\usage{
init_cache(cm = 50 * 1024^2, ...)
}
\arguments{
\item{cm}{A memoise-compatible cache object. 
E.g. \code{cachem::cache_mem()}, \code{cachem::cache_disk()}, or 
\code{cachem::cache_layered()}. If \code{cm} is numeric, it will
be used to define the maximum cache size (in bytes) for an in-memory 
cache. Default: \code{50 * 1024^2} \emph{(50 MB)}}

\item{...}{Other arguments passed on to \code{cachem::cache_mem}. 
For example: \code{max_size = 512 * 1024^2}, 
\code{max_age = Inf}, \code{max_n = Inf}, etc.}
}
\value{
The cache object, invisibly.
}
\description{
Speeds up repetitive computations by storing results of functions calls. 
This cache initialization function can only be called once. Subsequent calls 
return a warning.
}
\examples{
    library(rbiom)
    
    init_cache()
    system.time(x <- adiv_table(hmp50, "multi"))
    system.time(y <- adiv_table(hmp50, "multi"))
    identical(x, y)

}
