% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{is.rarefied}
\alias{is.rarefied}
\title{Checks if a BIOM object is rarefied.}
\usage{
is.rarefied(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
\code{TRUE} if the BIOM object is rarefied, \code{FALSE} otherwise.
}
\description{
Checks if a BIOM object is rarefied.
}
\examples{
    library(rbiom)
    
    is.rarefied(hmp50)
    
    rarefy(hmp50, 1000) \%>\% is.rarefied()

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
