% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{is_rarefied}
\alias{is_rarefied}
\title{Checks if a BIOM object is rarefied.}
\usage{
is_rarefied(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
\code{TRUE} if the BIOM object is rarefied, \code{FALSE} otherwise.
}
\description{
Checks if a BIOM object is rarefied.
}
\examples{
    library(rbiom)
    
    is_rarefied(hmp50)
    
    rarefy(hmp50, 1000) \%>\% is_rarefied()

}
\seealso{
Other accessors: 
\code{\link{comments}()},
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
