% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.r
\name{matrix_ops}
\alias{matrix_ops}
\alias{mtx_rarefy}
\alias{mtx_percent}
\alias{mtx_rescale}
\alias{rarefy_cols}
\alias{rescale_rows}
\alias{rescale_cols}
\title{Transform a counts matrix.}
\usage{
mtx_rarefy(
  mtx,
  margin = 2L,
  depth = 0.1,
  n = NULL,
  seed = 0L,
  upsample = NULL,
  cpus = NULL
)

mtx_percent(mtx, margin = 2L)

mtx_rescale(mtx, margin = 2L, range = c(0, 1))

rarefy_cols(mtx, depth = 0.1, n = NULL, seed = 0L, cpus = NULL)

rescale_rows(mtx)

rescale_cols(mtx)
}
\arguments{
\item{mtx}{A matrix-like object.}

\item{margin}{Apply the transformation to the matrix's rows (\code{margin=1L})
or columns (\code{margin=2L}). Instead of \code{1L} and \code{2L}, you may also use
\code{'rows'} and \code{'cols'}.
Default: \code{2L} (column-wise, aka sample-wise for otu tables)}

\item{depth}{How many observations to keep per sample. When
\verb{0 < depth < 1}, it is taken as the minimum percentage of the
dataset's observations to keep. Ignored when \code{n} is specified.
Default: \code{0.1}}

\item{n}{The number of samples to keep. When \verb{0 < n < 1}, it is taken as
the percentage of samples to keep. If negative, that number or
percentage of samples is dropped. If \code{0}, all samples are kept. If
\code{NULL}, \code{depth} is used instead.
Default: \code{NULL}}

\item{seed}{Random seed for permutations. Must be a non-negative integer.
Default: \code{0}}

\item{upsample}{If the count data is in percentages, provide an integer
value here to scale each sample's observations to integers that sum
to this value. Generally not recommended, but can be used to
'shoehorn' metagenomic abundance estimates into rbiom's functions
that were designed for amplicon datasets. When invoked, \code{depth}, \code{n},
and \code{seed} are ignored. The default, \code{NULL}, will throw an error if
the counts are not all integers.}

\item{cpus}{The number of CPUs to use. Set to \code{NULL} to use all available,
or to \code{1} to disable parallel processing. Default: \code{NULL}}

\item{range}{When rescaling, what should the minimum and maximum values
be? Default: \code{c(0, 1)}}
}
\value{
The transformed matrix. If \code{mtx} was a sparse matrix from the \code{Matrix}
package, then the result will also be a sparse matrix,
otherwise the result will be a base R matrix.
}
\description{
A collection of transformations that operate directly on matrices.\cr\cr
Note: \code{rarefy_cols()}, \code{rescale_rows()}, and \code{rescale_cols()} are deprecated.
}
\examples{
    library(rbiom)
    
    # mtx_rarefy --------------------------------------
    biom <- hmp50$clone()
    sample_sums(biom) \%>\% head(10)

    biom$counts \%<>\% mtx_rarefy(depth=1000)
    sample_sums(biom) \%>\% head(10)
    
    
    # rescaling ----------------------------------------
    mtx <- matrix(sample(1:20), nrow=4)
    mtx
    
    colSums(mtx)
    
    colSums(mtx_rarefy(mtx))
    
    colSums(mtx_percent(mtx))
    
    apply(mtx_rescale(mtx), 2L, max)

}
\seealso{
Other transformations: 
\code{\link{modify_metadata}},
\code{\link{rarefy}()},
\code{\link{slice_metadata}},
\code{\link{subset}()},
\code{\link{with}()}
}
\concept{low_level}
\concept{transformations}
