% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.r
\name{metadata<-}
\alias{metadata<-}
\title{Set the \code{BIOM} object's metadata.}
\usage{
metadata(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read.biom}.}

\item{value}{A data.frame with the metadata. All the \code{rownames()} must 
be in \code{sample.names(biom)}. If there are fewer rows of 
\code{data} than samples, then the \code{biom} object will be subset.}
}
\description{
Set the \code{BIOM} object's metadata.
}
\examples{
    library(rbiom)
    
    md <- metadata(hmp50)
    md <- md[,c('Sex', 'Body Site')]
    metadata(hmp50) <- md
    head(metadata(hmp50))

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sample.names<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa.names<-}()},
\code{\link{taxa.ranks<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
