% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{metadata_names}
\alias{metadata_names}
\alias{metadata_names<-}
\title{Get or set the metadata column names.}
\usage{
metadata_names(biom)

metadata_names(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{value}{A character vector of the new metadata column names. Names
cannot start with a \code{.} character. Leading and trailing
whitespace is automatically removed.}
}
\value{
A character vector of the metadata column names in \code{biom}.
}
\description{
Get or set the metadata column names.
}
\examples{
    library(rbiom) 
    
    biom <- hmp50
    
    metadata_names(biom)
    
    metadata_names(biom) <- sub(' ', '_', tolower(metadata_names(biom)))
    metadata_names(biom)
}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
