% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{metadata_numeric}
\alias{metadata_numeric}
\title{Check which metadata columns are numeric.}
\usage{
metadata_numeric(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
A named logical vector indicating if each metadata column is
numeric.
}
\description{
Check which metadata columns are numeric.
}
\examples{
    library(rbiom)
    metadata_numeric(hmp50)

}
\seealso{
Other metadata: 
\code{\link{bdply}()},
\code{\link{blply}()},
\code{\link{metadata_names}()},
\code{\link{n_metadata}()},
\code{\link{sample_metadata}()},
\code{\link{sample_select}()},
\code{\link{subset.rbiom}()},
\code{\link{within.rbiom}()},
\code{\link{write_metadata}()}
}
\concept{metadata}
