% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.r
\name{metadata_params}
\alias{metadata_params}
\title{Convert all *.by params into long form.}
\usage{
metadata_params(params, contraints = list())
}
\description{
Examples values for color.by:
}
\details{
"Body Site"
c("Body Site" = "okabe")
list("Body Site")
list("Body Site" = "okabe")
list("Body Site" = list(values = "okabe", name = "Swab Location"))
list("Body Site" = list(values = c("red", "green", "blue"), name = "Swab Location"))
list("Body Site" = list(values = c(Saliva = "red", Stool = "green"), name = "Swab Location"))


c("Body Site", "Sex")
c("Body Site" = "okabe", "Sex")
c("Body Site" = "okabe", "Sex" = "muted")
list("Body Site", "Sex")
}
