% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{n_metadata}
\alias{n_metadata}
\title{Report the number of metadata fields in an rbiom object.}
\usage{
n_metadata(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
The number of metadata fields present.
}
\description{
Report the number of metadata fields in an rbiom object.
}
\examples{
    library(rbiom)
    
    n_metadata(hmp50)

}
\seealso{
Other metadata: 
\code{\link{bdply}()},
\code{\link{blply}()},
\code{\link{metadata_names}()},
\code{\link{metadata_numeric}()},
\code{\link{sample_metadata}()},
\code{\link{sample_select}()},
\code{\link{subset.rbiom}()},
\code{\link{within.rbiom}()},
\code{\link{write_metadata}()}
}
\concept{metadata}
