% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{n_otus}
\alias{n_otus}
\title{Report the number of otus in an rbiom object.}
\usage{
n_otus(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
The number of otus present.
}
\description{
Report the number of otus in an rbiom object.
}
\examples{
    library(rbiom)
    
    n_otus(hmp50)
    
}
\seealso{
Other otus: 
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{write_counts}()}
}
\concept{otus}
