% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{n_ranks}
\alias{n_ranks}
\title{Report the number of ranks in an rbiom object.}
\usage{
n_ranks(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
The number of taxonomic ranks present.
}
\description{
Report the number of ranks in an rbiom object.
}
\examples{
    library(rbiom)
    
    n_ranks(hmp50)

}
\seealso{
Other taxonomy: 
\code{\link{otu_taxonomy}()},
\code{\link{taxa_ranks}()},
\code{\link{write_taxonomy}()}
}
\concept{taxonomy}
