% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{n_samples}
\alias{n_samples}
\title{Report the number of samples in an rbiom object.}
\usage{
n_samples(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
The number of samples present.
}
\description{
Report the number of samples in an rbiom object.
}
\examples{
    library(rbiom)
    
    n_samples(hmp50)

}
\seealso{
Other samples: 
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{sample_select}()},
\code{\link{sample_sums}()},
\code{\link{subset.rbiom}()},
\code{\link{within.rbiom}()}
}
\concept{samples}
