% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{nsamples}
\alias{nsamples}
\title{Number of samples in a BIOM.}
\usage{
nsamples(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
The number of samples present.
}
\description{
Number of samples in a BIOM.
}
\examples{
    library(rbiom)
    nsamples(hmp50)

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is.rarefied}()},
\code{\link{metadata}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
