% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate.r
\name{ordinate}
\alias{ordinate}
\title{Reduce a distance matrix to two or three dimensions}
\usage{
ordinate(
  biom,
  ord = "nmds",
  method = "bray-curtis",
  weighted = TRUE,
  tree = NULL,
  md = FALSE,
  k = 2,
  safe = TRUE,
  ...
)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.
Alternatively, a distance matrix, such as from \link{beta.div}.}

\item{ord}{Method for reducing the dimensionality of a distance matrix.
Default: "nmds". Options are:
\describe{
    \item{pcoa}{
      Principal coordinate analysis, via \code{\link[ape]{pcoa}}.
    }
    \item{nmds}{
      Nonmetric multidimensional scaling, via \code{\link[vegan]{metaMDS}}.
    }
    \item{tsne}{
      t-distributed stochastic neighbor embedding, via \code{\link[tsne]{tsne}}.
    }
   }}

\item{method}{The distance algorithm to use. Options are:
\bold{"manhattan"}, \bold{"euclidean"}, 
\bold{"bray-curtis"}, \bold{"jaccard"}, and
\bold{"unifrac"}. Non-ambiguous abbreviations of the method 
names are also accepted. A phylogenetic tree must be present in 
\code{biom} or explicitly provided via \code{tree=} to use the 
UniFrac methods. Ignored if \code{biom} is a distance matrix.
Default: "bray-curtis"}

\item{weighted}{Take relative abundances into account. When 
\code{weighted=FALSE}, only presence/absence is considered.
 Ignored if \code{biom} is a distance matrix.
(Default: \code{TRUE})}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Will be taken from the tree
embedded in the \code{biom} object if not explicitly specified. Only
required for computing UniFrac distance matrices.  Ignored if 
\code{biom} is a distance matrix.}

\item{md}{Include metadata in the output data frame? Ignored if 
\code{biom} is a distance matrix. Options are: 
\describe{
  \item{\code{FALSE}}{ Don't include metadata. (Default) }
  \item{\code{TRUE}}{ Include all metadata. }
  \item{\emph{character vector}}{ Include only the specified metadata columns. }
}}

\item{k}{Number of dimensions to return. (Default: 2)}

\item{safe}{If \code{FALSE} (the default) autogenerated columns be named
"Axis.1", "Axis.2", ..., "Axis.k". If \code{TRUE}, they will be named
".axis.1", ".axis.2", ..., ".axis.k" to avoid conflicting with
metadata column names.}

\item{...}{Additional arguments to pass on to \code{\link[ape]{pcoa}}, 
\code{\link[vegan]{metaMDS}}, or \code{\link[tsne]{tsne}}.}
}
\value{
A data.frame with columns Axis.1, Axis.2, ..., Axis.k as well as columns
        given by \code{md}. Rownames are the sample IDs. For pcoa ordinations, 
        \code{attr(, 'eig')} will contain the eigenvalues useful for construction 
        of ".. % variation explained" labels.
}
\description{
Reduce a distance matrix to two or three dimensions
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom   <- read.biom(infile)
    ord    <- ordinate(biom, "pcoa", "bray-curtis")
    head(ord)
    
}
