% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{otu_names}
\alias{otu_names}
\alias{otu_names<-}
\title{Get or set the OTU names.}
\usage{
otu_names(biom)

otu_names(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{value}{A character vector with the new OTU names.}
}
\value{
A character vector of the taxa names in \code{biom}.
}
\description{
Get or set the OTU names.
}
\examples{
    library(rbiom) 
    
    biom <- hmp50
    
    head(otu_names(biom))
    
    otu_names(biom) <- sub('Unc', 'Uncultured_', otu_names(biom))
    head(otu_names(biom))
}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{metadata_names}()},
\code{\link{otu_matrix}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
