% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{otu_sequences}
\alias{otu_sequences}
\alias{otu_sequences<-}
\title{Get or set the nucleotide sequences associated with each OTU.}
\usage{
otu_sequences(biom)

otu_sequences(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{value}{A file name with fasta-formatted sequences or a named character
vector. Names must match \code{otu_names(biom)}.}
}
\value{
A named character vector of sequences in \code{biom}. If this data
is not present, then returns \code{NULL}.
}
\description{
Get or set the nucleotide sequences associated with each OTU.
}
\examples{
    library(rbiom) 
    
    substr(otu_sequences(hmp50)[1:4], 1, 10)
    
\dontrun{
    # Write to a compressed fasta file in the temporary directory:
    seqs <- otu_sequences(hmp50)
    conn <- bzfile(file.path(tempdir(), "Sequences.fa.bz2"), "w")
    cat(sprintf(">\%s\n\%s", names(seqs), seqs), file=conn, sep="\n")
    close(conn)
    
    # You can also use the write_fasta function for this task:
    write_fasta(hmp50, file.path(tempdir(), "Sequences.fa.gz"))
    
    # Set/replace the sequences
    biom <- read_biom('path/to/file.biom')
    otu_sequences(biom) <- read_fasta('path/to/sequences.fa')
    otu_sequences(biom) <- c(OTU1 = 'ATCGGGTA', OTU2 = 'GGCATTAGC')
}

}
\seealso{
Other sequences: 
\code{\link{has_sequences}()},
\code{\link{write_fasta}()}
}
\concept{sequences}
