% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_taxonomy.r
\name{otu_taxonomy}
\alias{otu_taxonomy}
\alias{otu_taxonomy<-}
\title{Get or set the taxonomy table.}
\usage{
otu_taxonomy(biom, ranks = NULL, unc = "asis")

otu_taxonomy(x) <- value
}
\arguments{
\item{biom, x}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{ranks}{The taxonomic ranks to return in the matrix, or \code{NULL}
for all of them. Default: \code{NULL}}

\item{unc}{How to handle unclassified, uncultured, and similarly ambiguous
taxa names. Default: \code{"asis"}
\itemize{
\item{\code{"asis"} - }{ Don't check/modify any taxa names. }
\item{\code{"singly"} - }{ Replace with "Unc. \if{html}{\out{<OTU ID>}}". }
\item{\code{"grouped"} - }{ Replace with "Unc. \if{html}{\out{<Higher Rank>}}". }
\item{\code{"drop"} - }{ Don't include in the returned matrix. }
}}

\item{value}{A character matrix with rownames \code{otu_names(x)}. If
there are more rownames than taxa names, the matrix will be subset.
May also be a character vector of length one with a file or URL
where the matrix is saved in either comma- or tab-separated format.}
}
\value{
A character matrix with taxa/OTU IDs as row names and taxa ranks as
column names. An 'OTU' column is always added as the last column and
matches the row names.
}
\description{
Get or set the taxonomy table.
}
\examples{
    library(rbiom)
    
    otu_taxonomy(hmp50)[1:4,]
    otu_taxonomy(hmp50, c("Family", "Genus"))[1:4,]
    head(otu_taxonomy(hmp50, "Genus")[,])
    
    # Sometimes taxonomic names are incomplete
    otu_taxonomy(hmp50)[c(53,107,139), 2:6]
    
    # rbiom can insert more descriptive placeholders
    otu_taxonomy(hmp50, unc = "singly")[c(53,107,139), 3:6]
    otu_taxonomy(hmp50, unc = "grouped")[c(53,107,139), 4:6]

}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_tree}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
