% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{otu_tree}
\alias{otu_tree}
\alias{otu_tree<-}
\title{Get or set the phylogenetic tree.}
\usage{
otu_tree(biom)

otu_tree(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{value}{A \code{phylo} class object, file name, or newick string with
leaf names matching \code{otu_names(x)}. If there are more leafs than
taxa names, then the tree will be subset.}
}
\value{
A \code{phylo} class object of the tree in \code{biom}.
}
\description{
Get or set the phylogenetic tree.
}
\examples{
\dontrun{
    library(rbiom)
    
    biom <- hmp50
    otu_tree(biom) <- read_tree('path/to/newick.tre')
}

}
\seealso{
Other phylogeny: 
\code{\link{has_tree}()},
\code{\link{read_tree}()},
\code{\link{tree_data}()},
\code{\link{tree_subset}()},
\code{\link{tree_tips}()},
\code{\link{write_tree}()}
}
\concept{phylogeny}
