% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{phylogeny}
\alias{phylogeny}
\title{Get the phylogenetic tree.}
\usage{
phylogeny(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
A \code{phylo} class object of the tree in \code{biom}.
}
\description{
Get the phylogenetic tree.
}
\examples{
    library(rbiom)
    
    summary(phylogeny(hmp50))

}
\seealso{
Other accessors: 
\code{\link{comments}()},
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
