% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot}
\alias{plot}
\alias{plot.BIOM}
\title{Visualize diversity or abundance as a boxplot, dotplot, etc.}
\usage{
\method{plot}{BIOM}(
  x,
  formula,
  layers = "box",
  color.by = NULL,
  facet.by = NULL,
  stats = TRUE,
  vline = "ci95",
  xlab.angle = "auto",
  ...
)
}
\arguments{
\item{x}{A BIOM object, as returned from \link{read.biom}.}

\item{formula}{Definition of the y-axis metric and metadata column to plot along the x-axis. 
            For example, \bold{Shannon ~ `Body Site`}. Y-axis options are:
\describe{
    \item{Alpha Diversity Metrics (one or more)}{
      \bold{OTUs}, \bold{Shannon}, \bold{Chao1}, \bold{Simpson}, and/or 
      \bold{InvSimpson}. Use \bold{Diversity} to get all five metrics.
    }
    \item{Beta Diversity Metrics (one only)}{
      \bold{manhattan}, \bold{euclidean}, \bold{bray-curtis}, \bold{jaccard}, 
      or \bold{unifrac}. \bold{Distance} will use \bold{unifrac} if a phylogenetic 
      tree is present, or \bold{bray-curtis} otherwise. Use in combination with
      the \code{weighted} parameter. Metadata column names can be prefixed with
      \bold{==} or \bold{!=} to limit distance calculations to \emph{within} or
      \emph{between}, respectively, those categories. See examples below.
    }
    \item{Taxa Abundances (one only)}{
      \bold{Kingdom}, \bold{Phylum}, \bold{Class}, \bold{Order}, \bold{Family}, 
      \bold{Genus}, \bold{Species}, \bold{Strain}, or \bold{OTU}. Supported ranks 
      will vary by biom. Run \code{taxa.ranks(biom)} to see the available options. 
      Specifying \bold{Abundance} will default to the most precise rank possible.
    }
   }}

\item{layers}{What kind of plot to create. Options are \bold{box}, \bold{violin}, \bold{dot}, 
\bold{strip}, \bold{crossbar}, \bold{errorbar}, \bold{linerange}, and/or 
\bold{pointrange}. Single letter abbreviations are also accepted. For instance,
\code{c("box", "dot")} is equivalent to \code{c("b", "d")} and \code{"bd"}.}

\item{color.by}{Metadata column to color by. If that column is a \code{factor}, the ordering of levels
will be maintained in the plot.}

\item{facet.by}{A character vector of one or two metadata columns to facet by. Ordering of \code{factor}
levels will be maintained in the plot.}

\item{stats}{How to display significance brackets and p-values on the plot. Options are: 
\describe{
  \item{\code{TRUE}}{ Auto-select \bold{pairwise} or \bold{group}. (Default) }
  \item{\code{FALSE}}{ Do not display significance values on the plot. }
  \item{\bold{"p"} or \bold{"pairwise"}}{ Compare all against all. }
  \item{\bold{"g"} or \bold{"group"}}{ Compute a single p-value per group. }
  \item{\emph{data.frame}}{ Data to use for statistical annotations. I.e. from \link{stats.table}. }
}}

\item{vline}{How to calculate min/max of the \bold{crossbar}, \bold{errorbar}, \bold{linerange},
and \bold{pointrange} layers. Options are \bold{ci} (confidence interval), \bold{sd} 
(standard deviation), \bold{se} (standard error), and \bold{mad} (median absolute 
deviation). You may optionally append a number to \bold{ci} to specify the confidence
level, for instance \code{vline = "ci95"} (the default), will calculate the 95%
confidence interval, whereas \code{vline = "ci99"} will give the 99% confidence
interval.}

\item{xlab.angle}{How to rotate the tick labels on the x-axis. \bold{'auto'} (the default), 
automatically selects a rotation value. \bold{0}, \bold{30}, and \bold{90} sets the 
angle to horizontal, angled, and vertical, respectively.}

\item{...}{Parameters passed on to ggplot2 functions. Prefixing a parameter name with \bold{b.}, 
\bold{v.}, \bold{d.}, or \bold{s.} forces that parameter to be passed to, and only to,
\bold{geom_boxplot}, \bold{geom_violin}, \bold{geom_dotplot}, or \bold{geom_jitter},
respectively. Otherwise, parameters are passed along by matching against formal arguments.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will be attached as
        \code{attr(p, 'data')} and \code{attr(p, 'stats')}, respectively.
}
\description{
Visualize diversity or abundance as a boxplot, dotplot, etc.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom <- read.biom(infile)
    
    plot(biom, Shannon ~ `Body Site`)
    plot(biom, Shannon ~ Sex, layers="vb", color.by="Body Site")
    
    plot(biom, Diversity ~ `Body Site`, layers="p", color.by="Sex", xlab.angle=30)
    
    #plot(biom, UniFrac ~ `==Body Site`)
    
    plot(biom, bray ~ nmds)
    

}
\seealso{
\code{\link{stats.table}}
}
