% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print}
\alias{print}
\alias{print.BIOM}
\title{Summarize the contents of a BIOM object}
\usage{
\method{print}{BIOM}(x, ...)
}
\arguments{
\item{x}{A BIOM object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{...}{Not used.}
}
\value{
NULL (invisibly)
}
\description{
Summarize the contents of a BIOM object
}
\examples{
    library(rbiom)
    
    print(hmp50)

}
