% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{properties}
\alias{properties}
\alias{has_tree}
\alias{has_sequences}
\alias{is_rarefied}
\title{Check specific properties of a BIOM object.}
\usage{
has_tree(biom)

has_sequences(biom)

is_rarefied(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}
}
\value{
\code{TRUE} if it has the property, \code{FALSE} otherwise.
}
\description{
Check specific properties of a BIOM object.
}
\examples{
    library(rbiom)
    
    has_tree(hmp50)
    has_sequences(hmp50)
    is_rarefied(hmp50)
}
