% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_barplot.r
\name{rare_barplot}
\alias{rare_barplot}
\title{Visualize the number of observations per sample.}
\usage{
rare_barplot(
  biom,
  rline = TRUE,
  counts = TRUE,
  labels = TRUE,
  trans = "log10",
  ...
)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{rline}{Where to draw a horizontal line on the plot, intended to show
a particular rarefaction depth. Set to \code{TRUE} to show an
auto-selected rarefaction depth, \code{FALSE} to not show a line, or
an integer for a custom position.
Default: \code{TRUE}.}

\item{counts}{Display the number of samples and reads remaining after
rarefying to \code{rline} reads per sample. Default: \code{TRUE}.}

\item{labels}{Show sample names under each bar. Default: \code{TRUE}.}

\item{trans}{Y-axis transformation. Options are \code{"log10"} or
\code{NULL}.  Default: \code{"log10"}.}

\item{...}{Additional parameters to pass along to ggplot2 functions.
Prefix a parameter name with \code{r.} to ensure it gets
passed to (and only to) \link[ggplot2]{geom_hline}. For instance,
\code{r.color = "black"} ensures only the horizontal rarefaction line
has its color set to \code{"black"}.}
}
\value{
A \code{ggplot2} plot. \cr The computed data points, ggplot command,
and object history are available as \code{$data}, \code{$code}, and
\code{$history}, respectively.
}
\description{
Visualize the number of observations per sample.
}
\examples{
    library(rbiom)
    
    rare_barplot(hmp50)
    
    rare_barplot(hmp50, rline = 500, r.size = 2, r.linetype = "twodash")
    
    fig <- rare_barplot(hmp50, counts = FALSE)
    fig$code
    
}
\seealso{
Other rarefaction: 
\code{\link{is_rarefied}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_depth}()},
\code{\link{rare_multiplot}()},
\code{\link{rare_suggest}()},
\code{\link{sample_rarefy}()},
\code{\link{sample_sums}()}

Other visualization: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_barplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()}
}
\concept{rarefaction}
\concept{visualization}
