% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_corrplot.r
\name{rare_corrplot}
\alias{rare_corrplot}
\title{Visualize rarefaction curves with scatterplots and trendlines.}
\usage{
rare_corrplot(
  biom,
  metric = "OTUs",
  depths = NULL,
  points = FALSE,
  color.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  ci = 95,
  rline = NULL,
  caption = TRUE,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{metric}{Alpha diversity metric(s) to use. Options are: 
\code{"OTUs"}, \code{"Shannon"}, \code{"Chao1"}, \code{"Simpson"}, 
and/or \code{"InvSimpson"}. Default: \code{"OTUs"}.}

\item{depths}{Rarefaction depths to show in the plot. Passed to
\link{adiv_table}. The default, \code{"multi_even"}, uses a heuristic
to pick 10 evenly spaced depths.}

\item{color.by, facet.by, limit.by}{Metadata columns to use for aesthetics 
and partitioning. See below for details. Default: \code{NULL}}

\item{ci}{The confidence interval to display around the fitted curve. Set
to \code{FALSE} to hide the confidence interval. Default: \code{95}.}

\item{rline}{Where to draw a horizontal line on the plot, intended to show
a particular rarefaction depth. Set to \code{TRUE} to show an 
auto-selected rarefaction depth or \code{NULL} to not show a line.
Default: \code{NULL}.}

\item{caption}{Display information about the method used for trendline
fitting beneath the plot. Default: \code{TRUE}.}

\item{...}{Additional parameters to pass along to ggplot2 
functions. Prefix a parameter name with either \code{p.}, 
\code{r.}/\code{v.}, or \code{s.} to ensure it gets passed to (and
only to)  \link[ggplot2]{geom_point}, \link[ggplot2]{vline}, or 
\link[ggplot2]{geom_smooth}, respectively. For instance, 
\code{p.size = 2} ensures only the points have their size set to 
\code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
}
\description{
Visualize rarefaction curves with scatterplots and trendlines.
}
\section{Aesthetics and Partitions}{


Metadata can be used to flexibly subset, partition, and apply aesthetics 
when creating a plot. Common use cases are provided below. More thorough 
documentation is available at \url{https://cmmr.github.io/rbiom}.

\preformatted{  ## Colors ----------------------------
  color.by = "Body Site"
  color.by = list('Body Site' = "bright")
  color.by = list('Body Site' = c("Stool", "Saliva"))
  color.by = list('Body Site' = list('values' = c("Stool", "Saliva"), 'colors' = "bright"))
  color.by = list('Body Site' = c('Stool' = "blue", 'Saliva' = "green"))
  
  ## Facets ----------------------------
  facet.by = "Body Site"
  facet.by = c("Body Site", "Sex")
  facet.by = list('Body Site' = c("Stool", "Saliva"), "Sex")
  
  ## Limits ----------------------------
  limit.by = list('Sex' = "Male", 'Age' = c(20,40))
  limit.by = list('Body Site' = c("Saliva", "Anterior nares"), 'Age' = c(NA,35))
}

\itemize{
  \item{\code{color.by} - }{Any metadata column. (Max 1)}
  \item{\code{facet.by} - }{Only categorical metadata column(s).}
  \item{\code{limit.by} - }{Any metadata column(s).}
}

All built-in color palettes are colorblind-friendly.

The available categorical palette names are: \code{"okabe"}, \code{"carto"}, 
\code{"r4"}, \code{"polychrome"}, \code{"tol"}, \code{"bright"}, 
\code{"light"}, \code{"muted"}, \code{"vibrant"}, \code{"tableau"}, 
\code{"classic"}, \code{"alphabet"}, \code{"tableau20"}, \code{"kelly"}, 
and \code{"fishy"}.

The available numeric palette names are: \code{"reds"}, \code{"oranges"}, 
\code{"greens"}, \code{"purples"}, \code{"grays"}, \code{"acton"}, 
\code{"bamako"}, \code{"batlow"}, \code{"bilbao"}, \code{"buda"}, 
\code{"davos"}, \code{"devon"}, \code{"grayC"}, \code{"hawaii"}, 
\code{"imola"}, \code{"lajolla"}, \code{"lapaz"}, \code{"nuuk"}, 
\code{"oslo"}, \code{"tokyo"}, \code{"turku"}, \code{"bam"}, 
\code{"berlin"}, \code{"broc"}, \code{"cork"}, \code{"lisbon"}, 
\code{"roma"}, \code{"tofino"}, \code{"vanimo"}, \code{"vik"}
}

\examples{
    library(rbiom)
    
    rare_corrplot(hmp50, color.by="Body Site", metric=c("shannon", "otus"), facet.by="Sex") 
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()},
\code{\link{tree_plot}()}
}
\concept{plotting}
