% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{rare_depth}
\alias{rare_depth}
\title{If the samples are rarefied, report the level (depth).}
\usage{
rare_depth(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
The rarefaction depth, or \code{NULL} if \code{biom} is not
rarefied.
}
\description{
This value is initally set by \code{\link[=read_biom]{read_biom()}} and only subsequently changed by
calls to \code{\link[=sample_rarefy]{sample_rarefy()}}. Calling \code{\link[=as_percent]{as_percent()}} will NOT set
\code{\link[=rare_depth]{rare_depth()}} to \code{1}, even though that function sets all
\code{\link[=sample_sums]{sample_sums()}} to \code{1}.
}
\examples{
    library(rbiom)
    
    rare_depth(hmp50)
    
    biom <- sample_rarefy(hmp50)
    rare_depth(biom)

}
\seealso{
\code{\link[=sample_sums]{sample_sums()}}
}
