% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_multiplot.r
\name{rare_multiplot}
\alias{rare_multiplot}
\title{Combines rare_corrplot and depths_barplot into a single figure.}
\usage{
rare_multiplot(
  biom,
  rline = TRUE,
  metric = "OTUs",
  depths = NULL,
  points = FALSE,
  color.by = NULL,
  facet.by = NULL,
  colors = NULL,
  facets = NULL,
  ci = 95,
  caption = FALSE,
  labels = FALSE,
  trans = "log10",
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{rline}{Where to draw a vertical line on the plot, intended to show
a particular rarefaction depth. Default: \code{TRUE} (no line).}

\item{metric}{Alpha diversity metric(s) to use. Options are: 
\code{"OTUs"}, \code{"Shannon"}, \code{"Chao1"}, \code{"Simpson"}, 
and/or \code{"InvSimpson"}. Default: \code{"OTUs"}.}

\item{depths}{Rarefaction depths to show in the plot. Passed to
\link{adiv_table}. The default, \code{"multi_even"}, uses a heuristic
to pick 10 evenly spaced depths.}

\item{points}{Overlay a scatter plot. Default: \code{FALSE}.}

\item{color.by, facet.by}{Metadata column to color and/or facet by. If that
column is a \code{factor}, the ordering of levels will be maintained 
in the plot. Default: \code{color.by = NULL, facet.by = NULL}}

\item{colors, facets}{Names of the colors and/or facets values to use in
the plot. Available values for colors are given by \code{colors()}. 
Use a named character vector to map them to specific factor levels in
the metadata. \code{facets} are coerced to unnamed character vectors.
If the length of these vectors is less than the values present in 
their corresponding metadata column, then the data set will be 
subseted accordingly. Default: \code{colors = NULL, facets = NULL}}

\item{ci}{The confidence interval to display around the fitted curve. Set
to \code{FALSE} to hide the confidence interval. Default: \code{95}.}

\item{caption}{Display information about the method used for trendline
fitting beneath the plot. Default: \code{FALSE}.}

\item{labels}{Show sample names under each bar. Default: \code{FALSE}.}

\item{trans}{Y-axis transformation. Options are \code{"log10"} or 
\code{NULL}.  Default: \code{"log10"}.}

\item{...}{Additional parameters to pass along to ggplot2 
functions. Prefix a parameter name with either \code{p.}, 
\code{r.}/\code{v.}, or \code{s.} to ensure it gets passed to (and
only to)  \link[ggplot2]{geom_point}, \link[ggplot2]{vline}, or 
\link[ggplot2]{geom_smooth}, respectively. For instance, 
\code{p.size = 2} ensures only the points have their size set to 
\code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
}
\description{
Combines rare_corrplot and depths_barplot into a single figure.
}
\examples{
    library(rbiom)
    
    rare_multiplot(hmp50, color.by="Body Site") 
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()},
\code{\link{tree_plot}()}
}
\concept{plotting}
