% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rarefy.r
\name{rare_suggest}
\alias{rare_suggest}
\title{Suggest a 'good' rarefaction depth.}
\usage{
rare_suggest(biom)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM}
object, as returned from \code{\link[=read_biom]{read_biom()}}. For matrices, the rows and
columns are assumed to be the taxa and samples, respectively.}
}
\value{
An integer.
}
\description{
Suggest a 'good' rarefaction depth.
}
