\docType{package}
\name{rbiom-package}
\alias{rbiom-package}
\title{This is an R package for interfacing with the biom format.}
\description{
  This is an R package for interfacing with the
  biom-format. It includes basic utilities for reading and
  writing BIOM format using R, as well as a formal
  object-oriented representation of BIOM objects that maps
  to functionality in other languages that support
  biom-format, like python.
}
\author{
  Paul J. McMurdie II \email{mcmurdie@stanford.edu}
}
\references{
  \url{http://www.biom-format.org/}
}
\keyword{package}

