% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbiom_objects.r
\name{rbiom_objects}
\alias{rbiom_objects}
\title{Working with rbiom Objects.}
\description{
Rbiom objects make it easy to access and manipulate your BIOM data, ensuring
all the disparate components remain in sync. These objects behave largely
like lists, in that you can access and assign to them using the \code{$}
operator. The sections below list all the fields which can be read and/or
written, and the helper functions for common tasks like rarefying and
subsetting. To create an rbiom object, see \code{\link[=as_rbiom]{as_rbiom()}}.
\cr\cr
}
\section{Readable Fields}{


Reading from fields will not change the rbiom object.\tabular{ll}{
   \strong{Accessor} \tab \strong{Content} \cr
   \verb{$counts} \tab Abundance of each OTU in each sample. \cr
   \verb{$metadata} \tab Sample mappings to metadata (treatment, patient, etc). \cr
   \verb{$taxonomy} \tab OTU mappings to taxonomic ranks (genus, phylum, etc). \cr
   \verb{$otus}, \verb{$n_otus} \tab OTU names. \cr
   \verb{$samples}, \verb{$n_samples} \tab Sample names. \cr
   \verb{$fields}, \verb{$n_fields} \tab Metadata field names. \cr
   \verb{$ranks}, \verb{$n_ranks} \tab Taxonomic rank names. \cr
   \verb{$tree}, \verb{$sequences} \tab Phylogenetic tree / sequences for the OTUs, or \code{NULL}. \cr
   \verb{$id}, \verb{$comment} \tab Arbitrary strings for describing the dataset, or \code{NULL}. \cr
   \verb{$depth} \tab Rarefaction depth, or \code{NULL} if unrarefied. \cr
}


\cr
}

\section{Writable Fields}{


Assigning new values to these components will trigger
validation checks and inter-component synchronization.\tabular{ll}{
   \strong{Component} \tab \strong{What can be assigned.} \cr
   \verb{$counts} \tab matrix of abundances; OTUs (rows) by samples (columns) \cr
   \verb{$metadata} \tab data.frame with \code{'.sample'} as the first column \cr
   \verb{$taxonomy} \tab data.frame with \code{'.otu'} as the first column \cr
   \verb{$tree} \tab phylo object with the phylogenetic tree for OTUs \cr
   \verb{$sequences} \tab character vector of OTU reference sequences \cr
   \verb{$id}, \verb{$comment} \tab string with a title or comment for the dataset \cr
}


\cr
}

\section{Transformations}{
\tabular{ll}{
   \strong{Function} \tab \strong{Transformation} \cr
   \verb{$clone()} \tab Safely duplicate an rbiom object. \cr
   \code{\link[=subset]{subset()}} \tab Subset samples according to metadata properties. \cr
   \code{\link[=select]{select()}} \tab Keep just the specified sample names or indices. \cr
   \code{\link[=rarefy]{rarefy()}} \tab Sub-sample OTU counts to an even sampling depth. \cr
}


All functions return an rbiom object.
\cr\cr
\subsection{Clone an Object}{

Use \verb{$clone()} to create a copy of an rbiom object. This is necessary
because rbiom objects are \strong{passed by reference}. The usual \verb{<-} assignment
operator will simply create a second reference to the same object - it will
not create a second object.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{$clone(deep = FALSE)
}\if{html}{\out{</div>}}
}

\subsection{Arguments}{

\itemize{
\item{\code{deep}}{ Not used. }
}
}

\subsection{Examples}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{biom <- hmp50$clone()
}\if{html}{\out{</div>}}
}

}
}

\keyword{internal}
