% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tree.r
\name{read.tree}
\alias{read.tree}
\title{Read a newick formatted phylogenetic tree.}
\usage{
read.tree(file = NULL, text = NULL)
}
\arguments{
\item{file}{Path to the newick file.}

\item{text}{The newick tree as a string.}
}
\value{
A \code{phylo} class object.
}
\description{
A phylogenetic tree is required for computing UniFrac distance matrices.
You can load a tree either from a file or by providing the tree string
directly.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "newick.tre", package = "rbiom")
    tree <- read.tree(infile)

}
