% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biom.r
\name{read_biom}
\alias{read_biom}
\title{Extracts counts, metadata, taxonomy, and phylogeny from a biom file.}
\usage{
read_biom(src, tree = "auto", cleanup = FALSE, prune = cleanup)
}
\arguments{
\item{src}{Input data as either a file path, URL, or JSON string.
BIOM files can be formatted according to
version 1.0 (JSON) or 2.1 (HDF5)
\href{http://biom-format.org/documentation/}{specifications}, or as
classical tabular format. URLs must begin with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://}. JSON files must
have \code{\{} as their first character. Compressed (gzip or bzip2)
BIOM files are also supported. NOTE: to read HDF5 formatted BIOM
files, the BioConductor R package \code{rhdf5} must be installed.}

\item{tree}{By default, the tree will be read from the BIOM file specified
in \code{src}. Specifying \code{tree=TRUE} will do the same, but will
generate an error message if a tree is not present. Setting
\code{tree=FALSE} will return a \code{BIOM} object without any tree
data. You may also provide a file path, URL, or Newick string to load
that tree data into the returned \code{BIOM} object.
Default: \code{"auto"}}

\item{cleanup}{Renames ambiguous taxons and removes leading underscores.
Also converts character metadata into factors and dates based on
heuristics. Default: \code{FALSE}}

\item{prune}{Should samples and taxa with zero observations be discarded?
Default: \code{cleanup}}
}
\value{
A \code{BIOM} class object containing the parsed data. This object
can be treated as a list with the following named elements:
\itemize{
\item{\code{$counts} - }{
A numeric \code{slam} sparse matrix of observation counts. Taxa
(OTUs) as rows and samples as columns. }
\item{\code{$metadata} - }{
A data frame containing any embedded metadata. Row names are
sample IDs. }
\item{\code{$taxonomy} - }{
Character matrix of taxonomic names, if given. Row names are taxa
(OTU) IDs. Column rows are named Kingdom, Phylum, Class, Order,
Family, Genus, Species, and Strain, or TaxLvl.1, TaxLvl.2, ... ,
TaxLvl.N when more than 8 levels of taxonomy are encoded in the
biom file. }
\item{\code{$phylogeny} - }{
An object of class \code{phylo} defining the phylogenetic
relationships between the taxa. Although the official
specification for BIOM only includes phylogenetic trees in BIOM
version 2.1, if a BIOM version 1.0 file includes a
\code{phylogeny} entry with newick data, then it will be loaded
here as well. The \pkg{ape} package has additional functions for
working with \code{phylo} objects.}
\item{\code{$sequences} - }{
A named character vector, where the names are taxonomic
identifiers and the values are the sequences they represent.
These values are not part of the official BIOM specification, but
will be read and written when defined. }
\item{\code{$info} - }{
A list of other attributes defined in the BIOM file, such as
\code{id}, \code{type}, \code{format}, \code{format_url},
\code{generated_by}, \code{date}, \code{matrix_type},
\code{matrix_element_type}, \code{Comment}, and \code{shape}. }
}
\code{metadata}, \code{taxonomy}, and \code{phylogeny} are optional
components of the BIOM file specification and therefore will be empty
in the returned object when they are not provided by the BIOM file.
}
\description{
Extracts counts, metadata, taxonomy, and phylogeny from a biom file.
}
\examples{
    library(rbiom)

    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom <- read_biom(infile)

    print(biom)

    # Taxa Abundances
    as.matrix(biom$counts[1:4,1:4])

    top5 <- names(head(rev(sort(slam::row_sums(biom$counts))), 5))
    biom$taxonomy[top5,c('Family', 'Genus')]
    as.matrix(biom$counts[top5, 1:6])

    # Metadata
    table(biom$metadata$Sex, biom$metadata$`Body Site`)
    sprintf("Mean age: \%.1f", mean(biom$metadata$Age))

    # Phylogenetic tree
    tree <- biom$phylogeny
    top5.tree <- tree_subset(tree, top5)
    ape::plot.phylo(top5.tree)

}
