% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{.}
\alias{ddply}
\alias{ldply}
\alias{llply}
\alias{\%>\%}
\alias{\%<>\%}
\alias{glimpse}
\alias{na.omit}
\alias{left_join}
\alias{mutate}
\alias{pull}
\alias{relocate}
\alias{rename}
\alias{select}
\alias{slice}
\alias{slice_head}
\alias{slice_tail}
\alias{slice_min}
\alias{slice_max}
\alias{slice_sample}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_head}}, \code{\link[dplyr:slice]{slice_max}}, \code{\link[dplyr:slice]{slice_min}}, \code{\link[dplyr:slice]{slice_sample}}, \code{\link[dplyr:slice]{slice_tail}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}

  \item{pillar}{\code{\link[pillar]{glimpse}}}

  \item{plyr}{\code{\link[plyr:quoted]{.}}, \code{\link[plyr]{ddply}}, \code{\link[plyr]{ldply}}, \code{\link[plyr]{llply}}}

  \item{stats}{\code{\link[stats:na.fail]{na.omit}}}
}}

