% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy.r
\name{rescale}
\alias{rescale}
\alias{rarefy_cols}
\alias{rescale_cols}
\alias{rescale_rows}
\title{Transform a counts matrix.}
\usage{
rarefy_cols(mtx, depth = 0.1, n = NULL, seed = 0)

rescale_cols(mtx)

rescale_rows(mtx)
}
\arguments{
\item{mtx}{A matrix-like object.}

\item{depth}{The number of observations to keep, per column (sample). If
set to \code{'auto'}, a depth will be automatically selected. Samples
that have fewer than this number of observations will be dropped. If
called on data with non-integer abundances, values will be re-scaled
to integers between 1 and \code{depth} such that they sum to \code{depth}.
Default: \code{'auto'}}

\item{n}{The number of samples to retain, or if negative, the number of
samples to drop. If \code{NULL}, the number kept/dropped is automatically
selected. When not \code{NULL}, overrides any value given to \code{depth.}
Default: \code{NULL}}

\item{seed}{An integer to use for seeding the random number generator. If
you need to create different random rarefactions of the same
matrix, set this seed value to a different number each time.}
}
\value{
The rarefied or rescaled matrix.
}
\description{
Rarefaction subset counts so that all samples have the same number of
observations. Rescaling rows or cols scales the matrix values so that row
sums or column sums equal 1.
}
\examples{
    library(rbiom)
    
    # rarefy_cols --------------------------------------
    biom <- hmp50$clone()
    sample_sums(biom) \%>\% head(10)

    biom$counts \%<>\% rarefy_cols(depth=1000)
    sample_sums(biom) \%>\% head(10)
    
    
    # rescaling ----------------------------------------
    mtx <- matrix(sample(1:20), nrow=4)
    mtx
    
    rowSums(mtx)
    rowSums(rescale_rows(mtx))
    
    colSums(mtx)
    colSums(rescale_cols(mtx))

}
\seealso{
Other rarefaction: 
\code{\link{pull.rbiom}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{rarefy}()},
\code{\link{sample_sums}()}

Other transformations: 
\code{\link{rarefy}()},
\code{\link{subset}()}
}
\concept{rarefaction}
\concept{transformations}
