% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{sample.names<-}
\alias{sample.names<-}
\title{Set the sample names.}
\usage{
sample.names(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read.biom}.}

\item{value}{A character vector. A named character vector can be used to
only change some of the sample names.}
}
\description{
Set the sample names.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    
    head(sample.names(biom))
    
    sample.names(biom) <- sub('HMP', 'Sample_', sample.names(biom))
    head(sample.names(biom))
    
    sample.names(biom) <- c('Sample_02' = 'One', 'Sample_03' = 'Two')
    head(sample.names(biom))

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa.names<-}()},
\code{\link{taxa.ranks<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
