% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{sample.sums}
\alias{sample.sums}
\title{Sum the observations in each sample.}
\usage{
sample.sums(biom, long = FALSE, md = FALSE)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
A named numeric vector of the number of observations in each 
        sample. The names are the sample IDs.
}
\description{
Sum the observations in each sample.
}
\examples{
    library(rbiom)
    
    sample.sums(hmp50) \%>\% sort() \%>\% head()
    
    hist(sample.sums(hmp50))

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is.rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
