% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.r
\name{sample_metadata}
\alias{sample_metadata}
\alias{sample_metadata<-}
\title{Get or set the sample metadata.}
\usage{
sample_metadata(biom, field = NULL, id = NULL, cleanup = FALSE)

sample_metadata(x) <- value
}
\arguments{
\item{biom, x}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{field}{The name of a single metadata column to retrieve. If provided,
a named vector will be returned instead of a data.frame.
Default: \code{NULL}}

\item{id}{Copy the sample names (rownames) into this column.
Default: \code{NULL}}

\item{cleanup}{Change character columns to factor or dates. Date formats
recognized are: "\%Y-\%m-\%d" or "\%m/\%d/\%Y" (four-digit year). Also
replace '.' with '_' in column names when '.' is the first character
in order to avoid conflicts with auto-generated columns.
Default: \code{FALSE}}

\item{value}{A data.frame with the metadata. All the \code{rownames()} must
be in \code{sample_names(biom)}. If there are fewer rows of
\code{value} than samples, then the \code{biom} object will be subset.}
}
\value{
A data frame of the metadata in \code{biom}. If \code{field} is
given, will return a named vector of the field's values.
}
\description{
Get or set the sample metadata.
}
\examples{
    library(rbiom)
    
    sample_metadata(hmp50)[1:4,1:3]
    
    head(sample_metadata(hmp50, "Body Site"))
    
    head(sample_metadata(hmp50, "Body Site", cleanup=TRUE))
    
    md <- sample_metadata(hmp50)
    md <- md[,c('Sex', 'Body Site')]
    sample_metadata(hmp50) <- md
    head(sample_metadata(hmp50))

}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
