% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{sample_names<-}
\alias{sample_names<-}
\title{Set the sample names.}
\usage{
sample_names(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read_biom}.}

\item{value}{A character vector. A named character vector can be used to
only change some of the sample names.}
}
\description{
Set the sample names.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    
    head(sample_names(biom))
    
    sample_names(biom) <- sub('HMP', 'Sample_', sample_names(biom))
    head(sample_names(biom))
    
    sample_names(biom) <- c('Sample_02' = 'One', 'Sample_03' = 'Two')
    head(sample_names(biom))

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa_names<-}()},
\code{\link{taxa_ranks<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
