% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{sample_names}
\alias{sample_names}
\alias{sample_names<-}
\title{Get or set the sample names.}
\usage{
sample_names(biom)

sample_names(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{value}{A character vector of the new sample names.}
}
\value{
A character vector of the sample names in \code{biom}.
}
\description{
Get or set the sample names.
}
\examples{
    library(rbiom) 
    
    biom <- hmp50
    
    head(sample_names(biom))
    
    sample_names(biom) <- sub('HMP', 'Sample_', sample_names(biom))
    head(sample_names(biom))
}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{metadata_names}()},
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_metadata}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
