% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rarefy.r
\name{sample_rarefy}
\alias{sample_rarefy}
\title{Subset counts so that all samples have the same number of observations.}
\usage{
sample_rarefy(biom, depth = NULL, seed = 0)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{depth}{The number of observations to keep, per sample. If set to
\code{NULL}, a depth will be automatically selected. Samples that
have fewer than this number of observations will be dropped. If
called on data with non-integer abundances, values will be re-scaled
to integers between 1 and \code{depth} such that they sum to
\code{depth}.}

\item{seed}{An integer to use for seeding the random number generator. If
you need to create different random rarefactions of the same
\code{rbiom} object, set this seed value to a different number each
time.}
}
\value{
The rarefied \code{rbiom} object.
}
\description{
Subset counts so that all samples have the same number of observations.
}
\examples{
    library(rbiom)

    range(sample_sums(hmp50))

    biom <- sample_rarefy(hmp50, depth=1000)
    range(sample_sums(biom))

}
\seealso{
Other rarefaction: 
\code{\link{is_rarefied}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_depth}()},
\code{\link{rare_multiplot}()},
\code{\link{rare_suggest}()},
\code{\link{sample_sums}()}
}
\concept{rarefaction}
