% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_select.r
\name{sample_select}
\alias{sample_select}
\title{Reduce samples to a specific list}
\usage{
sample_select(biom, samples = NULL, nTop = NULL, nRandom = NULL, seed = 0)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{samples}{Sample names, indices, or a logical vector identifying
the samples to keep. The latter two should be based on the order of
sample names given by \code{sample_names(biom)}.}

\item{nTop}{Selects this number of samples, taking the sample with the most
observations first, then the sample with the second-most observations,
etc. Ties will be randomly ordered. If \code{nTop} is higher than the
number of samples in the dataset, the entire dataset will be returned.
See note.}

\item{nRandom}{Randomly selects this number of samples. If higher than the
number of samples in the dataset, the entire dataset will be returned.
See note.}

\item{seed}{Random seed, used when selecting \code{nRandom} samples.

Note: Generally, you will specify only one of the filters: \code{samples},
\code{nTop}, or \code{nRandom}. However, specifying multiple filters is
allowed; they will be applied in the order listed above.}
}
\value{
A \code{rbiom} object, subsetted as requested.
}
\description{
Reduce samples to a specific list
}
\examples{
  \dontrun{
    library(rbiom) 
    
    ex1 <- sample_select(hmp50, c('HMP14', 'HMP22', 'HMP03'))
    ex2 <- sample_select(hmp50, c(32, 11, 28, 16, 46, 5))
    ex3 <- sample_select(hmp50, 1:50 \%\% 6 == 0)
    ex4 <- sample_select(hmp50, nRandom = 10)
    ex5 <- sample_select(hmp50, nTop = 5)
    ex6 <- sample_select(hmp50, samples = 10:40, nTop = 20, nRandom = 10)
 }
}
\seealso{
Other metadata: 
\code{\link{bdply}()},
\code{\link{blply}()},
\code{\link{metadata_names}()},
\code{\link{metadata_numeric}()},
\code{\link{n_metadata}()},
\code{\link{sample_metadata}()},
\code{\link{subset.rbiom}()},
\code{\link{within.rbiom}()},
\code{\link{write_metadata}()}

Other samples: 
\code{\link{n_samples}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{subset.rbiom}()},
\code{\link{within.rbiom}()}
}
\concept{metadata}
\concept{samples}
