% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{sample_sums}
\alias{sample_sums}
\title{Sum the observations in each sample.}
\usage{
sample_sums(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
A named numeric vector of the number of observations in each
sample. The names are the sample IDs.
}
\description{
Sum the observations in each sample.
}
\examples{
    library(rbiom)
    
    sample_sums(hmp50) \%>\% sort() \%>\% head()
    
    hist(sample_sums(hmp50))
}
\seealso{
\code{\link[=rare_depth]{rare_depth()}}, \code{\link[=rare_barplot]{rare_barplot()}}

Other accessors: 
\code{\link{biom_info}()},
\code{\link{metadata_numeric}()},
\code{\link{n_samples}()}
}
\concept{accessors}
