% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.r
\name{select}
\alias{select}
\title{Reduce samples to a specific list}
\usage{
select(biom, samples)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read.biom}.}

\item{samples}{Sample names, indices, or a logical vector identifying
the samples to keep. The latter two should be based on the order of
sample names given by \code{colnames(biom$counts)}.}
}
\value{
A \code{BIOM} object.
}
\description{
Reduce samples to a specific list
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    
    ex1 <- select(biom, c('HMP14', 'HMP22', 'HMP03'))
    ex2 <- select(biom, c(32, 11, 28, 16, 46, 5))
    ex3 <- select(biom, 1:50 \%\% 6 == 0)

}
\seealso{
\code{\link{subset}}
}
