% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{sequences<-}
\alias{sequences<-}
\title{Set nucleotide sequences associated with each taxonomic identifier.}
\usage{
sequences(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read.biom}.}

\item{value}{A named character vector. Names must match \code{taxa.names(x)}.}
}
\description{
Set nucleotide sequences associated with each taxonomic identifier.
}
\examples{
\dontrun{
    library(rbiom)
    
    sequences(biom) <- read.fasta('path/to/sequences.fa')
    sequences(biom) <- c(OTU1 = 'ATCGGGTA', OTU2 = 'GGCATTAGC')
}

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sample.names<-}()},
\code{\link{taxa.names<-}()},
\code{\link{taxa.ranks<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
