% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{sequences}
\alias{sequences}
\title{Nucleotide sequences associated with each taxonomic identifier.}
\usage{
sequences(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
A named character vector of sequences in \code{biom}. If this data
          is not present, then returns \code{NULL}.
}
\description{
Nucleotide sequences associated with each taxonomic identifier.
}
\examples{
    library(rbiom)
    
    sequences(hmp50)[1:4]
    
    # Write to a compressed fasta file in the temporary directory:
    seqs <- sequences(hmp50)
    conn <- bzfile(file.path(tempdir(), "Sequences.fa.bz2"), "w")
    cat(sprintf(">\%s\n\%s", names(seqs), seqs), file=conn, sep="\n")
    close(conn)
    
    # You can also use the write_fasta function for this task:
    write_fasta(hmp50, file.path(tempdir(), "Sequences.fa.gz"))

}
\seealso{
Other accessors: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
