% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.r
\name{slice}
\alias{slice}
\alias{slice.rbiom}
\alias{slice_head.rbiom}
\alias{slice_tail.rbiom}
\alias{slice_min.rbiom}
\alias{slice_max.rbiom}
\alias{slice_sample.rbiom}
\title{Subset to a specific number of samples.}
\usage{
\method{slice}{rbiom}(biom, ..., clone = TRUE)

\method{slice_head}{rbiom}(biom, ..., clone = TRUE)

\method{slice_tail}{rbiom}(biom, ..., clone = TRUE)

\method{slice_min}{rbiom}(biom, ..., clone = TRUE)

\method{slice_max}{rbiom}(biom, ..., clone = TRUE)

\method{slice_sample}{rbiom}(biom, ..., clone = TRUE)
}
\arguments{
\item{biom}{An \link[=rbiom_objects]{rbiom object}, such as from \code{\link[=as_rbiom]{as_rbiom()}}.}

\item{...}{Passed on to \code{\link[dplyr:slice]{dplyr::slice()}}.}

\item{clone}{Create a copy of \code{biom} before modifying. If \code{FALSE}, \code{biom}
is modified in place as a side-effect. See \link[=speed_ups]{speed ups} for
use cases. Default: \code{TRUE}}
}
\value{
An \link[=rbiom_objects]{rbiom object}.
}
\description{
Subset to a specific number of samples.
}
\examples{
    library(rbiom)
    
    # The last 3 samples in the metadata table.
    biom <- slice_tail(hmp50, n = 3)
    biom$metadata
    
    # The 3 oldest subjects sampled.
    biom <- slice_max(hmp50, Age, n = 3)
    biom$metadata
    
    # Pick 3 samples at random.
    biom <- slice_sample(hmp50, n = 3)
    biom$metadata

}
\seealso{
Other transformations: 
\code{\link{mutate}()},
\code{\link{rarefy_cols}()},
\code{\link{rarefy}()},
\code{\link{subset}()}
}
\concept{transformations}
