% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_plot.r
\name{stacked_plot}
\alias{stacked_plot}
\title{Display taxa abundances as a stacked bar graph.}
\usage{
stacked_plot(
  biom,
  rank = NULL,
  taxa = 6,
  facet.by = NULL,
  facets = NULL,
  colors = NULL,
  patterns = FALSE,
  label.by = NULL,
  sort.by = NULL,
  dist = "euclidean",
  clust = "complete",
  other = TRUE,
  xlab.angle = 90,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read.biom}.}

\item{rank}{What rank(s) of taxa to display, for example \code{"Phylum"} 
or \code{"Genus"}. Run \code{taxa.ranks()} to see all options for a 
given BIOM object. The default, \code{NULL}, selects the lowest
level.}

\item{taxa}{Which taxa to give separate colors. An integer value will show
the top n most abundant taxa. A value 0 <= n < 1 will show all taxa 
with that mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. Default: \code{6}.}

\item{facet.by}{Split into multiple plots based on this metadata 
column(s). Default: \code{NULL}.}

\item{facets}{Optionally specify which facets to show and in what order. 
Default: \code{NULL}.}

\item{colors, patterns}{A character vector of colors or patterns to use in
the graph. A named character vector can be used to map taxon names to 
specific colors or patterns. Set to \code{NULL} or \code{TRUE} to
automatically select. Set to \code{FALSE} to disable per-taxa colors
or patterns. Default: \code{colors=NULL, patterns=isFALSE(colors)}.}

\item{label.by, sort.by}{What metadata column to use for labeling and/or
sorting the samples across the x-axis. When \code{sort.by=NULL},
samples are arranged based on \code{dist} and \code{clust}, below.
Default: \code{label.by=NULL, sort.by=NULL}.}

\item{dist, clust}{Distance (\link{dist}) and clustering (\link{hclust})
methods to use for automatically arranging samples along the x-axis 
to group samples with similar composition near one another.
Default: \code{dist="euclidean", clust="complete"}.}

\item{other}{Add an 'Other' taxa that ensures all bars sum to 100%. 
Default: \code{TRUE}.}

\item{xlab.angle}{How to rotate the tick labels on the x-axis.
Options are \code{90} (the default) or \code{30}.}

\item{...}{Parameters for underlying functions. Prefixing parameter names 
with a layer name ensures that a particular parameter is passed to, 
and only to, that layer.}
}
\value{
A \code{ggplot2} plot. The computed data points will be attached as 
        \code{attr(, 'data')}.
}
\description{
Display taxa abundances as a stacked bar graph.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    stacked_plot(biom, rank="Phylum")
    
}
