% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.table.r
\name{stats.table}
\alias{stats.table}
\title{Compute p-values for data vs metadata.}
\usage{
stats.table(
  biom,
  x,
  y,
  by = NULL,
  adj = "fdr",
  pairwise = FALSE,
  weighted = TRUE,
  digits = 3,
  y.pos = NULL
)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}. Alternatively,
a data.frame with column names expected by \code{x}, \code{y}, and \code{by}.}

\item{x}{The metadata column name of the category to use for comparisons.}

\item{y}{A \emph{character vector} with values from one of the following sets:
\describe{
    \item{Alpha Diversity Metrics (one or more)}{
      \bold{OTUs}, \bold{Shannon}, \bold{Chao1}, \bold{Simpson}, and/or \bold{InvSimpson}.
      You may also set \code{y = "Diversity"} to get all five metrics.
    }
    \item{Beta Diversity Metrics (one only)}{
      \bold{manhattan}, \bold{euclidean}, \bold{bray-curtis}, \bold{jaccard}, 
      or \bold{unifrac}. Use in combination with the \code{weighted} parameter.
      Metadata column names can be prefixed with \bold{==} or \bold{!=} to limit distance
      calculations to \emph{within} or \emph{between}, respectively, those categories. See
      examples below. Setting \code{y = "Distance"} will use \bold{unifrac} if a 
      phylogenetic tree is present, or \bold{bray-curtis} otherwise.
    }
    \item{Taxa Abundances (one only)}{
      \bold{Kingdom}, \bold{Phylum}, \bold{Class}, \bold{Order}, \bold{Family}, \bold{Genus}, 
      \bold{Species}, \bold{Strain}, or \bold{OTU}. Supported ranks will vary by biom. Run
      \code{taxa.ranks(biom)} to see the available options. Specifying \code{y = "Taxa"}
      will default to the most precise rank possible.
    }
   }}

\item{by}{Additional metadata columns to group by (e.g. facets in a plot).}

\item{adj}{Adjustment to use for multiple comparisons. See \code{stats::p.adjust.methods} 
for available options. Default: \bold{fdr}.}

\item{pairwise}{If \bold{FALSE} (the default), one p-value is generated for each \code{by} 
group, or 1 p-value when \code{by = NULL}. If \bold{TRUE}, unique values in
\code{x} are compared pairwise against each other.}

\item{weighted}{If \bold{TRUE}(the default), run beta diversity metrics in weighted mode. 
Set to \bold{FALSE} to use unweighted mode.}

\item{digits}{Round p-values to this many significant figures (default 3). 
Set to \bold{NULL} to disable rounding.}

\item{y.pos}{Add an extra column to the results to indicate the upper most value of each group.
Added for ggpubr::geom_bracket(). Options are \bold{NULL} for no y.pos column 
(the default), \bold{max} for the maximum value, \bold{box} for a box plot's the 
whisker upper bound, or \bold{violin} for the highest point of a violin plot.}
}
\value{
A data.frame with columns \bold{p.val}, \bold{adj.p}, \bold{adj.sig}, as well as
                columns for tracking for \code{x}, \code{y}, and \code{by} categories.
}
\description{
Compute p-values for data vs metadata.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom   <- rarefy(read.biom(infile))
    biom   <- subset(biom, `Body Site` \%in\% c('Saliva', 'Stool', 'Buccal mucosa'))
      
    
    # Overall, do the three body sites different have different Shannon
    #  diversity indices (alpha diversity)? - YES
    stats.table(biom, x = "Body Site", y = "Shannon")
    
    # Which body sites have the most different Shannon diversity
    #  indices? - Buccal mucosa vs. Saliva
    stats.table(biom, x = "Body Site", y = "Shannon", pairwise = TRUE)
    
    # Do males or females have more alpha diversity variation among 
    #  body sites? - Females
    stats.table(biom, x = "Body Site", y = "Shannon", by = "Sex")
    
    # Which phylum is most different overall between males and females?
    #  - Saccharibacteria
    stats.table(biom, x = "Sex", y = "Phylum") \%>\% head()
    
    # Which phyla is the most differentially abundant between males and 
    #  females on a particular body site? - Tenericutes in Saliva
    stats.table(biom, x = "Sex", y = "Phylum", by = "Body Site") \%>\% head()
    
    # Overall, are males and females characterized by their OTU 
    #  abundances? - NO
    stats.table(biom, x = "Sex", y = "unifrac")
    
    # What about on a particular body site? - Yes, Buccal mucosa.
    stats.table(biom, x = "==Sex", y = "unifrac", by = "==Body Site")
    
    # Recall that for distance metrics such as unifrac, the value of
    #  interest (the distance) is calculated from a pair of samples.
    #  Sometimes both samples will be from males, sometimes both from
    #  females, and sometimes one from each.
    # The '==' prefix on '==Sex' and '==Body Site' indicates that only
    #  pairs from the same sex and same body site should be considered.
    # Using the '!=' prefix does the opposite - considering only sample
    #  pairs from different sexes and body sites.
    
    # Removing the '==' prefix generates the complete set of 
    #  comparisons. Note how the test changes from Mann-Whitney to
    #  Kruskal-Wallis. This is because instead of a two group comparison
    #  ('Male', 'Female') we are now comparing three groups 
    #  ('Male', 'Female', 'Female vs Male').
    stats.table(biom, x = "Sex", y = "unifrac", by = "Body Site")

}
