% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_table.r
\name{stats_table}
\alias{stats_table}
\title{Compute p-values for data vs categorical metadata.}
\usage{
stats_table(
  biom,
  x,
  y,
  by = NULL,
  adj = "fdr",
  pairwise = FALSE,
  weighted = TRUE,
  digits = 3,
  y.pos = NULL,
  y.pos.facet = "Metric",
  adj.n = NULL
)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}. 
Alternatively, a data.frame with column names expected by \code{x}, 
\code{y}, and \code{by}.}

\item{x}{A CATEGORICAL metadata column name to use for comparisons.}

\item{y}{A \emph{character vector} with a value from one of the following 
sets:
\describe{
    \item{Alpha Diversity Metric}{
      \bold{OTUs}, \bold{Shannon}, \bold{Chao1}, \bold{Simpson}, 
      and/or \bold{InvSimpson}.
    }
    \item{Beta Diversity Metric}{
      \bold{manhattan}, \bold{euclidean}, \bold{bray-curtis}, 
      \bold{jaccard}, or \bold{unifrac}. Use in combination with the
      \code{weighted} parameter. Metadata column names can be 
      prefixed with \bold{==} or \bold{!=} to limit distance
      calculations to \emph{within} or \emph{between}, respectively, 
      those categories. See examples below.
    }
    \item{Taxonomic Rank}{
      \bold{Kingdom}, \bold{Phylum}, \bold{Class}, \bold{Order}, 
      \bold{Family}, \bold{Genus}, \bold{Species}, \bold{Strain}, or 
      \bold{OTU}. Supported ranks will vary by biom. Run
      \code{taxa_ranks(biom)} to see the available options.
    }
   }}

\item{by}{Additional metadata columns to group by (e.g. facets in a plot).}

\item{adj}{Adjustment to use for multiple comparisons. See 
\code{stats::p.adjust.methods} for available options. 
Default: \bold{fdr}.}

\item{pairwise}{If \bold{FALSE} (the default), one p-value is generated 
for each \code{by} group, or 1 p-value when \code{by = NULL}. If 
\bold{TRUE}, unique values in \code{x} are compared pairwise against 
each other.}

\item{weighted}{If \bold{TRUE}(the default), run beta diversity metrics in 
weighted mode. Set to \bold{FALSE} to use unweighted mode.}

\item{digits}{Round p-values to this many significant figures (default 3). 
Set to \bold{NULL} to disable rounding.}

\item{y.pos}{Add an extra column to the results to indicate the upper most 
value of each group. Added for ggpubr::geom_bracket(). Options are 
\bold{NULL} for no y.pos column (the default), \bold{max} for the 
maximum value, \bold{box} for a box plot's the whisker upper bound, 
or \bold{violin} for the highest point of a violin plot.}

\item{adj.n}{Manually set the number of comparisons to correct for. Only
set this if you know what you are doing! Default: \code{NULL}}
}
\value{
A data.frame with columns \bold{.p.val} and \bold{.adj.p}, as well 
        as columns for tracking \code{x}, \code{y}, and \code{by} values.
}
\description{
Compute p-values for data vs categorical metadata.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    biom <- subset(biom, `Body Site` \%in\% c('Saliva', 'Stool', 'Buccal mucosa'))
      
    
    # Overall, do the three body sites different have different Shannon
    #  diversity indices (alpha diversity)? - YES
    stats_table(biom, x = "Body Site", y = "Shannon")
    
    # Which body sites have the most different Shannon diversity
    #  indices? - Buccal mucosa vs. Saliva
    stats_table(biom, x = "Body Site", y = "Shannon", pairwise = TRUE)
    
    # Do males or females have more alpha diversity variation among 
    #  body sites? - Females
    stats_table(biom, x = "Body Site", y = "Shannon", by = "Sex")
    
    # Which phylum is most different overall between males and females?
    #  - Saccharibacteria
    stats_table(biom, x = "Sex", y = "Phylum") \%>\% head()
    
    # Which phyla is the most differentially abundant between males and 
    #  females on a particular body site? - Tenericutes in Saliva
    stats_table(biom, x = "Sex", y = "Phylum", by = "Body Site") \%>\% head()
    
    # Overall, are males and females characterized by their OTU 
    #  abundances? - NO
    stats_table(biom, x = "Sex", y = "unifrac")
    
    # What about on a particular body site? - Yes, Buccal mucosa.
    stats_table(biom, x = "==Sex", y = "unifrac", by = "==Body Site")
    
    # Recall that for distance metrics such as unifrac, the value of
    #  interest (the distance) is calculated from a pair of samples.
    #  Sometimes both samples will be from males, sometimes both from
    #  females, and sometimes one from each.
    # The '==' prefix on '==Sex' and '==Body Site' indicates that only
    #  pairs from the same sex and same body site should be considered.
    # Using the '!=' prefix does the opposite - considering only sample
    #  pairs from different sexes and body sites.
    
    # Removing the '==' prefix generates the complete set of 
    #  comparisons. Note how the test changes from Mann-Whitney to
    #  Kruskal-Wallis. This is because instead of a two group comparison
    #  ('Male', 'Female') we are now comparing three groups 
    #  ('Male', 'Female', 'Female vs Male').
    stats_table(biom, x = "Sex", y = "unifrac", by = "Body Site")

}
