% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.r
\name{subsample}
\alias{subsample}
\title{Provides a table of sample diversity values at multiple read depths.}
\usage{
subsample(biom, x = 10, progressbar = FALSE)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read.biom}.}

\item{x}{A numeric vector giving the depths to rarefy at. If the vector
has only a length of one, the value is taken to mean the number of
different depths to automatically select.}

\item{progressbar}{Whether to display a progress bar and status messages
(TRUE/FALSE). Will automatically tie in with \pkg{shiny} if run within a
\pkg{shiny} session.}
}
\value{
A \code{data frame} with column names \bold{Sample}, 
    \bold{Depth}, \bold{OTUs}, \bold{Shannon}, \bold{Simpson}, and 
    \bold{Chao1}.
}
\description{
A useful function for making rarefaction curve plots.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    rc <- subsample(biom, x=5)
    head(rc)

}
