% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.r
\name{subset}
\alias{subset}
\alias{subset.rbiom}
\alias{slice.rbiom}
\alias{mutate.rbiom}
\title{Transform an rbiom object by its metadata.}
\usage{
\method{subset}{rbiom}(biom, ..., clone = TRUE)

\method{slice}{rbiom}(biom, ..., clone = TRUE)

\method{mutate}{rbiom}(biom, ..., clone = TRUE)
}
\arguments{
\item{biom}{An \link[=rbiom_objects]{rbiom object}, such as from \code{\link[=as_rbiom]{as_rbiom()}}.}

\item{...}{Passed on to \code{\link[base:subset]{base::subset()}}, \code{\link[dplyr:slice]{dplyr::slice()}}, or \code{\link[dplyr:mutate]{dplyr::mutate()}}.}

\item{clone}{Create a copy of \code{biom} before modifying. If \code{FALSE}, \code{biom}
is modified in place as a side-effect. See \link{cloning} for use cases.
Default: \code{TRUE}}
}
\value{
An \link[=rbiom_objects]{rbiom object}.
}
\description{
Transform an rbiom object by its metadata.
}
\examples{
    library(rbiom)
    
    subset(hmp50, `Body Site` \%in\% 'Saliva' & Age < 25)$metadata
    
    slice(hmp50, 1:4)$metadata
    
    mutate(hmp50, Obsese = BMI >= 30)$metadata

}
\seealso{
Other transformations: 
\code{\link{rarefy}()},
\code{\link{rescale}}
}
\concept{transformations}
