% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{summary}
\alias{summary}
\alias{summary.BIOM}
\title{Summarize the contents of a BIOM object}
\usage{
\method{summary}{BIOM}(object, ...)
}
\arguments{
\item{object}{A BIOM object, as returned from \link{read.biom}.}

\item{...}{Not used.}
}
\value{
A distance matrix.
}
\description{
Summarize the contents of a BIOM object
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    
    summary(biom)

}
