% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.sums.r
\name{taxa.means}
\alias{taxa.means}
\title{Get the taxa abundances.}
\usage{
taxa.means(biom, rank = NULL)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}

\item{rank}{The taxonomic rank to return means for. The default, 
\code{NULL}, return OTU means in the same order as they appear in 
\link{counts}. If not \code{NULL}, returned means are ordered from
most abundance to least abundant.}
}
\value{
A numeric vector, named with the taxa names.
}
\description{
Get the taxa abundances.
}
\examples{
    library(rbiom)
    
    taxa.means(hmp50, 'Genus') \%>\% head(10)

}
