% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{taxa.names<-}
\alias{taxa.names<-}
\title{Set the taxa names.}
\usage{
taxa.names(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read.biom}.}

\item{value}{A character vector. A named character vector can be used to
only change some of the taxa names.}
}
\description{
Set the taxa names.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    
    head(taxa.names(biom))
    
    taxa.names(biom) <- sub('Unc', 'Uncultured_', taxa.names(biom))
    head(taxa.names(biom))
    
    taxa.names(biom) <- c('AnmMass2' = 'One', 'PreBivi6' = 'Two')
    head(taxa.names(biom))

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sample.names<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa.ranks<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
