% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{taxa.ranks<-}
\alias{taxa.ranks<-}
\title{Set the taxa rank names.}
\usage{
taxa.ranks(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read.biom}.}

\item{value}{A character vector. A named character vector can be used to
only change some of the taxa rank names.}
}
\description{
Set the taxa rank names.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    
    taxa.ranks(biom)
    
    taxa.ranks(biom) <- c("OTU" = "ASV")
    taxa.ranks(biom)
    
    taxa.ranks(biom) <- paste0("Level", 1:7)
    taxa.ranks(biom)

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sample.names<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa.names<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
