% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{taxa.ranks}
\alias{taxa.ranks}
\title{Get the taxa ranks.}
\usage{
taxa.ranks(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
A character vector of the taxa ranks in \code{biom}.
}
\description{
Get the taxa ranks.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    
    taxa.ranks(biom)

}
\seealso{
\code{\link{sample.names}}, \code{\link{taxa.names}}, 
    \code{\link{taxa.ranks}}, \code{\link{counts}}, 
    \code{\link{taxonomy}}, \code{\link{phylogeny}},
    \code{\link{metadata}}, \code{\link{info}}

Other accessor functions: \code{\link{counts}},
  \code{\link{info}}, \code{\link{metadata}},
  \code{\link{phylogeny}}, \code{\link{sample.names}},
  \code{\link{taxa.names}}, \code{\link{taxonomy}}
}
