% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.rollup.r
\name{taxa.rollup}
\alias{taxa.rollup}
\title{Generate a matrix of samples by taxa, at the specified taxonomic rank.}
\usage{
taxa.rollup(biom, rank = "OTU", lineage = FALSE, sparse = FALSE,
  progressbar = FALSE)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read.biom}.}

\item{rank}{The taxonomic rank. E.g. \bold{\dQuote{OTU}}, 
\bold{\dQuote{Phylum}}, etc. May also be given numerically: 0 for
OTU, 1 for the highest level (i.e. Kingdom), and extending to the number
of taxonomic ranks encoded in the original biom file. See example below
to fetch the names of all available ranks.}

\item{lineage}{Include all ranks in the name of the taxa. For instance,
setting to \code{TRUE} will produce 
\code{Bacteria; Actinobacteria; Coriobacteriia; Coriobacteriales}. 
Whereas setting to \code{FALSE} (the default) will return simply
\code{Coriobacteriales}. You want to set this to TRUE if you have
genus names (such as \emph{Incertae_Sedis}) that map to multiple higher
level ranks.}

\item{sparse}{If true, returns a sparse matrix as described by 
\link[slam]{simple_triplet_matrix}, otherwise returns a normal R
matrix object. Sparse matrices will likely be considerably more
memory efficient in this scenario.}

\item{progressbar}{Whether to display a progress bar and status messages
(logical). Will automatically tie in with \pkg{shiny} if run within a
\pkg{shiny} session.}
}
\value{
A numeric matrix with samples as column names, and taxonomic
    identifiers as row names.
}
\description{
Generate a matrix of samples by taxa, at the specified taxonomic rank.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    
    colnames(biom$taxonomy)
    
    phyla <- taxa.rollup(biom, 'Phylum')
    phyla[1:4,1:6]

}
