% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_rollup.r
\name{taxa_matrix}
\alias{taxa_matrix}
\title{Taxa abundances per sample, at the specified taxonomic rank.}
\usage{
taxa_matrix(
  biom,
  rank = "OTU",
  taxa = NULL,
  map = NULL,
  lineage = FALSE,
  sparse = FALSE,
  unc = "singly",
  other = FALSE
)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM}
object, as returned from \code{\link[=read_biom]{read_biom()}}. For matrices, the rows and
columns are assumed to be the taxa and samples, respectively.}

\item{rank}{The taxonomic rank. E.g. \code{rank = "OTU"},
\code{rank = "Phylum"}, etc. May also be given numerically: 0 for
OTU, 1 for the highest level (i.e. Kingdom), and extending to the number
of taxonomic ranks encoded in the original biom file. Run
\code{taxa_ranks(biom)} to fetch the names of all available ranks.}

\item{taxa}{Limit the number of taxa returned in the matrix. Depending on
the value type, a different filter is applied.
\itemize{
\item{\code{NULL} - }{ Retain all taxa. }
\item{\emph{integer} >= 1 - }{
The N most abundant taxa are retained. \cr
Example: \code{taxa_matrix(biom, 'Genus', 5)} returns the five
most abundant genera. }
\item{\emph{numeric} between 0 and 1 - }{
Retain taxa if their abundance is greater than or equal to N. \cr
Example: \code{taxa_matrix(biom, 'Phylum', 0.1)} returns all
phyla with a relative abundance of at least 10 percent. }
\item{\emph{character vector} - }{
Only these taxa names are retained. \cr
Example:
\code{taxa_matrix(biom, 'Phylum', c("Firmicutes", "Bacteroidetes"))}
will only retain those two phyla. }
}
Default: \code{NULL}}

\item{map}{A character matrix defining the value that each taxa IDs is
assigned for each taxonomic rank. If \code{map=NULL} and \code{biom} is a
\code{BIOM} class object, the map will be automatically loaded from
\code{biom$taxonomy}. \code{map} must not be \code{NULL} when \code{biom}
is a \code{matrix} or \code{simple_triplet_matrix}. See the example below
for an example of \code{map}'s structure. Default: \code{NULL}}

\item{lineage}{Include all ranks in the name of the taxa. For instance,
setting to \code{TRUE} will produce
\code{Bacteria; Actinobacteria; Coriobacteriia; Coriobacteriales}.
Whereas setting to \code{FALSE} (the default) will return simply
\code{Coriobacteriales}. You want to set this to TRUE if you have
genus names (such as \emph{Incertae_Sedis}) that map to multiple higher
level ranks. Default: \code{FALSE}}

\item{sparse}{If true, returns a sparse matrix as described by
\code{slam::simple_triplet_matrix}, otherwise returns a normal R
matrix object. Sparse matrices will likely be considerably more
memory efficient in this scenario. Default: \code{FALSE}}

\item{unc}{How to handle unclassified, uncultured, and similarly ambiguous
taxa names. Options are:
\itemize{
\item{\code{"singly"} - }{ Replaces them with the OTU name. }
\item{\code{"grouped"} - }{ Replaces them with a higher rank's name. }
\item{\code{"drop"} - }{ Excludes them from the result. }
\item{\code{"asis"} - }{ To not check/modify any taxa names. }
}
Default: \code{"singly"} \cr\cr
Non-ambiguous abbreviations are allowed}

\item{other}{Sum all non-itemized taxa into an "Other" taxa. When
\code{FALSE}, only returns taxa matched by the \code{`taxa`}
argument. Specifying \code{TRUE} adds "Other" to the returned set.
A string can also be given to imply \code{TRUE}, but with that
value as the name to use instead of "Other".
Default: \code{FALSE}}
}
\value{
A numeric matrix with samples as row names, and taxonomic
identifiers as column names.
}
\description{
Taxa abundances per sample, at the specified taxonomic rank.
}
\examples{
    library(rbiom)
    
    taxa_ranks(hmp50)
    
    phyla <- taxa_matrix(hmp50, 'Phylum')
    phyla[1:4,1:6]
    
    # Custom matrices should be formatted like so:
    counts <- otu_matrix(hmp50)
    map    <- otu_taxonomy(hmp50)
    
    counts[1:3,1:6]
    map[1:3,1:4]
    
    phyla <- taxa_matrix(counts, 'Phylum', map=map)
    phyla[1:3,1:6]

}
