% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{taxa_names<-}
\alias{taxa_names<-}
\title{Set the taxa names.}
\usage{
taxa_names(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read_biom}.}

\item{value}{A character vector. A named character vector can be used to
only change some of the taxa names.}
}
\description{
Set the taxa names.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    
    head(taxa_names(biom))
    
    taxa_names(biom) <- sub('Unc', 'Uncultured_', taxa_names(biom))
    head(taxa_names(biom))
    
    taxa_names(biom) <- c('AnmMass2' = 'One', 'PreBivi6' = 'Two')
    head(taxa_names(biom))

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sample_names<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa_ranks<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
