% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{taxa_ranks<-}
\alias{taxa_ranks<-}
\title{Set the taxa rank names.}
\usage{
taxa_ranks(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read_biom}.}

\item{value}{A character vector. A named character vector can be used to
only change some of the taxa rank names.}
}
\description{
Set the taxa rank names.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    
    taxa_ranks(biom)
    
    taxa_ranks(biom) <- c("OTU" = "ASV")
    taxa_ranks(biom)
    
    taxa_ranks(biom) <- paste0("Level", 1:7)
    taxa_ranks(biom)

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sample_names<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa_names<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
