% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{taxa_ranks}
\alias{taxa_ranks}
\alias{taxa_ranks<-}
\title{Get or set the names of the taxonomic ranks.}
\usage{
taxa_ranks(biom)

taxa_ranks(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{value}{A character vector the new taxa rank names.}
}
\value{
A character vector of the taxa rank names in \code{biom}.
}
\description{
Get or set the names of the taxonomic ranks.
}
\examples{
    library(rbiom) 
    
    biom <- hmp50
    
    taxa_ranks(biom)
    
    taxa_ranks(biom) <- c('.otu', paste0("Level", seq_len(n_ranks(biom) - 1)))
    taxa_ranks(biom)
}
\seealso{
Other taxonomy: 
\code{\link{n_ranks}()},
\code{\link{otu_taxonomy}()},
\code{\link{write_taxonomy}()}
}
\concept{taxonomy}
