% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_stacked.r
\name{taxa_stacked}
\alias{taxa_stacked}
\title{Display taxa abundances as a stacked bar graph.}
\usage{
taxa_stacked(
  biom,
  rank = NULL,
  taxa = 6,
  colors = TRUE,
  patterns = FALSE,
  label.by = NULL,
  order.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  dist = "euclidean",
  clust = "complete",
  other = TRUE,
  xlab.angle = 90,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{rank}{What rank(s) of taxa to display, for example \code{"Phylum"} 
or \code{"Genus"}. Run \code{taxa_ranks()} to see all options for a 
given BIOM object. The default, \code{NULL}, selects the lowest
level.}

\item{taxa}{Which taxa to give separate colors. An integer value will show
the top n most abundant taxa. A value 0 <= n < 1 will show all taxa 
with that mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. Default: \code{6}.}

\item{colors, patterns}{A character vector of colors or patterns to use in
the graph. A named character vector can be used to map taxon names to 
specific colors or patterns. Set to \code{TRUE} to auto-select colors
or patterns, or to \code{FALSE} to disable per-taxa colors
or patterns. Default: \code{colors=TRUE, patterns=FALSE}.}

\item{label.by, order.by}{What metadata column to use for labeling and/or
sorting the samples across the x-axis. When \code{order.by=NULL},
samples are arranged based on \code{dist} and \code{clust}, below.
Default: \code{NULL}.}

\item{facet.by, limit.by}{Metadata columns to 
use for data partitioning. Default: \code{NULL}}

\item{dist, clust}{Distance (\link{dist}) and clustering (\link{hclust})
methods to use for automatically arranging samples along the x-axis. 
Group samples with similar composition near one another.
Default: \code{dist="euclidean", clust="complete"}.}

\item{other}{Add an 'Other' taxa that ensures all bars sum to 100%. 
Default: \code{TRUE}.}

\item{xlab.angle}{How to rotate the tick labels on the x-axis.
Options are \code{90} (the default) or \code{30}.}

\item{...}{Parameters for underlying functions. Prefixing parameter names 
with a layer name ensures that a particular parameter is passed to, 
and only to, that layer.}
}
\value{
A \code{ggplot2} plot. The computed data points will be attached as 
        \code{attr(, 'data')}.
}
\description{
Display taxa abundances as a stacked bar graph.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    taxa_stacked(biom, rank="Phylum")
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{tree_plot}()}
}
\concept{plotting}
