% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.r
\name{taxonomy<-}
\alias{taxonomy<-}
\title{Set the taxonomy table.}
\usage{
taxonomy(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read_biom}.}

\item{value}{A character matrix with rownames \code{taxa_names(x)}. If
there are more rownames than taxa names, the matrix will be subset.}
}
\description{
Set the taxonomy table.
}
\examples{
    library(rbiom)
    
    taxonomy(hmp50)[1:4,]
    
    # Sometimes taxonomic names are incomplete
    taxonomy(hmp50)[c(107,170,194), 'Genus', drop=FALSE]
    
    # rbiom can insert more descriptive placeholders
    taxonomy(hmp50, fix.names = TRUE)[c(107,170,194), 'Genus', drop=FALSE]

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sample_names<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa_names<-}()},
\code{\link{taxa_ranks<-}()}
}
\concept{setters}
