% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.r
\name{taxonomy}
\alias{taxonomy}
\title{Get the taxonomy table.}
\usage{
taxonomy(biom, ranks = NULL, unc = "asis")
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}

\item{ranks}{The taxonomic ranks to return in the matrix, or \code{NULL} 
for all of them. Default: \code{NULL}}

\item{unc}{How to handle unclassified, uncultured, and similarly ambiguous
taxa names. Default: \code{"asis"}
\itemize{
  \item{\code{"asis"} - }{ Don't check/modify any taxa names. }
  \item{\code{"singly"} - }{ Replace with "Unc. <OTU ID>". }
  \item{\code{"grouped"} - }{ Replace with "Unc. <Higher Rank>". }
  \item{\code{"drop"} - }{ Don't include in the returned matrix. }
}}
}
\value{
A character matrix with taxa/OTU IDs as row names and taxa ranks as
        column names. An 'OTU' column is always added as the last column and
        matches the row names.
}
\description{
Get the taxonomy table.
}
\examples{
    library(rbiom)
    
    taxonomy(hmp50)[1:4,]
    taxonomy(hmp50, c("Family", "Genus"))[1:4,]
    head(taxonomy(hmp50, "Genus")[,])
    
    # Sometimes taxonomic names are incomplete
    taxonomy(hmp50)[c(53,107,139), 2:6]
    
    # rbiom can insert more descriptive placeholders
    taxonomy(hmp50, unc = "singly")[c(53,107,139), 3:6]
    taxonomy(hmp50, unc = "grouped")[c(53,107,139), 4:6]

}
\seealso{
Other accessors: 
\code{\link{comments}()},
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()}
}
\concept{accessors}
