% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.r
\name{tree_data}
\alias{tree_data}
\title{Provides a 'treedata' S4 object for use in ggtree functions.}
\usage{
tree_data(biom, reads = TRUE, clades = TRUE)
}
\arguments{
\item{biom}{An rbiom object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{reads}{Include a 'reads' column indicating the sum of taxa
observations belonging to each node/leaf. Default: \code{TRUE}.}

\item{clades}{Notate top-most monophyletic nodes. The default,
\code{TRUE}, adds a column for every rank in the biom object. A
character vector of ranks can also be passed in. If the vector is
named, then those names are used for naming the columns in the
returned treedata object. Set to \code{NULL} to not return any
clade notations.}
}
\description{
Provides a 'treedata' S4 object for use in ggtree functions.
}
\examples{
    library(rbiom)
    
    # Needs the optional ggtree package
    if (nzchar(system.file(package = "ggtree"))) {
    
      tree_data(hmp50)
      
    }
}
