% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.r
\name{tree_plot}
\alias{tree_plot}
\title{Display a dendrogram of the phylogenetic tree.}
\usage{
tree_plot(
  biom,
  layout = "rectangular",
  tiplab = NULL,
  color.by = NULL,
  label = NULL,
  cladelab = NULL,
  top = NULL,
  right = NULL,
  bottom = NULL,
  left = NULL,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{layout}{Any layout option supported by \code{\link[ggtree:ggtree]{ggtree::ggtree()}}:
\code{"rectangular"}, \code{"dendrogram"}, \code{"slanted"},
\code{"ellipse"}, \code{"roundrect"}, \code{"fan"},
\code{"circular"}, \code{"inward_circular"}, \code{"radial"},
\code{"equal_angle"}, \code{"daylight"} or \code{"ape"}.
Default: \code{"rectangular"}.}

\item{tiplab}{Label the tree leafs with taxa names. Options are
\code{NULL} (no tip labels) or a taxonomic rank (for example
\code{tiplab = "Genus"}). Default: \code{NULL}.}

\item{color.by}{How to color the tree. Currently only supports the options
\code{NULL} (no coloring) or \code{'.reads'} (color by number of taxa
observations). Default: \code{NULL}.}

\item{label, cladelab}{Label monophyletic clades. You can specify the
same or different taxonomic ranks for internal (label) and external
(cladelab) annotations. Default: \code{NULL}.}

\item{top, right, bottom, left}{Add additional space around the tree.
Sometimes necessary for wide text annotations. Set as fraction of
tree's width/height. For instance, \code{right = 1} reserves the
right half of the plotting area for non-tree elements.
Default: \code{NULL}.}
}
\description{
Display a dendrogram of the phylogenetic tree.
}
\examples{
    library(rbiom)
    
    hmp1 <- sample_select(hmp50, samples = "HMP10")
    
    # Needs the optional ggtree package
    if (nzchar(system.file(package = "ggtree")))
      tree_plot(hmp50)
      
    if (nzchar(system.file(package = "ggtree")))
      tree_plot(hmp1, cladelab = "Phylum", layout = "roundrect")
    
    if (nzchar(system.file(package = "ggtree")))
      tree_plot(hmp1, tiplab = "Genus", layout = "fan", color.by = ".reads")
}
