% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot.r
\name{tree_tips}
\alias{tree_tips}
\title{Names of a phylogenetic tree's tips/leafs.}
\usage{
tree_tips(x)
}
\arguments{
\item{x}{A phylo object, as returned from \code{\link[=read_tree]{read_tree()}}.}
}
\value{
A character vector with the leaf names.
}
\description{
Names of a phylogenetic tree's tips/leafs.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "newick.tre", package = "rbiom")
    tree  <- read_tree(infile)
    leafs <- tree_tips(tree)
    head(leafs)

}
