% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifrac.r
\name{unifrac}
\alias{unifrac}
\title{Compute Weighted and Unweighted UniFrac distance matrices.}
\usage{
unifrac(biom, weighted = TRUE, tree = NULL, progressbar = FALSE)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read.biom}. A 
\code{matrix} or \code{simple_triplet_matrix} is also acceptable, when
the names of the rows and columns are the taxa and samples, respectively.}

\item{weighted}{Use weighted UniFrac, which takes abundance into account
rather than simply presence/absence.}

\item{tree}{A \code{phylo} object providing a phylogenetic tree for the
taxa names in \code{biom}. If \code{tree=NULL}, then the tree will be
loaded from \code{biom}, if encoded there.}

\item{progressbar}{Whether to display a progress bar and status messages
(logical). Will automatically tie in with \pkg{shiny} if run within a
\pkg{shiny} session.}
}
\value{
A distance matrix of class \code{dist}.
}
\description{
This is the function called internally by \link{beta.div}, but is made
visible here so you can use it with matrices and trees without having to
first convert them to \code{BIOM} objects.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    biom <- select(biom, 1:10)
    
    dm <- unifrac(biom)
    plot(hclust(dm), cex=.8)
    as.matrix(dm)[1:4,1:4]
    
    # Using a custom matrix and tree
    mtx <- matrix(sample.int(12*20), ncol=20)
    dimnames(mtx) <- list(LETTERS[1:12], letters[1:20])
    tree <- ape::as.phylo(hclust(dist(mtx)))
    
    dm <- unifrac(mtx, tree=tree)
    as.matrix(dm)[1:4,1:4]
    
}
