% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{validate_metrics}
\alias{validate_metrics}
\title{Cleanup dynamic options that we can recognize}
\usage{
validate_metrics(biom, metrics, mode = NULL, multi = FALSE, mixed = FALSE, ...)
}
\description{
Cleanup dynamic options that we can recognize
}
