% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.r
\name{with}
\alias{with}
\alias{with.rbiom}
\alias{within.rbiom}
\title{Evaluate expressions on metadata.}
\usage{
\method{with}{rbiom}(biom, ...)

\method{within}{rbiom}(biom, ..., clone = TRUE)
}
\arguments{
\item{biom}{An \link[=rbiom_objects]{rbiom object}, such as from \code{\link[=as_rbiom]{as_rbiom()}}.}

\item{...}{Passed on to \code{\link[base:with]{base::with()}} or \code{\link[base:with]{base::within()}}.}
}
\value{
See description.
}
\description{
\code{with()} will return the result of your expression. \code{within()} will return
an rbiom object.
}
\examples{
    library(rbiom) 
    
    with(hmp50, table(`Body Site`, Sex))
    
    biom <- within(hmp50, {
      age_bin = cut(Age, 5)
      bmi_bin = cut(BMI, 5)
    })
    biom$metadata

}
\seealso{
Other transformations: 
\code{\link{mutate}()},
\code{\link{rarefy}()},
\code{\link{rarefy_cols}()},
\code{\link{slice}()},
\code{\link{subset}()}
}
\concept{transformations}
