% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{with_cache}
\alias{with_cache}
\title{Speeds up repetitive computations by storing results of functions calls.}
\usage{
with_cache(expr)
}
\arguments{
\item{expr}{An expression to cache the output of.}
}
\value{
The result of \code{expr}.
}
\description{
Speeds up repetitive computations by storing results of functions calls.
}
\examples{
    library(rbiom)
    
    system.time(x <- adiv_table(hmp50, "multi"))
    system.time(y <- adiv_table(hmp50, "multi"))
    identical(x, y)

}
