% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.r
\name{within.rbiom}
\alias{within.rbiom}
\title{Modify sample metadata.}
\usage{
\method{within}{rbiom}(biom, expr)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{expr}{A single or compound expression to evaluate.}
}
\value{
An \code{rbiom} object with the updated metadata.
}
\description{
Modify sample metadata.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    sample_metadata(biom)
    
    biom \%<>\% within(Sex2 <- as.numeric(Sex))
    sample_metadata(biom)
    
    biom \%<>\% within({
      `Body Site` <- substr(`Body Site`, 1, 4)
      remove("BMI", "Age", "Sex2")
    })
    sample_metadata(biom)

}
\seealso{
Other metadata: 
\code{\link{bdply}()},
\code{\link{blply}()},
\code{\link{metadata_names}()},
\code{\link{metadata_numeric}()},
\code{\link{n_metadata}()},
\code{\link{sample_metadata}()},
\code{\link{sample_select}()},
\code{\link{subset.rbiom}()},
\code{\link{write_metadata}()}

Other samples: 
\code{\link{n_samples}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{sample_select}()},
\code{\link{sample_sums}()},
\code{\link{subset.rbiom}()}
}
\concept{metadata}
\concept{samples}
