% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_biom.r
\name{write_biom}
\alias{write_biom}
\title{Write counts, metadata, taxonomy, and phylogeny to a biom file.}
\usage{
write_biom(biom, file, format = "json")
}
\arguments{
\item{biom}{The \code{rbiom} object to save to the file.}

\item{file}{Path to the output file. If the file name ends in \code{".gz"}
or \code{".bz2"}, the file contents will be compressed accordingly.}

\item{format}{Options are \code{"tab"}, \code{"json"}, and \code{"hdf5"},
corresponding to classic tabular format, BIOM format version 1.0 and
biom version 2.1, respectively.
See \url{http://biom-format.org/documentation/} for details.
NOTE: to write HDF5 formatted BIOM files, the BioConductor R package
\code{rhdf5} must be installed.}
}
\value{
The normalized filepath that was written to (invisibly).
}
\description{
Write counts, metadata, taxonomy, and phylogeny to a biom file.
}
\seealso{
Other biom: 
\code{\link{as_percent}()},
\code{\link{as_rbiom}()},
\code{\link{bdply}()},
\code{\link{biom_build}()},
\code{\link{biom_id}()},
\code{\link{biom_info}()},
\code{\link{biom_merge}()},
\code{\link{biom_repair}()},
\code{\link{blply}()},
\code{\link{read_biom}()},
\code{\link{write_xlsx}()}

Other writers: 
\code{\link{write_counts}()},
\code{\link{write_fasta}()},
\code{\link{write_metadata}()},
\code{\link{write_taxonomy}()},
\code{\link{write_tree}()},
\code{\link{write_xlsx}()}
}
\concept{biom}
\concept{writers}
