% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.r
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences from a BIOM object to a file in fasta format.}
\usage{
write_fasta(seqs, outfile)
}
\arguments{
\item{seqs}{A named character vector where names are sequence names and
values are the sequences. Also accepts a \code{BIOM}
object which contains sequences.}

\item{outfile}{Path to the output fasta file. Files ending in \code{.gz} or
\code{.bz2} will be compressed accordingly.}
}
\value{
On success, returns \code{NULL} invisibly.
}
\description{
Write sequences from a BIOM object to a file in fasta format.
}
