% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.r
\name{write_taxonomy}
\alias{write_taxonomy}
\title{Write BIOM taxonomy map to a tab-separated value (tsv) file.}
\usage{
write_taxonomy(biom, file)
}
\arguments{
\item{biom}{An rbiom object}

\item{file}{Path to the output file. Filenames ending in
\code{.gz} or \code{.bz2} will be compressed accordingly.}
}
\value{
The normalized filepath that was written to (invisibly).
}
\description{
Write BIOM taxonomy map to a tab-separated value (tsv) file.
}
