% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.r
\name{write_tree}
\alias{write_tree}
\title{Write a newick formatted phylogenetic tree.}
\usage{
write_tree(tree, file = NULL)
}
\arguments{
\item{tree}{A \code{phylo} object, as returned from \code{\link[=read_tree]{read_tree()}}. Also
accepts an \code{rbiom} object if it has a phylogentic tree.}

\item{file}{Filename or connection to write the newick file to (optional).}
}
\value{
If file is NULL, the newick string as a character vector. Otherwise,
the normalized filepath that was written to (invisibly).
}
\description{
Write a newick formatted phylogenetic tree.
}
\examples{
    library(rbiom) 
    
    infile <- system.file("extdata", "newick.tre", package = "rbiom")
    tree   <- read_tree(infile)
    newick <- write_tree(tree)

}
\seealso{
Other phylogeny: 
\code{\link{has_tree}()},
\code{\link{otu_tree}()},
\code{\link{read_tree}()},
\code{\link{tree_data}()},
\code{\link{tree_subset}()},
\code{\link{tree_tips}()}

Other writers: 
\code{\link{write_biom}()},
\code{\link{write_counts}()},
\code{\link{write_fasta}()},
\code{\link{write_metadata}()},
\code{\link{write_taxonomy}()},
\code{\link{write_xlsx}()}
}
\concept{phylogeny}
\concept{writers}
