% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.r
\name{write_wrapper}
\alias{write_wrapper}
\title{Common code for writers.}
\usage{
write_wrapper(outfile, callback)
}
\arguments{
\item{outfile}{The file to open a connection to.}

\item{callback}{A function that takes a connection as its only
argument and writes content to it.}
}
\value{
The normalized filepath that was written to (invisibly).
}
\description{
Common code for writers.
}
