% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adiv_boxplot.r
\name{adiv_boxplot}
\alias{adiv_boxplot}
\title{Visualize alpha diversity with boxplots.}
\usage{
adiv_boxplot(
  biom,
  x = NULL,
  adiv = "Shannon",
  layers = "bld",
  color.by = NULL,
  pattern.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  flip = FALSE,
  stripe = NULL,
  p.adj = "fdr",
  p.label = 0.05,
  ci = "ci",
  level = 0.95,
  outliers = NULL,
  xlab.angle = "auto",
  caption = TRUE,
  ...
)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{x}{A categorical metadata column name to use for the x-axis. The
default, \code{NULL}, groups all samples into a single category.}

\item{adiv}{Alpha diversity metric(s) to use. Options are: \code{"OTUs"},
\code{"Shannon"}, \code{"Chao1"}, \code{"Simpson"}, and/or
\code{"InvSimpson"}. Set \code{adiv=".all"} to use all metrics.
Default: \code{"Shannon"} \cr\cr
Multiple values allowed. Non-ambiguous abbreviations are allowed.}

\item{layers}{\code{"bar"}, \code{"box" ("x")}, \code{"violin"},
\code{"dot"}, \code{"strip"}, \code{"crossbar"}, \code{"errorbar"},
\code{"linerange"}, and \code{"pointrange"}. Single letter
abbreviations are also accepted. For instance, \code{c("box", "dot")}
is equivalent to \code{c("x", "d")} and \code{"xd"}.
See \code{vignette("boxplots")} for examples of each.
Default: \code{"bld"}.}

\item{color.by, pattern.by, shape.by, facet.by, limit.by}{Metadata columns to
use for aesthetics and partitioning. Default: \code{NULL}}

\item{flip}{Transpose the axes, so that taxa are present as rows instead
of columns. Default: \code{FALSE}}

\item{stripe}{Shade every other x position. Default: \emph{same as flip}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}.}

\item{p.label}{Minimum adjusted p-value to display on the plot with a
bracket.
\itemize{
\item{\code{p.label = 0.05} - }{ Show p-values that are <= 0.05. }
\item{\code{p.label = 0} - }{ Don't show any p-values on the plot. }
\item{\code{p.label = 1} - }{ Show all p-values on the plot. }
}
If a numeric vector with more than one value is
provided, they will be used as breaks for asterisk notation.
Default: \code{0.05}}

\item{ci}{How to calculate min/max of the \bold{crossbar},
\bold{errorbar}, \bold{linerange}, and \bold{pointrange} layers.
Options are: \code{"ci"} (confidence interval), \code{"range"},
\code{"sd"} (standard deviation), \code{"se"} (standard error), and
\code{"mad"} (median absolute deviation).
The center mark of \bold{crossbar} and \bold{pointrange} represents
the mean, except for code{"mad"} in which case it represents
the median. Default: \code{"ci"}}

\item{level}{The confidence level for calculating a confidence interval.
Default: \code{0.95}}

\item{outliers}{Show boxplot outliers? \code{TRUE} to always show.
\code{FALSE} to always hide. \code{NULL} to only hide them when
overlaying a dot or strip chart.  Default: \code{NULL}}

\item{xlab.angle}{Angle of the labels at the bottom of the plot.
Options are \code{"auto"}, \code{'0'}, \code{'30'}, and \code{'90'}.
Default: \code{"auto"}.}

\item{caption}{Add methodology caption beneath the plot.
Default: \code{TRUE}.}

\item{...}{Parameters are matched to formal arguments of ggplot2
functions. Prefixing parameter names with a layer name ensures that
a particular parameter is passed to, and only to, that layer. For
instance, \code{dot.size = 2} or \code{d.size = 2} ensures only the
dotplot layer has its size set to \code{2}. The special prefix
\code{pt.} will control both the dot and strip layers.}
}
\value{
A \code{ggplot2} plot. \cr The computed data points, statistics,
ggplot command, and object history are available as \code{$data},
\code{$stats}, \code{$code}, and \code{$history}, respectively.
}
\description{
Visualize alpha diversity with boxplots.
}
\examples{
    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    
    adiv_boxplot(biom, x="Sex", adiv=c("otu", "shan"), color.by="Body Site")
    
    adiv_boxplot(biom, x="body", adiv=".all", layers="p", color.by="sex", flip=TRUE)
    
    adiv_boxplot(biom, x="Body Site", color.by="Body Site")
    
    
    # Each plot object includes additional information.
    fig <- adiv_boxplot(biom, x="Body Site")
    
    ## Provenance History ---------------------
    fig$history
    
    ## Computed Data Points -------------------
    fig$data
    
    ## Statistics Table -----------------------
    fig$stats
    
    ## ggplot2 Command ------------------------
    fig$code
    
}
\seealso{
Other alpha_diversity: 
\code{\link{adiv_corrplot}()},
\code{\link{adiv_stats}()},
\code{\link{adiv_table}()}

Other visualization: 
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_barplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()}
}
\concept{alpha_diversity}
\concept{visualization}
