% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adiv_corrplot.r
\name{adiv_corrplot}
\alias{adiv_corrplot}
\title{Visualize alpha diversity with scatterplots and trendlines.}
\usage{
adiv_corrplot(
  biom,
  x,
  metric = "OTUs",
  layers = "t",
  color.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  model = "linear",
  p.adj = "fdr",
  ci = 95,
  caption = FALSE,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{x}{A numeric metadata column name to use for the x-axis. Required.}

\item{metric}{Alpha diversity metric(s) to use. Options are: 
\code{"OTUs"}, \code{"Shannon"}, \code{"Chao1"}, \code{"Simpson"}, 
and/or \code{"InvSimpson"}. Default: \code{"OTUs"}.}

\item{layers}{\code{"trend"}, \code{"scatter"}. Single letter 
abbreviations are also accepted. For instance, 
\code{c("trend", "scatter")} is equivalent to \code{c("t", "s")} and 
\code{"ts"}. See \code{vignette("corrplots")} for examples of each. 
Default: \code{"t"}.}

\item{color.by, facet.by, limit.by}{Metadata columns to use for aesthetics 
and partitioning. See below for details. Default: \code{NULL}}

\item{model}{What type of trendline to fit to the data. Options are: 
\code{"linear"}, \code{"logarithmic"}, or \code{"local"}. You can
alternatively provide a list of length two where the first element is
a character vector of length 1 naming a function, and the second 
element is a list of arguments to pass to that function. One of the 
function's arguments must be named 'formula'. For example, 
\code{model = list("stats::lm", list(formula = y ~ x))}.
Default: \code{"linear"}.}

\item{p.adj}{Method to use for multiple comparisons adjustment of p-values.
Run \code{p.adjust.methods} for a list of available options.
Default: \code{"fdr"}.}

\item{ci}{The confidence interval to display around the fitted curve. Set
to \code{FALSE} to hide the confidence interval. Default: \code{95}.}

\item{caption}{Display information about the method used for trendline
fitting beneath the plot. Default: \code{FALSE}.}

\item{...}{Additional parameters to pass along to ggplot2
functions. Prefix a parameter name with either \code{t.} or 
\code{s.}/\code{pt.} to ensure it gets passed to (and only to) 
\link[ggplot2]{geom_smooth} or \link[ggplot2]{geom_point}, 
respectively. For instance, \code{s.size = 2} ensures only the 
scatterplot points have their size set to \code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
}
\description{
Visualize alpha diversity with scatterplots and trendlines.
}
\section{Aesthetics and Partitions}{


Metadata can be used to flexibly subset, partition, and apply aesthetics 
when creating a plot. Common use cases are provided below. More thorough 
documentation is available at \url{https://cmmr.github.io/rbiom}.

\preformatted{  ## Colors ----------------------------
  color.by = "Body Site"
  color.by = list('Body Site' = "bright")
  color.by = list('Body Site' = c("Stool", "Saliva"))
  color.by = list('Body Site' = list('values' = c("Stool", "Saliva"), 'colors' = "bright"))
  color.by = list('Body Site' = c('Stool' = "blue", 'Saliva' = "green"))
  
  ## Facets ----------------------------
  facet.by = "Body Site"
  facet.by = c("Body Site", "Sex")
  facet.by = list('Body Site' = c("Stool", "Saliva"), "Sex")
  
  ## Limits ----------------------------
  limit.by = list('Sex' = "Male", 'Age' = c(20,40))
  limit.by = list('Body Site' = c("Saliva", "Anterior nares"), 'Age' = c(NA,35))
}

\itemize{
  \item{\code{color.by} - }{Any metadata column. (Max 1)}
  \item{\code{facet.by} - }{Only categorical metadata column(s).}
  \item{\code{limit.by} - }{Any metadata column(s).}
}

All built-in color palettes are colorblind-friendly.

The available categorical palette names are: \code{"okabe"}, \code{"carto"}, 
\code{"r4"}, \code{"polychrome"}, \code{"tol"}, \code{"bright"}, 
\code{"light"}, \code{"muted"}, \code{"vibrant"}, \code{"tableau"}, 
\code{"classic"}, \code{"alphabet"}, \code{"tableau20"}, \code{"kelly"}, 
and \code{"fishy"}.

The available numeric palette names are: \code{"reds"}, \code{"oranges"}, 
\code{"greens"}, \code{"purples"}, \code{"grays"}, \code{"acton"}, 
\code{"bamako"}, \code{"batlow"}, \code{"bilbao"}, \code{"buda"}, 
\code{"davos"}, \code{"devon"}, \code{"grayC"}, \code{"hawaii"}, 
\code{"imola"}, \code{"lajolla"}, \code{"lapaz"}, \code{"nuuk"}, 
\code{"oslo"}, \code{"tokyo"}, \code{"turku"}, \code{"bam"}, 
\code{"berlin"}, \code{"broc"}, \code{"cork"}, \code{"lisbon"}, 
\code{"roma"}, \code{"tofino"}, \code{"vanimo"}, \code{"vik"}
}

\examples{
    library(rbiom)
    
    adiv_corrplot(rarefy(hmp50), "Age", color.by="Body Site", metric=c("shannon", "otus"), facet.by = "Sex", ci = 90) 
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()},
\code{\link{tree_plot}()}
}
\concept{plotting}
