% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adiv_corrplot.r
\name{adiv_corrplot}
\alias{adiv_corrplot}
\title{Visualize alpha diversity with scatterplots and trendlines.}
\usage{
adiv_corrplot(
  biom,
  x,
  layers = "ps",
  metric = "OTUs",
  color.by = NULL,
  facet.by = NULL,
  colors = NULL,
  facets = NULL,
  method = "lm",
  formula = y ~ x,
  ci = 95,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{x}{A numeric metadata column name to use for the x-axis. Required.}

\item{layers}{\code{"point"}, and/or \code{"smooth"}. Single letter 
abbreviations are also accepted. For instance, 
\code{c("point", "smooth")} is equivalent to \code{c("p", "s")} and 
\code{"ps"}. Default: \code{"ps"}.}

\item{metric}{Alpha diversity metric(s) to use. Options are: 
\code{"OTUs"}, \code{"Shannon"}, \code{"Chao1"}, \code{"Simpson"}, 
and/or \code{"InvSimpson"}. Default: \code{"OTUs"}.}

\item{color.by, facet.by}{Metadata column to color and/or facet by. If that
column is a \code{factor}, the ordering of levels will be maintained 
in the plot. Default: \code{color.by = NULL, facet.by = NULL}}

\item{colors, facets}{Names of the colors and/or facets values to use in
the plot. Available values for colors are given by \code{colors()}. 
Use a named character vector to map them to specific factor levels in
the metadata. \code{facets} are coerced to unnamed character vectors.
If the length of these vectors is less than the values present in 
their corresponding metadata column, then the data set will be 
subseted accordingly. Default: \code{colors = NULL, facets = NULL}}

\item{method, formula}{Passed to \link[ggplot2]{stat_smooth} to
define the fitted curve.
Default: \code{method = "lm", formula = y ~ x}.}

\item{ci}{The confidence interval to display around the fitted curve. Set
to \code{FALSE} to hide the confidence interval. Default: \code{95}.}

\item{...}{Additional parameters to pass along to ggplot2
functions. Prefix a parameter name with either \code{p.} or \code{s.}
to ensure it gets passed to (and only to) 
\link[ggplot2]{geom_point} or 
\link[ggplot2]{geom_smooth}, respectively. For instance, 
\code{p.size = 2} ensures only the points have their size set to 
\code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
}
\description{
Visualize alpha diversity with scatterplots and trendlines.
}
\examples{
    library(rbiom)
    
    adiv_corrplot(rarefy(hmp50), "Age", color.by="Body Site", metric=c("shannon", "otus"), facet.by = "Sex", ci = 90)
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()}
}
\concept{plotting}
