% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adiv_table.r
\name{adiv_table}
\alias{adiv_table}
\title{Calculate the alpha diversity of each sample.}
\usage{
adiv_table(biom, rarefy = FALSE, adiv = "Shannon", md = TRUE)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{rarefy}{Control how/whether rarefactions are done prior to alpha
diversity computations. Options are:
\itemize{
\item{\code{FALSE} - }{
Use each sample's current set of observations without applying
any rarefaction.}
\item{\code{TRUE} - }{
Automatically select and apply a single rarefaction. }
\item{\code{"multi"}, \code{"multi_log"}, \code{"multi_even"} - }{
Automatically select and apply multiple rarefactions.
\code{"multi"} provides \code{"multi_log"} at the low end
and \code{"multi_even"} at the high end. }
\item{\emph{integer vector} - }{
Rarefy at the specified depth(s). }
}
Default: \code{FALSE}}

\item{adiv}{Alpha diversity metric(s) to use. Options are: \code{"OTUs"},
\code{"Shannon"}, \code{"Chao1"}, \code{"Simpson"}, and/or
\code{"InvSimpson"}. The shortcut \code{adiv="all"} also exists.
Default: \code{"Shannon"} \cr\cr
Multiple values allowed. Non-ambiguous abbreviations are allowed.}

\item{md}{Include metadata in the output data frame? Options are:
\itemize{
\item{\code{NULL} - }{ Don't include metadata. }
\item{\code{TRUE} - }{ Include all metadata. }
\item{\emph{character vector} - }{
Include only the specified metadata columns. }
}
Default: \code{TRUE}}
}
\value{
A data frame of alpha diversity values. \cr
Each combination of sample/depth/\code{adiv} has its own row. \cr
Column names are \bold{.sample}, \bold{.depth}, \bold{.adiv},
and \bold{.diversity}, followed by any metadata fields requested by
\code{md}.
}
\description{
Calculate the alpha diversity of each sample.
}
\examples{
    library(rbiom) 
    
    biom <- sample_select(hmp50, 1:6)
    
    adiv_table(biom)
    
    adiv_table(sample_rarefy(biom), adiv = "all", md = FALSE)
}
\seealso{
Other alpha_diversity: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{adiv_stats}()}
}
\concept{alpha_diversity}
