% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.div.r
\name{alpha.div}
\alias{alpha.div}
\title{Estimate the diversity of each sample.}
\usage{
alpha.div(
  biom,
  rarefy = FALSE,
  metrics = "all",
  long = FALSE,
  md = FALSE,
  safe = FALSE
)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM}
object, as returned from \link{read.biom}. For matrices, the rows and
columns are assumed to be the taxa and samples, respectively.}

\item{rarefy}{Control how/whether rarefactions are done prior to alpha 
diversity computations. Options are: 
\describe{
  \item{\code{FALSE}}{ Use each sample's current set of observations without
                  applying any rarefaction. (Default) }
  \item{\code{TRUE}}{ Automatically select and apply a single rarefaction. }
  \item{\code{"multi"}}{ Automatically select and apply multiple rarefactions.}
  \item{\emph{integer vector}}{ Rarefy at the specified depth(s). }
}}

\item{metrics}{Character vector of one or more of the following: \code{Depth}, 
\code{OTUs}, \code{Shannon}, \code{Chao1}, \code{Simpson}, \code{InvSimpson}.
Non-ambiguous abbreviations are also accepted. The default, \code{all},
returns all of them.}

\item{long}{Pivot the returned data to long format?
\describe{
  \item{\bold{FALSE}}{ Each metric has its own column. (Default) }
  \item{\bold{TRUE}}{ "Sample", "Metric" and "Diversity" are the columns 
                  returned. Rows are added to attain all combinations of 
                  samples x metrics. }
}}

\item{md}{Include metadata in the output data frame? Options are: 
\describe{
  \item{\code{FALSE}}{ Don't include metadata. (Default) }
  \item{\code{TRUE}}{ Include all metadata. }
  \item{\emph{character vector}}{ Include only the specified metadata columns. }
}}

\item{safe}{Should autogenerated columns be prefixed with a "." to avoid conflicting
with metadata column names? (Default: FALSE)}
}
\value{
A data frame of four diversity values for each sample in
    \code{biom}. The column names are \bold{Sample}, \bold{Depth} and the 
    diversity metrics: \bold{OTUs}, \bold{Shannon}, \bold{Chao1}, \bold{Simpson}, 
    and \bold{InvSimpson}. The row names are the sample names, except when
    multiple rarefactions are done.
}
\description{
Estimate the diversity of each sample.
}
\examples{
    library(rbiom)
    
    ad <- alpha.div(hmp50)
    head(ad)
    
    biom <- subset(hmp50, `Body Site` == "Saliva" & Age < 26)
    ad   <- alpha.div(biom, "multi")
    boxplot(Shannon ~ Depth, data=ad, xlab="Reads", ylab="Diversity")
    
}
