% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.r
\name{as.list.rbiom}
\alias{as.list.rbiom}
\title{Convert an rbiom object to a base R list.}
\usage{
\method{as.list}{rbiom}(x, ...)
}
\arguments{
\item{x}{An \link[=rbiom_objects]{rbiom object}, such as from \code{\link[=as_rbiom]{as_rbiom()}}.}

\item{...}{Not used.}
}
\value{
A list with names
\code{c('counts', 'metadata', 'taxonomy', 'tree', 'sequences', 'id', 'comment', 'date', 'generated_by')}.
}
\description{
Convert an rbiom object to a base R list.
}
\seealso{
Other conversion: 
\code{\link{as.matrix.rbiom}()}
}
\concept{conversion}
