% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_percent.r
\name{as_percent}
\alias{as_percent}
\title{Convert absolute counts to relative abundances.}
\usage{
as_percent(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
An \code{rbiom} object.
}
\description{
Convert absolute counts to relative abundances.
}
\examples{
    library(rbiom)
    
    otu_matrix(hmp50)[1:4,1:5]
    
    biom <- as_percent(hmp50)
    otu_matrix(biom)[1:4,1:5]

}
