% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_rbiom.r
\name{as_rbiom}
\alias{as_rbiom}
\title{Convert BIOM data to an rbiom object.}
\usage{
as_rbiom(biom)
}
\arguments{
\item{biom}{Object which can be coerced to an \code{rbiom}-class object.
For example:
\itemize{
\item{\emph{file} - }{ Filepath or URL to a biom file. }
\item{\emph{matrix} - }{ An abundance matrix with OTUs in rows and samples in columns. }
\item{\code{phyloseq}-class object - }{ From the phyloseq Bioconductor R package. }
}}
}
\value{
An \code{rbiom}-class object.
}
\description{
Convert BIOM data to an rbiom object.
}
\seealso{
\code{\link[=biom_build]{biom_build()}}, \code{\link[=read_biom]{read_biom()}}
}
