% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_boxplot.r
\name{bdiv_boxplot}
\alias{bdiv_boxplot}
\title{Visualize BIOM data with boxplots.}
\usage{
bdiv_boxplot(
  biom,
  x = NULL,
  metric = "Bray-Curtis",
  layers = "rls",
  color.by = NULL,
  pattern.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  xvals = NULL,
  colors = NULL,
  patterns = NULL,
  shapes = NULL,
  facets = NULL,
  p.adj = "fdr",
  p.label = TRUE,
  ci = 95,
  xlab.angle = "auto",
  weighted = TRUE,
  tree = NULL,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{x}{Metadata column name(s). Must be categorical. Prefix with 
\code{==} or \code{!=} to limit comparisons to within or between
groups, respectively. The default, \code{NULL} groups all distances 
into a single column.}

\item{metric}{Beta diversity metric(s) to use. Options are 
\code{"Manhattan"}, \code{"Euclidean"}, \code{"Bray-Curtis"}, 
\code{"Jaccard"}, and \code{"UniFrac"}. UniFrac requires a 
phylogenetic tree. Default: \code{"Bray-Curtis"}.}

\item{layers}{\code{"box"}, \code{"bar" ("r")}, \code{"violin"}, 
\code{"dot"}, \code{"strip"}, \code{"crossbar"}, \code{"errorbar"}, 
\code{"linerange"}, and \code{"pointrange"}. Single letter 
abbreviations are also accepted. For instance, \code{c("box", "dot")} 
is equivalent to \code{c("b", "d")} and \code{"bd"}.
Default: \code{"rls"}.}

\item{color.by, pattern.by, shape.by, label.by, sort.by, facet.by}{Metadata 
column to color, pattern, shape, label, sort, and/or facet by. If 
that column is a \code{factor}, the ordering of levels will be 
maintained in the plot.}

\item{colors, patterns, shapes, facets, xvals}{Names of the colors, patterns,
shapes, facets, and/or x values to use in the plot. Available values 
for colors, patterns, and shapes are given by \code{colors()}, 
\code{gridpattern::names_magick}, and \code{0:25},
respectively. Use a named character vector to map them to specific
factor levels in the metadata. \code{facets} and \code{xvals} are
coerced to unnamed character vectors. If the length of these vectors
is less than the values present in their corresponding metadata 
column, then the data set will be subseted accordingly.}

\item{p.adj}{Method to use for multiple comparisons adjustment of p-values.
Run \code{p.adjust.methods} for a list of available options.
Default: \code{fdr}}

\item{p.label}{Minimum adjusted p-value to display on the plot with a 
bracket, for example \code{p.label = 0.05}.
Some special cases are:
\itemize{
  \item{\code{p.label = TRUE} - }{ equivalent to \code{p.label = 0.05} }
  \item{\code{p.label = FALSE} - }{ do not show any stats on the plot }
  \item{\code{p.label = Inf} - }{ display all p-values }
  \item{\code{p.label = NULL} - }{ do not calculate stats }
}
If a numeric vector with more than one value is 
provided, they will be used as breaks for asterisk notation.
Default: \code{TRUE}}

\item{ci}{How to calculate min/max of the \bold{crossbar}, 
\bold{errorbar}, \bold{linerange}, and \bold{pointrange} layers.
Provide a number between 75 and 100 to define a confidence interval's
confidence level, commonly 95 or 97.5. Other options are: 
\bold{range}, 
\bold{sd} (standard deviation), 
\bold{se} (standard error), and 
\bold{mad} (median absolute deviation). 
The center mark of \code{crossbar} and \code{pointrange} represents
the mean, except for \bold{mad} in which case it represents
the median. Trendlines require a confidence interval value. 
Set to \code{NULL} to disable. Default: \code{95}}

\item{xlab.angle}{How to rotate the tick labels on the x-axis. 
\bold{'auto'} (the default), automatically selects a rotation value. 
\bold{0}, \bold{30}, and \bold{90} sets the angle to horizontal, 
angled, and vertical, respectively.}

\item{weighted}{When employing a beta diversity metric, use the weighted
version. Default: \code{TRUE}.}

\item{tree}{A phylogenetic tree for use in calculating UniFrac distance.
The default, \code{NULL}, will use the BIOM object's tree.}

\item{...}{Parameters are matched to formal arguments of ggplot2
functions. Prefixing parameter names with a layer name ensures that
a particular parameter is passed to, and only to, that layer. For
instance, \code{dot.size = 2} or \code{d.size = 2} ensures only the 
dotplot layer has its size set to \code{2}.}

\item{filter.fun}{A function that takes a \code{bdiv_table()} as input and
returns a \code{bdiv_table()} as output. This modified table will be
used for plotting. Default: \code{NULL}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
        
Shapes can also be given as their string values, defined in pch_table here:
https://github.com/tidyverse/ggplot2/blob/master/R/geom-point.r . Note that
some shapes have a colored outline given by `color`, some are filled with 
`color`, and some are outlined in `color` and filled with `fill`. See
https://blog.albertkuo.me/post/point-shape-options-in-ggplot/ for details.

To expand the low end of the y axis, you can set \code{y.trans = "sqrt"} or
\code{y.trans = "log1p"}. The former applies a square-root transformation, 
and the latter plots log(y + 1). Both of these methods work well with data
that contains zeroes.
}
\description{
Visualize BIOM data with boxplots.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    bdiv_boxplot(biom, x="==Body Site", metric="UniFrac", color.by="Body Site")
    

}
\seealso{
\code{\link{stats_table}}

Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()},
\code{\link{tree_plot}()}
}
\concept{plotting}
