% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_boxplot.r
\name{bdiv_boxplot}
\alias{bdiv_boxplot}
\title{Visualize BIOM data with boxplots.}
\usage{
bdiv_boxplot(
  biom,
  x = NULL,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  layers = "bld",
  color.by = NULL,
  pattern.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  within = NULL,
  between = NULL,
  flip = FALSE,
  stripe = NULL,
  p.adj = "fdr",
  p.label = 0.05,
  ci = "ci",
  level = 0.95,
  outliers = NULL,
  xlab.angle = "auto",
  ...
)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{x}{A categorical metadata column name. Prefix the column name with
\code{==} or \code{!=} to limit comparisons to within or between
groups, respectively. The default, \code{NULL} groups all distances
into a single column.}

\item{bdiv}{Beta diversity distance algorithm(s) to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. For \code{"UniFrac"}, a
phylogenetic tree must be present in \code{biom} or explicitly
provided via \code{tree=}. Default: \code{"Bray-Curtis"} \cr\cr
Multiple/abbreviated values allowed.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Default: \code{TRUE} \cr\cr
Multiple values allowed.}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Only required when
computing UniFrac distances. Default: \code{otu_tree(biom)}}

\item{layers}{\code{"bar"}, \code{"box" ("x")}, \code{"violin"},
\code{"dot"}, \code{"strip"}, \code{"crossbar"}, \code{"errorbar"},
\code{"linerange"}, and \code{"pointrange"}. Single letter
abbreviations are also accepted. For instance, \code{c("box", "dot")}
is equivalent to \code{c("x", "d")} and \code{"xd"}.
See \code{vignette("boxplots")} for examples of each.
Default: \code{"bld"}.}

\item{color.by, pattern.by, shape.by, facet.by, limit.by}{Metadata columns to
use for data partitioning. Prefix the column name with
\code{==} or \code{!=} to limit comparisons to within or between
groups, respectively. Default: \code{NULL}}

\item{within, between}{Metadata field(s) for intra- or inter- sample
comparisons. Default: \code{NULL}}

\item{flip}{Transpose the axes, so that taxa are present as rows instead
of columns. Default: \code{FALSE}}

\item{stripe}{Shade every other x position. Default: \emph{same as flip}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}.}

\item{p.label}{Minimum adjusted p-value to display on the plot with a
bracket.
\itemize{
\item{\code{p.label = 0.05} - }{ Show p-values that are <= 0.05. }
\item{\code{p.label = 0} - }{ Don't show any p-values on the plot. }
\item{\code{p.label = 1} - }{ Show all p-values on the plot. }
}
If a numeric vector with more than one value is
provided, they will be used as breaks for asterisk notation.
Default: \code{0.05}}

\item{ci}{How to calculate min/max of the \bold{crossbar},
\bold{errorbar}, \bold{linerange}, and \bold{pointrange} layers.
Options are: \code{"ci"} (confidence interval), \code{"range"},
\code{"sd"} (standard deviation), \code{"se"} (standard error), and
\code{"mad"} (median absolute deviation).
The center mark of \bold{crossbar} and \bold{pointrange} represents
the mean, except for code{"mad"} in which case it represents
the median. Default: \code{"ci"}}

\item{level}{The confidence level for calculating a confidence interval.
Default: \code{0.95}}

\item{outliers}{Show boxplot outliers? \code{TRUE} to always show.
\code{FALSE} to always hide. \code{NULL} to only hide them when
overlaying a dot or strip chart.  Default: \code{NULL}}

\item{xlab.angle}{Angle of the labels at the bottom of the plot.
Options are \code{"auto"}, \code{'0'}, \code{'30'}, and \code{'90'}.
Default: \code{"auto"}.}

\item{...}{Parameters are matched to formal arguments of ggplot2
functions. Prefixing parameter names with a layer name ensures that
a particular parameter is passed to, and only to, that layer. For
instance, \code{dot.size = 2} or \code{d.size = 2} ensures only the
dotplot layer has its size set to \code{2}. The special prefix
\code{pt.} will control both the dot and strip layers.}
}
\value{
A \code{ggplot2} plot. \cr The computed data points, statistics,
ggplot command, and object history are available as \code{$data},
\code{$stats}, \code{$code}, and \code{$history}, respectively.
}
\description{
Visualize BIOM data with boxplots.
}
\examples{
    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    bdiv_boxplot(biom, x="==Body Site", bdiv="UniFrac", color.by="Body Site")
    

}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_distmat}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_stats}()},
\code{\link{bdiv_table}()},
\code{\link{distmat_stats}()}

Other visualization: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_barplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()}
}
\concept{beta_diversity}
\concept{visualization}
