% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_distmat.r
\name{bdiv_distmat}
\alias{bdiv_distmat}
\title{Make a distance matrix of samples vs samples.}
\usage{
bdiv_distmat(biom, bdiv = "Bray-Curtis", weighted = TRUE, tree = NULL)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM}
object (e.g. as returned from \code{\link[=read_biom]{read_biom()}}). For matrices, the
rows and columns are assumed to be the taxa and samples, respectively.}

\item{bdiv}{The beta diversity distance algorithm to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. Non-ambiguous abbreviations
of the algorithm names are also accepted. For \code{"UniFrac"}, a
phylogenetic tree must be present in \code{biom} or explicitly
provided via \code{tree=}. Default: \code{"Bray-Curtis"}.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Default: \code{TRUE}}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Will be taken from the tree
defined in \code{biom} if not explicitly specified. Only required
when computing UniFrac distance matrices. Default: \code{NULL}}
}
\value{
A distance matrix.
}
\description{
Make a distance matrix of samples vs samples.
}
\examples{
    library(rbiom)
    
    biom <- sample_select(hmp50, 1:10)
    dm   <- bdiv_distmat(biom, 'unifrac')
    as.matrix(dm)[1:4,1:4]
    plot(hclust(dm))

}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_table}()},
\code{\link{distmat_stats}()}
}
\concept{beta_diversity}
