% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_heatmap.r
\name{bdiv_heatmap}
\alias{bdiv_heatmap}
\title{Display beta diversities in an all vs all grid.}
\usage{
bdiv_heatmap(
  biom,
  metric = "Bray-Curtis",
  weighted = TRUE,
  color.by = NULL,
  order.by = NULL,
  colors = NULL,
  orders = NULL,
  xlab.angle = "auto",
  tree = NULL,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{metric}{Beta diversity metric(s) to use. Options are 
\code{"Manhattan"}, \code{"Euclidean"}, \code{"Bray-Curtis"}, 
\code{"Jaccard"}, and \code{"UniFrac"}. UniFrac requires a 
phylogenetic tree. Default: \code{"Bray-Curtis"}.}

\item{weighted}{Should the beta diversity metric be weighted by 
abundances? Default: \code{TRUE}.}

\item{color.by}{Add annotation tracks for these metadata column(s). 
Default: \code{NULL}.}

\item{order.by}{Which metadata columns to use for ordering the samples 
across the x-axis. The default setting, \code{order.by=NULL}, uses 
sample names as labels and arranges samples by similarity. See 
\link{plot_heatmap} for details.}

\item{colors}{A named \code{list()} with names matching \code{color.by}, 
each specifying a \bold{palette} as described in \link{plot_heatmap}.
For example: \code{colors = list(Sex = c(Male="blue", Female="red"))}.
Default: \code{NULL}.}

\item{orders}{Subset to just these values in the \code{order.by} metadata 
column. Default: \code{NULL}.}

\item{xlab.angle}{How to rotate the tick labels on the x-axis. 
\bold{'auto'} (the default), automatically selects a rotation value. 
\bold{0}, \bold{30}, and \bold{90} sets the angle to horizontal, 
angled, and vertical, respectively.}

\item{tree}{A phylogenetic tree for use in calculating UniFrac distance.
The default, \code{NULL}, will use the BIOM object's tree.}

\item{...}{Additional arguments to pass on to \link{plot_heatmap}.}
}
\value{
A \code{ggplot2} plot. The computed data points will be attached as 
        \code{attr(, 'data')}.
}
\description{
Display beta diversities in an all vs all grid.
}
\examples{
    library(rbiom)
    
    biom <- hmp50 \%>\% rarefy() \%>\% select(1:10)
    bdiv_heatmap(biom, color.by="Body Site")
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{plot_heatmap}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()}
}
\concept{plotting}
