% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_heatmap.r
\name{bdiv_heatmap}
\alias{bdiv_heatmap}
\title{Display beta diversities in an all vs all grid.}
\usage{
bdiv_heatmap(
  biom,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  grid = list(label = "Distance", colors = "-bilbao"),
  color.by = NULL,
  order.by = NULL,
  limit.by = NULL,
  label = TRUE,
  label_size = NULL,
  rescale = "none",
  ratio = 1,
  clust = "complete",
  trees = TRUE,
  tree_height = NULL,
  track_height = NULL,
  legend = "right",
  xlab.angle = "auto",
  ...
)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{bdiv}{Beta diversity distance algorithm(s) to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. For \code{"UniFrac"}, a
phylogenetic tree must be present in \code{biom} or explicitly
provided via \code{tree=}. Default: \code{"Bray-Curtis"} \cr\cr
Multiple/abbreviated values allowed.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Default: \code{TRUE} \cr\cr
Multiple values allowed.}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Only required when
computing UniFrac distances. Default: \code{otu_tree(biom)}}

\item{grid}{Color palette name, or a list with entries for \code{label},
\code{colors}, \code{range}, \code{bins}, \code{na.color}, and/or
\code{guide}. See the Track Definitions section for details.
Default: \code{list(label = "Distance", colors = "-bilbao")}.}

\item{color.by}{Add annotation tracks for these metadata column(s).
See "Annotation Tracks" section below for details.
Default: \code{NULL}}

\item{order.by}{Which metadata column(s) to use for ordering the samples
across the x and y axes. Overrides any \code{clust} argument.
See "Ordering and Limiting" section below for details.
Default: \code{NULL}}

\item{limit.by}{Metadata definition(s) to use for sample subsetting prior
to calculations.
See "Ordering and Limiting" section below for details.
Default: \code{NULL}}

\item{label}{Label the matrix rows and columns. You can supply a list
or logical vector of length two to control row labels and column
labels separately, for example
\code{label = c(rows = TRUE, cols = FALSE)}, or simply
\code{label = c(T, F)}. Other valid options are \code{"rows"},
\code{"cols"}, \code{"both"}, \code{"bottom"}, \code{"right"},
and \code{"none"}.
Default: \code{TRUE}.}

\item{label_size}{The font size to use for the row and column labels. You
can supply a numeric vector of length two to control row label sizes
and column label sizes separately, for example
\code{c(rows = 20, cols = 8)}, or simply \code{c(20, 8)}.
Default: \code{NULL}, which computes:
\code{pmax(8, pmin(20, 100 / dim(mtx)))}.}

\item{rescale}{Rescale rows or columns to all have a common min/max.
Options: \code{"none"}, \code{"rows"}, or \code{"cols"}.
Default: \code{"none"}.}

\item{ratio}{Height/width ratio for entire grid.
Default: \code{1} (square).}

\item{clust}{Clustering algorithm for reordering the rows and columns by
similarity. You can supply a list or character vector of length two to
control the row and column clustering separately, for example
\code{clust = c(rows = "complete", cols = NA)}, or simply
\code{clust = c("complete", NA)}. Options are:
\itemize{
\item{\code{FALSE} or \code{NA} - }{ Disable reordering. }
\item{An \code{hclust} class object}{ E.g. from \code{\link[stats:hclust]{stats::hclust()}}. }
\item{A method name - }{ \code{"ward.D"},
\code{"ward.D2"}, \code{"single"}, \code{"complete"},
\code{"average"}, \code{"mcquitty"}, \code{"median"}, or
\code{"centroid"}. }
}
Default: \code{"complete"}.}

\item{trees}{Draw a dendrogram for rows (left) and columns (top). You can
supply a list or logical vector of length two to control the row tree
and column tree separately, for example
\code{trees = c(rows = T, cols = F)}, or simply \code{trees = c(T, F)}.
Other valid options are \code{"rows"}, \code{"cols"}, \code{"both"},
\code{"left"}, \code{"top"}, and \code{"none"}.
Default: \code{TRUE}.}

\item{tree_height, track_height}{The height of the dendrogram or annotation
tracks in multiples (or fractions) of the smaller dimension of the
grid cell size. Use a numeric vector of length two to assign
\code{c(top, left)} independently.
Default: \code{NULL}, which computes:
\code{tree_height = sqrt(min(dim(mtx))), track_height = tree_height / 4}.}

\item{legend}{Where to place the legend. Options are: \code{"right"} or
\code{"bottom"}. Default: \code{"right"}.}

\item{xlab.angle}{Angle of the labels at the bottom of the plot.
Options are \code{"auto"}, \code{'0'}, \code{'30'}, and \code{'90'}.
Default: \code{"auto"}.}

\item{...}{Additional arguments to pass on to ggplot2::theme().
For example, \code{labs.title = "Plot Title"}.}
}
\value{
A \code{ggplot2} plot. \cr The computed data points, ggplot command,
and object history are available as \code{$data}, \code{$code}, and
\code{$history}, respectively.
}
\description{
Display beta diversities in an all vs all grid.
}
\section{Annotation Tracks}{


Metadata can be displayed as colored tracks above the heatmap. Common use
cases are provided below, with more thorough documentation available at
https://cmmr.github.io/rbiom .

\preformatted{## Categorical ----------------------------
color.by = "Body Site"
color.by = list('Body Site' = "bright")
color.by = list('Body Site' = c("Stool", "Saliva"), 'colors' = "bright")
color.by = list('Body Site' = c('Stool' = "blue", 'Saliva' = "green"))

## Numeric --------------------------------
color.by = "Age"
color.by = list('Age' = "reds")
color.by = list('Age' = c(20,NA), 'colors' = "reds") # at least 20 years old
color.by = list('Age' = c(20,40)) # between 20 and 40 years old (inclusive)

## Multiple Tracks ------------------------
color.by = c("Body Site", "Age")
color.by = list('Body Site' = "bright", 'Age' = "reds")
color.by = list(
  'Body Site' = c('Stool' = "blue", 'Saliva' = "green"),
  'Age'       = list(range = c(20,40), 'colors' = "reds") )
}

The following entries in the track definitions are understood:

\itemize{
\item{\code{colors} - }{ A pre-defined palette name or custom set of colors to map to. }
\item{\code{range} - }{ The c(min,max) to use for scale values. }
\item{\code{label} - }{ Label for this track. Defaults to the name of this list element. }
\item{\code{side} - }{ Options are \code{"top"} (default) or \code{"left"}. }
\item{\code{na.color} - }{ The color to use for \code{NA} values. }
\item{\code{bins} - }{ Bin a gradient into this many bins/steps. }
\item{\code{guide} - }{ A list of arguments for guide_colorbar() or guide_legend(). }
}

All built-in color palettes are colorblind-friendly.

Categorical palette names: \code{"okabe"}, \code{"carto"}, \code{"r4"},
\code{"polychrome"}, \code{"tol"}, \code{"bright"}, \code{"light"},
\code{"muted"}, \code{"vibrant"}, \code{"tableau"}, \code{"classic"},
\code{"alphabet"}, \code{"tableau20"}, \code{"kelly"}, and \code{"fishy"}.

Numeric palette names: \code{"reds"}, \code{"oranges"}, \code{"greens"},
\code{"purples"}, \code{"grays"}, \code{"acton"}, \code{"bamako"},
\code{"batlow"}, \code{"bilbao"}, \code{"buda"}, \code{"davos"},
\code{"devon"}, \code{"grayC"}, \code{"hawaii"}, \code{"imola"},
\code{"lajolla"}, \code{"lapaz"}, \code{"nuuk"}, \code{"oslo"},
\code{"tokyo"}, \code{"turku"}, \code{"bam"}, \code{"berlin"},
\code{"broc"}, \code{"cork"}, \code{"lisbon"}, \code{"roma"},
\code{"tofino"}, \code{"vanimo"}, and \code{"vik"}.
}

\section{Ordering and Limiting}{


\bold{\code{order.by}} controls which metadata column(s) are used to arrange
samples on the plot. It also enables subsetting to a particular set or
range of values. Prefix a column name with \code{-} to arrange values in
descending order rather than ascending.

\preformatted{## Categorical ----------------------------
order.by = "Body Site"
order.by = list('Body Site' = c("Stool", "Saliva"))

## Numeric --------------------------------
order.by = "-Age"
order.by = list('Age'  = c(20,NA)) # at least 20 years old
order.by = list('-Age' = c(20,40)) # between 20 and 40 years old (inclusive)

## Multiple / Mixed -----------------------
order.by = c("-Body Site", "Age")
order.by = list("Body Site", '-Age' = c(20,40))
}

\bold{\code{limit.by}} is used to specify a subset of samples without any
side-effects on aesthetics. It is especially useful for limiting the data
to a single categorical metadata value. Unlike the other *.by parameters,
\code{limit.by} must always be a named \code{list()}.

\preformatted{## Categorical ----------------------------
limit.by = list('Sex' = "Male")

## Numeric --------------------------------
limit.by = list('Age' = c(20,NA)) # at least 20 years old
limit.by = list('Age' = c(20,40)) # between 20 and 40 years old (inclusive)

## Multiple / Mixed -----------------------
limit.by = list(
  'Sex'       = "Male", 
  'Body Site' = c("Stool", "Saliva")
  'Age'       = c(20,40) )
}
}

\examples{
  library(rbiom)
  
  biom <- hmp50 \%>\% sample_rarefy() \%>\% sample_select(1:10)
  bdiv_heatmap(biom, color.by=c("Body Site", "Age"))
  
  bdiv_heatmap(biom, bdiv="uni", weighted=c(T,F), color.by="sex")
    
}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_distmat}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_stats}()},
\code{\link{bdiv_table}()},
\code{\link{distmat_stats}()}

Other visualization: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_barplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()}
}
\concept{beta_diversity}
\concept{visualization}
