% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_ord_plot.r
\name{bdiv_ord_plot}
\alias{bdiv_ord_plot}
\title{Ordinate samples and taxa on a 2D plane based on beta diversity distances.}
\usage{
bdiv_ord_plot(
  biom,
  bdiv = "Bray-Curtis",
  ord = "UMAP",
  weighted = TRUE,
  layers = "petm",
  color.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  tree = NULL,
  test = "adonis2",
  seed = 0,
  permutations = 999,
  rank = -1,
  taxa = 4,
  p.top = Inf,
  p.adj = "fdr",
  unc = "singly",
  caption = TRUE,
  ...
)
}
\arguments{
\item{biom}{An rbiom-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{bdiv}{Beta diversity distance algorithm(s) to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. For \code{"UniFrac"}, a
phylogenetic tree must be present in \code{biom} or explicitly
provided via \code{tree=}. Default: \code{"Bray-Curtis"} \cr\cr
Multiple/abbreviated values allowed.}

\item{ord}{Method for reducing dimensionality. Options are:
\itemize{
\item{\code{"UMAP"} - }{ Uniform manifold approximation and projection; \code{\link[uwot:umap]{uwot::umap()}}. }
\item{\code{"PCoA"} - }{ Principal coordinate analysis; \code{\link[ape:pcoa]{ape::pcoa()}}. }
\item{\code{"NMDS"} - }{ Nonmetric multidimensional scaling; \code{\link[vegan:metaMDS]{vegan::metaMDS()}}. }
\item{\code{"tSNE"} - }{ t-distributed stochastic neighbor embedding; \code{\link[tsne:tsne]{tsne::tsne()}}. }
}
Default: \code{"UMAP"} \cr\cr
Multiple values allowed. Non-ambiguous abbreviations are allowed.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Default: \code{TRUE} \cr\cr
Multiple values allowed.}

\item{layers}{What graphical elements to use for drawing. Options are:
\bold{point}, \bold{spider}, \bold{ellipse}, and \bold{name} for
samples, and \bold{mean}, \bold{taxon}, and \bold{arrow} for taxa.
Single letter abbreviations are also accepted. For instance,
\code{c("point", "ellipse")} is equivalent to \code{c("p", "e")} and
\code{"pe"}.
See the \code{vignette("ordination")} vignette for examples of each.
Default: \code{"pe"}}

\item{color.by, shape.by, facet.by, limit.by}{Metadata columns to use for
aesthetics and partitioning. See below for details.
Default: \code{NULL}}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Only required when
computing UniFrac distances. Default: \code{biom$tree}}

\item{test}{Permutational test for accessing significance. Options are:
\itemize{
\item{\code{"adonis2"} - }{ Permutational MANOVA; \code{\link[vegan:adonis]{vegan::adonis2()}}. }
\item{\code{"mrpp"} - }{ Multiple response permutation procedure; \code{\link[vegan:mrpp]{vegan::mrpp()}}. }
}
Default: \code{"adonis2"} \cr\cr
Abbreviations are allowed.}

\item{seed}{Random seed for permutations. Default: \code{0}}

\item{permutations}{Number of random permutations to use.
Default: \code{999}}

\item{rank}{What rank(s) of taxa to display. E.g. \code{"Phylum"},
\code{"Genus"}, \code{".otu"}, etc. An integer vector can also be
given, where \code{1} is the highest rank, \code{2} is the second
highest, \code{-1} is the lowest rank, \code{-2} is the second
lowest, and \code{0} is the OTU "rank". Run \code{biom$ranks} to
see all options for a given rbiom object. Default: \code{-1}.}

\item{taxa}{Which taxa to display. An integer value will show the top n
most abundant taxa. A value 0 <= n < 1 will show any taxa with that
mean abundance or greater (e.g. \code{0.1} implies >= 10\%). A
character vector of taxa names will show only those named taxa.
Default: \code{6}.}

\item{p.top}{Only display taxa with the most significant differences in
abundance. If \code{p.top} is >= 1, then the \code{p.top} most
significant taxa are displayed. If \code{p.top} is less than one, all
taxa with an adjusted p-value <= \code{p.top} are displayed.
Recommended to be used in combination with the \code{taxa} parameter
to set a lower bound on the mean abundance of considered taxa.
Default: \code{Inf}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}}

\item{unc}{How to handle unclassified, uncultured, and similarly ambiguous
taxa names. Options are:
\itemize{
\item{\code{"singly"} - }{ Replaces them with the OTU name. }
\item{\code{"grouped"} - }{ Replaces them with a higher rank's name. }
\item{\code{"drop"} - }{ Excludes them from the result. }
\item{\code{"asis"} - }{ To not check/modify any taxa names. }
}
Default: \code{"singly"} \cr\cr
Abbreviations are allowed.}

\item{caption}{Add methodology caption beneath the plot.
Default: \code{TRUE}}

\item{...}{Parameters for layer geoms (e.g. \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}).
Prefixing parameter names with a layer name ensures that a particular
parameter is passed to, and only to, that layer. For instance,
\code{point.size = 2} or \code{p.size = 2} ensures only the points
have their size set to \code{2}. Points can also be controlled with
the \code{pt.} prefix.}
}
\value{
A \code{ggplot2} plot. \cr
The computed sample coordinates and ggplot command
are available as \code{$data} and \code{$code} respectively. \cr
If \code{color.by} is given, then \code{$stats} and
\code{$stats$code} are set. \cr
If \code{rank} is given, then \code{$data$taxa_coords},
\code{$taxa_stats}, and \code{$taxa_stats$code} are set.
}
\description{
Ordinate samples and taxa on a 2D plane based on beta diversity distances.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    
    bdiv_ord_plot(biom, layers="pemt", color.by="Body Site", rank="g")
    
}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_distmat}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_stats}()},
\code{\link{distmat_stats}()}

Other ordination: 
\code{\link{bdiv_ord_table}()},
\code{\link{distmat_ord_table}()}

Other visualization: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_barplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()}
}
\concept{beta_diversity}
\concept{ordination}
\concept{visualization}
