% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_ord_table.r
\name{bdiv_ord_table}
\alias{bdiv_ord_table}
\title{Reduce a distance matrix to two or three dimensions}
\usage{
bdiv_ord_table(
  biom,
  dist = "Bray-Curtis",
  ord = "UMAP",
  weighted = TRUE,
  tree = NULL,
  md = NULL,
  k = 2,
  split.by = NULL,
  stat.by = NULL,
  seed = 0,
  perms = 1000,
  rank = NULL,
  taxa = 5,
  p.adj = "fdr",
  p.top = 5,
  ...
)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.
Alternatively, a distance matrix, such as from \link{bdiv_distmat}.}

\item{dist}{The distance algorithm to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"}, 
\code{"Jaccard"}, and \code{"UniFrac"}. A phylogenetic tree must be
present in \code{biom} or explicitly provided via \code{tree=} to use
the UniFrac methods. Ignored if \code{biom} is a distance matrix.
Multiple values allowed. Default: \code{"Bray-Curtis"}.}

\item{ord}{Method for reducing the dimensionality of a distance matrix.
Multiple values allowed. Options are:
\describe{
    \item{\code{"UMAP"} (default)}{
      Uniform manifold approximation and projection,
      via \code{\link[uwot]{umap}}. }
    \item{\code{"PCoA"}}{
      Principal coordinate analysis,
      via \code{\link[ape]{pcoa}}. }
    \item{\code{"NMDS"}}{
      Nonmetric multidimensional scaling,
      via \code{\link[vegan]{metaMDS}}. }
    \item{\code{"tSNE"}}{
      t-distributed stochastic neighbor embedding,
      via \code{\link[tsne]{tsne}}. }
   }}

\item{weighted}{Take relative abundances into account. When 
\code{weighted=FALSE}, only presence/absence is considered.
 Ignored if \code{biom} is a distance matrix. Multiple values 
 allowed. (Default: \code{TRUE})}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Will be taken from the tree
embedded in the \code{biom} object if not explicitly specified. Only
required for computing UniFrac distance matrices.  Ignored if 
\code{biom} is a distance matrix.}

\item{md}{Include metadata in the output data frame? Ignored if 
\code{biom} is a distance matrix. Options are: 
\describe{
  \item{\code{NULL}}{ Don't include metadata. (Default) }
  \item{\code{TRUE}}{ Include all metadata. }
  \item{\emph{character vector}}{ Include only the specified metadata columns. }
}}

\item{k}{Number of dimensions to return. (Default: \code{2})}

\item{split.by}{Name(s) of metadata columns that the data should be split
by prior to calculating distance matrices, ordinations, and statistics.}

\item{stat.by, seed, perms}{Passthrough parameters to \code{bdiv_distmat()} 
for computing adonis statistics. (Default: \code{stat.by=NULL, seed=0, 
perms=1000})}

\item{rank}{[Biplot] What rank of taxa to display (e.g. "Phylum"), or 
\code{NULL} for no biplot. Run \code{taxa_ranks()} to see all options 
for a given BIOM object. (Default: \code{NULL})}

\item{taxa}{[Biplot] Which taxa to display. An integer value will return 
the top n most abundant taxa. A value 0 <= n < 1 will return any taxa 
with that mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. (Default: \code{5})}

\item{p.adj}{[Biplot] Method to use for multiple comparisons adjustment of 
p-values. Run \code{p.adjust.methods} for a list of available options.
(Default: \code{fdr})}

\item{p.top}{[Biplot] Only returns taxa with the most significant 
differences in abundance. If \code{p.top} is >= 1, then the 
\code{p.top} most significant taxa are displayed. If \code{p.top} 
is less than one, all taxa with an adjusted p-value <= \code{p.top} 
are displayed. Recommended to be used in combination with the 
\code{taxa} parameter to set a lower bound on the mean abundance of 
considered taxa. (Default: \code{Inf})}

\item{...}{Additional arguments to pass on to \code{\link[ape]{pcoa}}, 
\code{\link[vegan]{metaMDS}}, or \code{\link[tsne]{tsne}}.}
}
\value{
A data.frame with columns .axis.1, .axis.2, ..., axis.k as well as 
        columns given by \code{md}, \code{split.by}, and \code{stat.by}. 
        Sample IDs are in column \code{'.sample'} and the metric used in 
        column \code{'.metric'}. For pcoa ordinations, 
        \code{attr(, 'eig')} will contain the eigenvalues useful for 
        construction of ".. % variation explained" labels. 
        \code{attr(, 'stats_raw')} and \code{attr(,'stats_tbl')} contain the 
        raw and tabular adonis statistics when \code{stat.by} is set. 
        Distance matrices are in \code{attr(, 'dm')}. When \code{rank} is
        non-NULL, or \code{taxa} is a list of taxon names, a data.frame of
        biplot coordinates will be given by \code{attr(,'biplot')}.
}
\description{
Reduce a distance matrix to two or three dimensions
}
\examples{
    library(rbiom)
    
    ord <- bdiv_ord_table(hmp50, "bray-curtis", "pcoa")
    head(ord)
    
}
