% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_stats.r
\name{bdiv_stats}
\alias{bdiv_stats}
\title{Test beta diversity vs categorical or numeric metadata.}
\usage{
bdiv_stats(
  biom,
  stat.by,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  test = "adonis2",
  within = NULL,
  between = NULL,
  split.by = NULL,
  seed = 0,
  permutations = 999,
  p.adj = "fdr"
)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{stat.by}{The categorical or numeric metadata field over which
statistics should be calculated. Required.}

\item{bdiv}{Beta diversity distance algorithm(s) to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. For \code{"UniFrac"}, a
phylogenetic tree must be present in \code{biom} or explicitly
provided via \code{tree=}. Default: \code{"Bray-Curtis"} \cr\cr
Multiple/abbreviated values allowed.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Default: \code{TRUE} \cr\cr
Multiple values allowed.}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Only required when
computing UniFrac distances. Default: \code{otu_tree(biom)}}

\item{test}{Permutational test for accessing significance. Options are:
\itemize{
\item{\code{"adonis2"} - }{ Permutational MANOVA; \code{\link[vegan:adonis]{vegan::adonis2()}}. }
\item{\code{"mrpp"} - }{ Multiple response permutation procedure; \code{\link[vegan:mrpp]{vegan::mrpp()}}. }
}
Default: \code{"adonis2"} \cr\cr
Abbreviations are allowed.}

\item{within, between}{Metadata field(s) for intra- or inter- sample
comparisons. Default: \code{NULL}}

\item{split.by}{Name(s) of metadata columns that the data should be split
by prior to any calculations. Default: \code{NULL}}

\item{seed}{Random seed for permutations. Default: \code{0}}

\item{permutations}{Number of random permutations to use.
Default: \code{999}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}.}
}
\value{
A tibble data frame with summary statistics. \cr
The R code or generating these statistics is in \code{$code}, and
the object history is in \code{$history}.
}
\description{
Test beta diversity vs categorical or numeric metadata.
}
\examples{
    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    
    bdiv_stats(biom, stat.by = "Body Site", split.by = "Sex")
      
    bdiv_stats(biom, stat.by = "BMI", bdiv = c("bray", "unifrac"))
    
    # The R code used to compute the stats is in $code.
    tbl <- bdiv_stats(biom, stat.by = "Sex")
    tbl$code
}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_distmat}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_table}()},
\code{\link{distmat_stats}()}

Other stats_tables: 
\code{\link{adiv_stats}()},
\code{\link{distmat_stats}()},
\code{\link{stats_table}()},
\code{\link{taxa_stats}()}
}
\concept{beta_diversity}
\concept{stats_tables}
