% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_distmat.r
\name{bdiv_table}
\alias{bdiv_table}
\title{Make a data.frame of distances between samples.}
\usage{
bdiv_table(
  biom,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  md = FALSE
)
}
\arguments{
\item{biom}{A \code{matrix}, \code{simple_triplet_matrix}, or \code{BIOM}
object (e.g. as returned from \code{\link[=read_biom]{read_biom()}}). For matrices, the
rows and columns are assumed to be the taxa and samples, respectively.}

\item{bdiv}{The beta diversity distance algorithm to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. Non-ambiguous abbreviations
of the algorithm names are also accepted. For \code{"UniFrac"}, a
phylogenetic tree must be present in \code{biom} or explicitly
provided via \code{tree=}. Default: \code{"Bray-Curtis"}.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Default: \code{TRUE}}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Will be taken from the tree
defined in \code{biom} if not explicitly specified. Only required
when computing UniFrac distance matrices. Default: \code{NULL}}

\item{md}{Include metadata in the output data frame? Options are:
\itemize{
\item{\code{FALSE} - }{ Don't include metadata. }
\item{\code{TRUE} - }{ Include all metadata. }
\item{\emph{character vector} - }{ Include only the specified metadata
columns. Column names can be prefixed with \bold{==} or \bold{!=} to
indicate that only within or between groupings, respectively, are to
be kept. See examples below. }
}
Default: \code{FALSE}}
}
\value{
A data.frame with first three columns named ".sample1", ".sample2",
and ".distance".
}
\description{
Make a data.frame of distances between samples.
}
\examples{
    library(rbiom)
    
    # Return in long format with metadata
    biom <- sample_select(hmp50, 18:21)
    bdiv_table(biom, 'unifrac', md = c("Body Site", "Sex"))
    
    # Only look at distances among the stool samples
    bdiv_table(biom, 'unifrac', md = c("==Body Site", "Sex"))
    
    # Or between males and females
    bdiv_table(biom, 'unifrac', md = c("Body Site", "!=Sex"))

}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_distmat}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{bdiv_ord_table}()},
\code{\link{distmat_stats}()}
}
\concept{beta_diversity}
