% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiv_table.r
\name{bdiv_table}
\alias{bdiv_table}
\title{Make a data.frame of distances between samples.}
\usage{
bdiv_table(
  biom,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  md = NULL,
  within = NULL,
  between = NULL
)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{bdiv}{Beta diversity distance algorithm(s) to use. Options are:
\code{"Bray-Curtis"}, \code{"Manhattan"}, \code{"Euclidean"},
\code{"Jaccard"}, and \code{"UniFrac"}. For \code{"UniFrac"}, a
phylogenetic tree must be present in \code{biom} or explicitly
provided via \code{tree=}. Default: \code{"Bray-Curtis"} \cr\cr
Multiple/abbreviated values allowed.}

\item{weighted}{Take relative abundances into account. When
\code{weighted=FALSE}, only presence/absence is considered.
Default: \code{TRUE} \cr\cr
Multiple values allowed.}

\item{tree}{A \code{phylo} object representing the phylogenetic
relationships of the taxa in \code{biom}. Only required when
computing UniFrac distances. Default: \code{otu_tree(biom)}}

\item{md}{Include metadata in the output data frame? Options are:
\itemize{
\item{\code{NULL} - }{ Don't include metadata. }
\item{\code{TRUE} - }{ Include all metadata. }
\item{\emph{character vector} - }{
Include only the specified metadata columns. }
}
Default: \code{NULL}}

\item{within, between}{Metadata field(s) for intra- or inter- sample
comparisons. Default: \code{NULL}}
}
\value{
A data.frame with columns named \code{.sample1}, \code{.sample2}, \code{.weighted},
\code{.bdiv}, \code{.distance}, and any fields requested by \code{md}.
}
\description{
Make a data.frame of distances between samples.
}
\examples{
    library(rbiom)
    
    # Return in long format with metadata
    biom <- sample_select(hmp50, 18:21)
    bdiv_table(biom, 'unifrac', md = c("Body Site", "Sex"))
    
    # Only look at distances among the stool samples
    bdiv_table(biom, 'unifrac', md = c("==Body Site", "Sex"))
    
    # Or between males and females
    bdiv_table(biom, 'unifrac', md = c("Body Site", "!=Sex"))

}
\seealso{
Other beta_diversity: 
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_distmat}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{bdiv_ord_table}()},
\code{\link{bdiv_stats}()},
\code{\link{distmat_stats}()}
}
\concept{beta_diversity}
