% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{biom_id}
\alias{biom_id}
\alias{biom_id<-}
\alias{biom_comment}
\alias{biom_comment<-}
\title{Get or set an \code{rbiom} object's id or comment.}
\usage{
biom_id(biom)

biom_id(biom) <- value

biom_comment(biom)

biom_comment(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{value}{The identifier to add (character vector of length 1).}
}
\value{
A length 1 character vector.
}
\description{
The BIOM specification includes \code{id} and \code{comment} fields
for free-form text.
}
\examples{
    library(rbiom) 
    biom <- hmp50
    
    biom_id(biom)
    biom_id(biom) <- "My new title/id"
    biom_id(biom)
    
    biom_comment(biom)
    biom_comment(biom) <- "A description of this study"
    biom_comment(biom)

}
\seealso{
Other biom: 
\code{\link{as_percent}()},
\code{\link{as_rbiom}()},
\code{\link{bdply}()},
\code{\link{biom_build}()},
\code{\link{biom_info}()},
\code{\link{biom_merge}()},
\code{\link{biom_repair}()},
\code{\link{blply}()},
\code{\link{read_biom}()},
\code{\link{write_biom}()},
\code{\link{write_xlsx}()}
}
\concept{biom}
