% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{biom_info}
\alias{biom_info}
\title{Get \code{rbiom} object's miscellaneous information.}
\usage{
biom_info(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
A list of the top-level metadata in \code{biom}.
}
\description{
Get \code{rbiom} object's miscellaneous information.
}
\examples{
    library(rbiom)
    biom_info(hmp50)

}
\seealso{
Other accessors: 
\code{\link{metadata_numeric}()},
\code{\link{n_samples}()},
\code{\link{sample_sums}()}
}
\concept{accessors}
