% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{biom_info}
\alias{biom_info}
\title{Get \code{rbiom} object's miscellaneous information.}
\usage{
biom_info(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
A list of the top-level metadata in \code{biom}.
}
\description{
Get \code{rbiom} object's miscellaneous information.
}
\examples{
    library(rbiom)
    biom_info(hmp50)

}
\seealso{
Other biom: 
\code{\link{as_percent}()},
\code{\link{as_rbiom}()},
\code{\link{bdply}()},
\code{\link{biom_build}()},
\code{\link{biom_id}()},
\code{\link{biom_merge}()},
\code{\link{biom_repair}()},
\code{\link{blply}()},
\code{\link{read_biom}()},
\code{\link{write_biom}()},
\code{\link{write_xlsx}()}
}
\concept{biom}
