% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biom_build.r
\name{biom_merge}
\alias{biom_merge}
\title{Combine several BIOM datasets into one.}
\usage{
biom_merge(...)
}
\arguments{
\item{...}{Any number of BIOM objects (e.g. from \code{\link[=read_biom]{read_biom()}}), lists of
BIOM objects, or valid arguments to the \code{src} parameter of
\code{\link[=read_biom]{read_biom()}} (for instance file names).}
}
\value{
A \code{BIOM} class object containing the parsed data. This object
can be treated as a list with the following named elements:
\itemize{
\item{\code{$counts} - }{
A numeric \code{slam} sparse matrix of observation counts. Taxa
(OTUs) as rows and samples as columns. }
\item{\code{$metadata} - }{
A data frame containing any embedded metadata. Row names are
sample IDs. }
\item{\code{$taxonomy} - }{
Character matrix of taxonomic names, if given. Row names are taxa
(OTU) IDs. Column rows are named Kingdom, Phylum, Class, Order,
Family, Genus, Species, and Strain, or TaxLvl.1, TaxLvl.2, ... ,
TaxLvl.N when more than 8 levels of taxonomy are encoded in the
biom file. }
\item{\code{$phylogeny} - }{
An object of class \code{phylo} defining the phylogenetic
relationships between the taxa. Although the official
specification for BIOM only includes phylogenetic trees in BIOM
version 2.1, if a BIOM version 1.0 file includes a
\code{phylogeny} entry with newick data, then it will be loaded
here as well. The \pkg{ape} package has additional functions for
working with \code{phylo} objects.}
\item{\code{$sequences} - }{
A named character vector, where the names are taxonomic
identifiers and the values are the sequences they represent.
These values are not part of the official BIOM specification, but
will be read and written when defined. }
\item{\code{$info} - }{
A list of other attributes defined in the BIOM file, such as
\code{id}, \code{type}, \code{format}, \code{format_url},
\code{generated_by}, \code{date}, \code{matrix_type},
\code{matrix_element_type}, \code{Comment}, and \code{shape}. }
}
\code{metadata}, \code{taxonomy}, and \code{phylogeny} are optional
components of the BIOM file specification and therefore will be empty
in the returned object when they are not provided by the BIOM file.
}
\description{
WARNING: It is generally ill-advised to merge BIOM datasets, as OTUs
mappings are dependent on upstream clustering and are not equivalent
between BIOM files.
}
\examples{
    library(rbiom)
    
    b1 <- sample_select(hmp50, 1:4)
    b2 <- sample_select(hmp50, 5:8)
    
    biom <- biom_merge(b1, b2)
    print(biom)
    
    otu_tree(biom) <- otu_tree(hmp50)
    print(biom)
}
