% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_rbiom.r
\name{biom_repair}
\alias{biom_repair}
\title{Run after manually editing an rbiom object's content.}
\usage{
biom_repair(biom)
}
\arguments{
\item{biom}{The \code{rbiom} object to repair.}
}
\value{
An \code{rbiom} object.
}
\description{
Run after manually editing an rbiom object's content.
}
