% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blply.r
\name{blply}
\alias{blply}
\title{Split BIOM by metadata, apply function, and return results in a list.}
\usage{
blply(biom, vars, FUN, ..., fast = TRUE)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read.biom}.}

\item{vars}{A character vector of metadata fields. Each unique combination
of values in these columns will be used to create a subsetted BIOM
object to pass to \code{FUN}. Unambiguous abbreviations of metadata
fields are also accepted.}

\item{FUN}{The function to execute on each BIOM subset. \code{FUN} may
return any object, all of which will be returned in a named list.}

\item{...}{Additional arguments to pass on to \code{FUN}.}

\item{fast}{If \code{TRUE} (the default), the subsetted BIOM objects will
still contain the full taxa table and phylogenetic tree. Set 
\code{fast = FALSE} to run the slow steps of subsetting these 
elements as well.}
}
\value{
A list with the function outputs.
}
\description{
Calls \code{plyr::dlply} internally.
}
\examples{
    library(rbiom)
    
    blply(hmp50, "Sex", nsamples)
    
    blply(hmp50, c("Body Site", "Sex"), function (b) {
      ad <- alpha.div(b)[,c("Shannon", "Simpson")]
      apply(ad, 2L, mean)
    })
    
    blply(hmp50, "Body Site", function (b) {
      r <- range(beta.div(b, "bray"))
      data.frame(bray.min = r[[1]], bray.max = r[[2]])
    })
    
    

}
