% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blply.r
\name{blply}
\alias{blply}
\title{Split an rbiom object by metadata, apply function, and return results in a
list.}
\usage{
blply(biom, vars, FUN, ..., iters = list(), prefix = FALSE)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{vars}{A character vector of metadata fields. Each unique combination
of values in these columns will be used to create a subsetted
\code{rbiom} object to pass to \code{FUN}. If \code{NULL},
\code{biom} will be passed to \code{FUN} unaltered. Unambiguous
abbreviations of metadata fields are also accepted.}

\item{FUN}{The function to execute on each \code{biom} subset. \code{FUN}
may return any object, all of which will be returned in a named list.}

\item{...}{Additional arguments to pass on to \code{FUN}.}

\item{iters}{A named list of values to pass to \code{FUN}. Unlike
\code{...}, these will be iterated over in all combinations.
Default: \code{list()}}

\item{prefix}{When \code{TRUE}, prefixes the names in in \code{iters} with
a '.' in the split_labels attribute of the returned object.
Default: \code{FALSE}}
}
\value{
A list with the function outputs.
}
\description{
Calls \code{plyr::dlply} internally. Consider setting
\code{otu_tree(biom) <- NULL} to speed up creation of subseted rbiom objects.
}
\examples{
    library(rbiom) 
    
    blply(hmp50, "Sex", n_samples)
    
    blply(hmp50, c("Body Site", "Sex"), function (b) {
      adm <- adiv_matrix(b)[,c("Shannon", "Simpson")]
      apply(adm, 2L, mean)
    })
    
    iters <- list(w = c(TRUE, FALSE), d = c("bray", "euclid"))
    blply(hmp50, "Sex", iters = iters, function (b, w, d) {
      r <- range(bdiv_distmat(biom = b, bdiv = d, weighted = w))
      round(data.frame(min = r[[1]], max = r[[2]]))
    })
    
    

}
