% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert_to}
\alias{convert_to}
\alias{convert_to_animalcules}
\alias{convert_to_biomformat}
\alias{convert_to_phyloseq}
\alias{convert_to_SE}
\alias{convert_to_TSE}
\title{Convert biom data to an external package class.}
\usage{
convert_to_animalcules(biom, ...)

convert_to_biomformat(biom, ...)

convert_to_phyloseq(biom, ...)

convert_to_SE(biom, ...)

convert_to_TSE(biom, ...)
}
\arguments{
\item{biom}{An \link[=rbiom_objects]{rbiom object}, or any value accepted by \code{\link[=as_rbiom]{as_rbiom()}}.}

\item{...}{Not Used.}
}
\value{
An animalcules (MultiAssayExperiment), biomformat (biom), phyloseq, SummarizedExperiment, or TreeSummarizedExperiment object.
}
\description{
Requires the relevant Bioconductor R package to be installed:
\describe{
\item{\code{convert_to_animalcules} - }{ \href{https://bioconductor.org/packages/animalcules/}{animalcules} }
\item{\code{convert_to_biomformat} - }{ \href{https://bioconductor.org/packages/biomformat/}{biomformat} }
\item{\code{convert_to_phyloseq} - }{ \href{https://bioconductor.org/packages/phyloseq/}{phyloseq} }
\item{\code{convert_to_SE} - }{ \href{https://bioconductor.org/packages/SummarizedExperiment/}{SummarizedExperiment} }
\item{\code{convert_to_TSE} - }{ \href{https://bioconductor.org/packages/TreeSummarizedExperiment/}{TreeSummarizedExperiment} }
}
}
\details{
animalcules, SummarizedExperiment, and biomformat objects include counts, metadata, and taxonomy.

phyloseq and TreeSummarizedExperiment additionally include the tree and sequences.
}
\examples{
\dontrun{
    library(rbiom)

    print(hmp50)

    # Requires 'animalcules', a Bioconductor R package
    if (nzchar(system.file(package = "animalcules"))) {
      ani <- convert_to_animalcules(hmp50)
      print(ani)
    }

    # Requires 'biomformat', a Bioconductor R package
    if (nzchar(system.file(package = "biomformat"))) {
      bio <- convert_to_biomformat(hmp50)
      print(bio)
    }

    # Requires 'phyloseq', a Bioconductor R package
    if (nzchar(system.file(package = "phyloseq"))) {
      phy <- convert_to_phyloseq(hmp50)
      print(phy)
    }

    # Requires 'SummarizedExperiment', a Bioconductor R package
    if (nzchar(system.file(package = "SummarizedExperiment"))) {
      se <- convert_to_SE(hmp50)
      print(se)
    }

    # Requires 'TreeSummarizedExperiment', a Bioconductor R package
    if (nzchar(system.file(package = "TreeSummarizedExperiment"))) {
      tse <- convert_to_TSE(hmp50)
      print(tse)
    }
}
}
