% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{counts}
\alias{counts}
\title{Get the abundance counts.}
\usage{
counts(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
A numeric matrix of the sample abundance counts in \code{biom}.
}
\description{
Get the abundance counts.
}
\examples{
    library(rbiom)
    
    counts(hmp50)[1:4,1:5]

}
\seealso{
Other accessors: 
\code{\link{comments}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
