% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{depth}
\alias{depth}
\title{The rarefaction depth of a BIOM object.}
\usage{
depth(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
The rarefaction depth. If the BIOM object is not rarefied, this will 
        be a sorted vector of all the unique depths.
}
\description{
The returned values reflect changes made after any \code{subset()} commands. 
To get the rarefaction level immediately after \code{read_biom()} or
\code{rarefy()} use \code{attr(biom, 'rarefaction')} instead.
}
\examples{
    library(rbiom)
    
    depth(hmp50) \%>\% head()
    rarefy(hmp50) \%>\% depth()

}
\seealso{
[sample_sums()]

Other accessors: 
\code{\link{comments}()},
\code{\link{counts}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
