% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_plot_return}
\alias{documentation_plot_return}
\title{documentation_plot_return}
\value{
A \code{ggplot2} plot. The computed data points and ggplot
command are available as \verb{$data} and \verb{$code},
respectively.
}
\description{
documentation_plot_return
}
\keyword{internal}
