% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.r
\name{documentation_plot_stats_return}
\alias{documentation_plot_stats_return}
\title{documentation_plot_stats_return}
\value{
A \code{ggplot2} plot. \cr The computed data points, statistics,
ggplot command, and object history are available as \code{$data},
\code{$stats}, \code{$code}, and \code{$history}, respectively.
}
\description{
documentation_plot_stats_return
}
\keyword{internal}
