% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gems}
\alias{gems}
\title{Global Enteric Multicenter Study (n = 1,006)}
\format{
An rbiom object with 1,006 samples.
Includes metadata and taxonomy.
\describe{
\item{diarrhea - }{Case or Control}
\item{age - }{0 - 4.8 (years old)}
\item{country - }{Bangladesh, Gambia, Kenya, or Mali}
}
}
\source{
\doi{10.1186/gb-2014-15-6-r76} and \doi{10.1093/nar/gkx1027}
}
\usage{
gems
}
\description{
Global Enteric Multicenter Study (n = 1,006)
}
\keyword{datasets}
