% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{has_tree}
\alias{has_tree}
\title{Check if a phylogenetic tree is included in an rbiom object.}
\usage{
has_tree(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
\code{TRUE} if it has the property, \code{FALSE} otherwise.
}
\description{
Check if a phylogenetic tree is included in an rbiom object.
}
\examples{
    library(rbiom)
    
    has_tree(hmp50)

}
\seealso{
Other phylogeny: 
\code{\link{otu_tree}()},
\code{\link{read_tree}()},
\code{\link{tree_data}()},
\code{\link{tree_subset}()},
\code{\link{tree_tips}()},
\code{\link{write_tree}()}
}
\concept{phylogeny}
