% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{id<-}
\alias{id<-}
\title{Set a \code{BIOM} object's id.}
\usage{
id(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read_biom}.}

\item{value}{The identifier to add (character vector of length 1).}
}
\description{
The BIOM specification includes \code{id} and \code{comment} fields
for free-form text.
}
\examples{
    library(rbiom)
    biom <- hmp50
    
    info(biom)$id
    
    id(biom) <- "My new title/id"
    info(biom)$id

}
\seealso{
Other setters: 
\code{\link{comments<-}()},
\code{\link{counts<-}()},
\code{\link{metadata<-}()},
\code{\link{phylogeny<-}()},
\code{\link{sample_names<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa_names<-}()},
\code{\link{taxa_ranks<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
