% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{id}
\alias{id}
\title{Get \code{BIOM} object's identifier / title.}
\usage{
id(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
A data frame of the metadata in \code{biom}.
}
\description{
Get \code{BIOM} object's identifier / title.
}
\examples{
    library(rbiom)
    id(hmp50)

}
\seealso{
Other accessors: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
