% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{info}
\alias{info}
\title{Get \code{BIOM} object's miscellaneous information.}
\usage{
info(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
A list of the top-level metadata in \code{biom}.
}
\description{
Get \code{BIOM} object's miscellaneous information.
}
\examples{
    library(rbiom)
    info(hmp50)

}
\seealso{
Other accessors: 
\code{\link{comments}()},
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
