% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{info}
\alias{info}
\title{Get \code{BIOM} object's miscellaneous information.}
\usage{
info(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
A list of the top-level metadata in \code{biom}.
}
\description{
Get \code{BIOM} object's miscellaneous information.
}
\examples{
    library(rbiom)
    info(hmp50)

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{id}()},
\code{\link{is.rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
