% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{is_rarefied}
\alias{is_rarefied}
\title{Check if the rbiom object is rarefied.}
\usage{
is_rarefied(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
\code{TRUE} if \code{biom} is rarefied, \code{FALSE} otherwise.
}
\description{
Check if the rbiom object is rarefied.
}
\examples{
    library(rbiom)
    
    is_rarefied(hmp50)

}
\seealso{
Other rarefaction: 
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_depth}()},
\code{\link{rare_multiplot}()},
\code{\link{rare_suggest}()},
\code{\link{sample_rarefy}()},
\code{\link{sample_sums}()}
}
\concept{rarefaction}
