% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{metadata}
\alias{metadata}
\title{Get the sample metadata.}
\usage{
metadata(biom, field = NULL, id = FALSE)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}

\item{field}{The name of the metadata field to retrieve (optional).}

\item{id}{Include the sample names in an 'id' column in addition to 
rownames (Default: FALSE).}
}
\value{
A data frame of the metadata in \code{biom}. If \code{field} is
   given, will return a named vector of the field's values.
}
\description{
Get the sample metadata.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom <- read.biom(infile)
    
    metadata(biom)[1:4,1:3]

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{info}()},
\code{\link{is.rarefied}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
