% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.r
\name{metadata}
\alias{metadata}
\title{Get the sample metadata.}
\usage{
metadata(biom, field = NULL, id = NULL, cleanup = FALSE)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}

\item{field}{The name of a single metadata column to retrieve. If provided,
a named vector will be returned instead of a data.frame.
(Default: NULL)}

\item{id}{Copy the sample names (rownames) into this column. 
(Default: NULL)}

\item{cleanup}{Change character columns to factor or dates. Date formats 
recognized are: "%Y-%m-%d" or "%m/%d/%Y" (four-digit year). Also
'.' with '_' in column names when '.' is the first character in order
to avoid conflicts with auto-generated columns. (Default: FALSE)}
}
\value{
A data frame of the metadata in \code{biom}. If \code{field} is
   given, will return a named vector of the field's values.
}
\description{
Get the sample metadata.
}
\examples{
    library(rbiom)
    
    metadata(hmp50)[1:4,1:3]
    
    head(metadata(hmp50, "Body Site"))
    
    head(metadata(hmp50, "Body Site", cleanup=TRUE))

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{info}()},
\code{\link{is.rarefied}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
