% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{metrics}
\alias{metrics}
\title{List all the options for each type of metric.}
\usage{
metrics(biom, mode = "all", tree = NULL)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{mode}{One of the following options:
\itemize{
  \item{\bold{ord} - }{ Ordination }
  \item{\bold{adiv} - }{ Alpha Diversity }
  \item{\bold{dist} - }{ Distance (Beta Diversity) }
  \item{\bold{clust} - }{ Clustering }
  \item{\bold{rank} - }{ Taxonomic Rank }
  \item{\bold{taxon} - }{ Taxa Names }
  \item{\bold{meta} - }{ Metadata Fields }
  \item{\bold{all} - }{ All of the Above }
}}
}
\value{
A character vector of supported values. 
        For \code{mode = "all"}, a named \code{list()} of character vectors.
}
\description{
List all the options for each type of metric.
}
\examples{
    library(rbiom)
    
    metrics(hmp50, 'adiv')
    metrics(hmp50, 'dist')

}
\seealso{
\code{\link{adiv_table}} \code{\link{bdiv_table}} \code{\link{ordinate}}
}
