% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{n_samples}
\alias{n_samples}
\alias{n_otus}
\alias{n_ranks}
\title{Report the number of samples, otus, or ranks in a BIOM.}
\usage{
n_samples(biom)

n_otus(biom)

n_ranks(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}
}
\value{
The number of samples, otus, or ranks present.
}
\description{
Report the number of samples, otus, or ranks in a BIOM.
}
\examples{
    library(rbiom)
    
    n_samples(hmp50)
    n_otus(hmp50)
    n_ranks(hmp50)

}
\seealso{
Other accessors: 
\code{\link{biom_info}()},
\code{\link{sample_sums}()}

Other accessors: 
\code{\link{biom_info}()},
\code{\link{sample_sums}()}

Other accessors: 
\code{\link{biom_info}()},
\code{\link{sample_sums}()}
}
\concept{accessors}
