% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{n_samples}
\alias{n_samples}
\alias{n_otus}
\alias{n_ranks}
\alias{n_metadata}
\title{Report the number of samples, otus, or ranks in an rbiom.}
\usage{
n_samples(biom)

n_otus(biom)

n_ranks(biom)

n_metadata(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
The number of samples, otus, taxa ranks, or metadata columns present.
}
\description{
Report the number of samples, otus, or ranks in an rbiom.
}
\examples{
    library(rbiom)
    
    n_samples(hmp50)
    n_otus(hmp50)
    n_ranks(hmp50)
    n_metadata(hmp50)

}
\seealso{
Other accessors: 
\code{\link{biom_info}()},
\code{\link{metadata_numeric}()},
\code{\link{sample_sums}()}
}
\concept{accessors}
