% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{nsamples}
\alias{nsamples}
\title{Number of samples in a BIOM.}
\usage{
nsamples(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
The number of samples present.
}
\description{
Number of samples in a BIOM.
}
\examples{
    library(rbiom)
    nsamples(hmp50)

}
\seealso{
Other accessors: 
\code{\link{comments}()},
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
