% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{otu_matrix}
\alias{otu_matrix}
\alias{otu_matrix<-}
\title{Get or set the abundance counts.}
\usage{
otu_matrix(biom)

otu_matrix(x) <- value
}
\arguments{
\item{biom, x}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{value}{A numeric matrix. Rownames and colnames must be identical to
the current \code{\link[=otu_matrix]{otu_matrix()}} value.}
}
\value{
A numeric matrix with samples as row names, and OTU identifiers as
column names.
}
\description{
Get or set the abundance counts.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    taxa <- c('Unc53100', 'Unc00c7g', 'Unc25731')
    
    otu_matrix(biom)[taxa,1:5]
    
    biom <- sample_rarefy(biom, 100)
    otu_matrix(biom)[taxa,1:5]
    
    otu_matrix(biom) <- otu_matrix(biom) / 100
    otu_matrix(biom)[taxa,1:5]
}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
