% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{otu_sequences}
\alias{otu_sequences}
\alias{otu_sequences<-}
\title{Get or set the nucleotide sequences associated with each OTU.}
\usage{
otu_sequences(biom)

otu_sequences(x) <- value
}
\arguments{
\item{biom, x}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{value}{A named character vector. Names must match \code{otu_names(x)}.}
}
\value{
A named character vector of sequences in \code{biom}. If this data
is not present, then returns \code{NULL}.
}
\description{
Get or set the nucleotide sequences associated with each OTU.
}
\examples{
    library(rbiom) 
    
    substr(otu_sequences(hmp50)[1:4], 1, 10)
    
\dontrun{
    # Write to a compressed fasta file in the temporary directory:
    seqs <- otu_sequences(hmp50)
    conn <- bzfile(file.path(tempdir(), "Sequences.fa.bz2"), "w")
    cat(sprintf(">\%s\n\%s", names(seqs), seqs), file=conn, sep="\n")
    close(conn)
    
    # You can also use the write_fasta function for this task:
    write_fasta(hmp50, file.path(tempdir(), "Sequences.fa.gz"))
    
    # Set/replace the sequences
    biom <- read_biom('path/to/file.biom')
    otu_sequences(biom) <- read_fasta('path/to/sequences.fa')
    otu_sequences(biom) <- c(OTU1 = 'ATCGGGTA', OTU2 = 'GGCATTAGC')
}

}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
