% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_taxonomy.r
\name{otu_taxonomy}
\alias{otu_taxonomy}
\alias{otu_taxonomy<-}
\title{Get or set the taxonomy table.}
\usage{
otu_taxonomy(biom, rank = NULL, unc = "asis", lineage = FALSE)

otu_taxonomy(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{rank}{A single taxonomic rank to return as a named vector, or
\code{NULL} to return the complete table. Default: \code{NULL}}

\item{unc}{How to handle unclassified, uncultured, and similarly ambiguous
taxa names. Options are:
\itemize{
\item{\code{"singly"} - }{ Replaces them with the OTU name. }
\item{\code{"grouped"} - }{ Replaces them with a higher rank's name. }
\item{\code{"drop"} - }{ Excludes them from the result. }
\item{\code{"asis"} - }{ To not check/modify any taxa names. }
}
Default: \code{"singly"} \cr\cr
Abbreviations are allowed.}

\item{lineage}{Include all ranks in the name of the taxa. For instance,
setting to \code{TRUE} will produce
\code{Bacteria; Actinobacteria; Coriobacteriia; Coriobacteriales}.
Otherwise the taxa name will simply be \code{Coriobacteriales}. You
want to set this to TRUE when \code{unc = "asis"} and you have taxa
names (such as \emph{Incertae_Sedis}) that map to multiple higher
level ranks. Default: \code{FALSE}}

\item{value}{An object coercible with \code{\link[tibble:as_tibble]{tibble::as_tibble()}}. Must either
have rownames or an '.otu' column with OTU names matching
\link{otu_names(x)}.}
}
\value{
Depending on \code{rank}, a named character vector or a tibble data
frame with '.otu' as the first column name.
}
\description{
Get or set the taxonomy table.
}
\examples{
    library(rbiom) 
    
    # Display the full taxonomic data for each OTU ------------------------
    otu_taxonomy(hmp50)
    
    # Only show the OTU -> Genus mapping  ---------------------------------
    otu_taxonomy(hmp50, "Genus") \%>\% head()
    
    # Sometimes taxonomic names are incomplete ----------------------------
    otu_taxonomy(hmp50)[,4:7] \%>\%
      dplyr::filter(.otu \%in\% c('GemAsacc', 'GcbBacte', 'Unc58411'))
    
    # rbiom can insert more descriptive placeholders ----------------------
    otu_taxonomy(hmp50, unc = "singly")[,4:7] \%>\%
      dplyr::filter(.otu \%in\% c('GemAsacc', 'GcbBacte', 'Unc58411'))
    
    # Or collapse them into groups ----------------------------------------
    otu_taxonomy(hmp50, unc = "grouped")[,4:7] \%>\%
      dplyr::filter(.otu \%in\% c('GemAsacc', 'GcbBacte', 'Unc58411'))

}
\seealso{
Other taxonomy: 
\code{\link{n_ranks}()},
\code{\link{taxa_ranks}()},
\code{\link{write_taxonomy}()}
}
\concept{taxonomy}
