% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{otu_tree}
\alias{otu_tree}
\alias{otu_tree<-}
\title{Get or set the phylogenetic tree.}
\usage{
otu_tree(biom)

otu_tree(x) <- value
}
\arguments{
\item{biom, x}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{value}{A \code{phylo} class object with tip.labels matching
\code{otu_names(x)}. If there are more tip.labels than
taxa names, then the tree will be subset.}
}
\value{
A \code{phylo} class object of the tree in \code{biom}.
}
\description{
Get or set the phylogenetic tree.
}
\examples{
\dontrun{
    library(rbiom)
    
    biom <- hmp50
    otu_tree(biom) <- read_tree('path/to/newick.tre')
}

}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
