% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{phylogeny<-}
\alias{phylogeny<-}
\title{Set the phylogenetic tree.}
\usage{
phylogeny(x) <- value
}
\arguments{
\item{x}{A \code{BIOM} object, as returned from \link{read_biom}.}

\item{value}{A \code{phylo} class object with tip.labels matching 
\code{taxa_names(x)}. If there are more tip.labels than
taxa names, then the tree will be subset.}
}
\description{
Set the phylogenetic tree.
}
\examples{
\dontrun{
    library(rbiom)
    
    biom <- hmp50
    phylogeny(biom) <- read_tree('path/to/newick.tre')
}

}
\seealso{
Other setters: 
\code{\link{counts<-}()},
\code{\link{id<-}()},
\code{\link{metadata<-}()},
\code{\link{sample_names<-}()},
\code{\link{sequences<-}()},
\code{\link{taxa_names<-}()},
\code{\link{taxa_ranks<-}()},
\code{\link{taxonomy<-}()}
}
\concept{setters}
