% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot}
\alias{plot}
\alias{plot.BIOM}
\title{Visualize BIOM data.}
\usage{
\method{plot}{BIOM}(
  x,
  formula,
  layers = "rls",
  color.by = NULL,
  pattern.by = NULL,
  shape.by = NULL,
  label.by = NULL,
  sort.by = NULL,
  facet.by = NULL,
  xvals = NULL,
  colors = NULL,
  patterns = NULL,
  shapes = NULL,
  facets = NULL,
  rank = NULL,
  taxa = NULL,
  p.top = Inf,
  p.adj = "fdr",
  p.label = 0.05,
  perms = 1000,
  ci = 95,
  rline = NULL,
  xlab.angle = "auto",
  weighted = TRUE,
  abbr = TRUE,
  other = FALSE,
  gradient = heat.colors(20),
  normalize.rows = TRUE,
  dist = "euclidean",
  model = y ~ x,
  regr = lm,
  safe = FALSE,
  ...
)
}
\arguments{
\item{x}{A BIOM object, as returned from \link{read.biom}.}

\item{formula}{Combination of terms to plot, either in \code{y ~ x} or
\code{c("x", "y")} form.}

\item{layers}{See "Layers" section for details. Options for super box
plots are \bold{box}, \bold{bar} (r), \bold{violin}, \bold{dot},
\bold{strip}, \bold{crossbar}, \bold{errorbar}, \bold{linerange}, and
\bold{pointrange}. Options for ordination plots are: \bold{point},
\bold{spider}, \bold{ellipse}, and \bold{name} for samples,
and \bold{mean}, \bold{taxon}, and \bold{arrow} for taxa biplots.
Single letter abbreviations are also accepted. For instance,
\code{c("box", "dot")} is equivalent to \code{c("b", "d")} and
\code{"bd"}. Default: "rls" for super box plots and "pce"/"p" for
ordinations with/without a \code{color.by} argument.}

\item{color.by, pattern.by, shape.by, label.by, sort.by, facet.by}{Metadata
column to color, pattern, shape, label, sort, and/or facet by. If
that column is a \code{factor}, the ordering of levels will be
maintained in the plot.}

\item{colors, patterns, shapes, facets, xvals}{Names of the colors, patterns,
shapes, facets, and/or x values to use in the plot. Available values
for colors, patterns, and shapes are given by \code{colors()},
\code{gridpattern::names_magick}, and \code{0:25},
respectively. Use a named character vector to map them  to specific
factor levels in the metadata. \code{facets} and \code{xvals} are
coerced to unnamed character vectors. If the length of these vectors
is less than the values present in their corresponding metadata
column, then the data set will be subseted accordingly.}

\item{rank}{What rank of taxa to display. E.g. "Phylum", "Genus", etc. Run
\code{taxa.ranks()} to see all options for a given BIOM object. The
default, \code{NULL}, selects the lowest level.}

\item{taxa}{Which taxa to display. An integer value will show the top n
most abundant taxa. A value 0 <= n < 1 will show any taxa with that
mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. (Default: \code{10} for
heatmaps, \code{5} otherwise.)}

\item{p.top}{For \code{RANK ~ .} (taxa abundance) plots or
\code{BDIV ~ ORD} ordination biplots, only display taxa with the most
significant differences in abundance. If \code{p.top} is >= 1, then
the \code{p.top} most significant taxa are displayed. If \code{p.top}
is less than one, all taxa with an adjusted p-value <= \code{p.top}
are displayed. Recommended to be used in combination with the
\code{taxa} parameter to set a lower bound on the mean abundance of
considered taxa. (Default: \code{Inf})}

\item{p.adj}{Method to use for multiple comparisons adjustment of p-values.
Run \code{p.adjust.methods} for a list of available options.
(Default: \code{fdr})}

\item{p.label}{Minimum adjusted p-value to display on the plot with a bracket.
Set to \code{Inf} to display all p-values, or \code{-Inf} for no brackets.
If a numeric vector with more than one value is provided, they will be
used as breaks for asterisk notation. For ordinations, \code{p.label}
applies to biplot taxa. (Default: \code{0.05})}

\item{perms}{Number of random permutations to use for estimating statistical
significance in \code{BDIV ~ ORD} ordinations. (Default: \code{1000})}

\item{ci}{How to calculate min/max of the \bold{crossbar},
\bold{errorbar}, \bold{linerange}, and \bold{pointrange} layers.
Provide a number between 75 and 100 to define a confidence interval's
confidence level, commonly 95 or 97.5. Other options are:
\bold{range},
\bold{sd} (standard deviation),
\bold{se} (standard error), and
\bold{mad} (median absolute deviation).
The center mark of \code{crossbar} and \code{pointrange} represents
the mean, except for \bold{mad} in which case it represents
the median. Trendlines require a confidence interval value.
Set to \code{NULL} to disable. Default: \code{95}}

\item{rline}{On rarefaction plots, highlight this rarefaction depth with a
dashed line. (Default: NULL)}

\item{xlab.angle}{How to rotate the tick labels on the x-axis.
\bold{'auto'} (the default), automatically selects a rotation value.
\bold{0}, \bold{30}, and \bold{90} sets the angle to horizontal,
angled, and vertical, respectively.}

\item{weighted}{When employing a beta diversity metric, use the weighted
version. (Default: \code{TRUE})}

\item{abbr}{When selecting taxa by name, allow abbreviated 'taxa' values,
e.g. \code{c('staph', 'lact')}. (Default: TRUE).}

\item{other}{Should non-selected taxa be displayed as an "Other" group?
(Default: \code{FALSE})}

\item{gradient}{For heatmaps, the color gradient to use for the cells.
(Default: \code{heat.colors(20)})}

\item{normalize.rows}{For heatmaps, should each row (taxon) have its
values rescaled from min-max to 0-1. (Default: \code{TRUE})}

\item{dist}{For \code{RANK ~ CLUST}, the distance metric to use.
(Default: \code{"euclidean"})}

\item{model}{For regressions, the formula to use in the smoothing function.
For example: \code{y ~ x}, \code{y ~ log(x)}, or \code{y ~ poly(x,2)}.
(Default: \code{y ~ x})}

\item{regr}{For \code{BDIV ~ ORD}, the number of random permutations to
use when calculating adonis statistics. (Default: \code{999})}

\item{safe}{If \code{FALSE}, data.frame column names such as
\code{".metric"} will be auto-converted to \code{"Metric"} to improve
human-readability. Conversion if aborted if a conflict is found with
a metadata column name. (Default: \code{FALSE})}

\item{...}{Parameters for underlying functions.
See "Additional Parameters" section for details.
For heatmaps, \code{...} is handled by \code{pheatmap::pheatmap()}.
Otherwise, parameters are matched to formal arguments of ggplot2
functions. Prefixing parameter names with a layer name ensures that
a particular parameter is passed to, and only to, that layer. For
instance, \code{dot.size = 2} or \code{d.size = 2} ensures only the
dotplot layer has its size set to \code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will
be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')},
respectively.

Shapes can also be given as their string values, defined in pch_table here:
https://github.com/tidyverse/ggplot2/blob/master/R/geom-point.r . Note that
some shapes have a colored outline given by \code{color}, some are filled with
\code{color}, and some are outlined in \code{color} and filled with \code{fill}. See
https://blog.albertkuo.me/post/point-shape-options-in-ggplot/ for details.

To expand the low end of the y axis, you can set \code{y.trans = "sqrt"} or
\code{y.trans = "log1p"}. The former applies a square-root transformation,
and the latter plots log(y + 1). Both of these methods work well with data
that contains zeroes.
}
\description{
Provide two terms to \code{plot()} and it will automatically produce the
appropriate chart type. The terms that \code{plot} understands and valid
combinations are listed in the Details section. Not all parameters are used
by all chart types.
}
\section{Terms}{

Terms come in the nine categories given below. The values are
case-insensitive and can be unambiguously abbreviated.
\itemize{
\item{\bold{Ordination Method (ORD): }}{
\code{PCoA}, \code{tSNE}, or \code{NMDS}.
}
\item{\bold{Alpha Diversity Metric (ADIV): }}{
\code{OTUs}, \code{Shannon}, \code{Chao1}, \code{Simpson},
and \code{InvSimpson}. Multiple allowed.
}
\item{\bold{Beta Diversity Metric (BDIV): }}{
\code{Manhattan}, \code{Euclidean}, \code{Bray-Curtis}, \code{Jaccard},
or \code{UniFrac}.
}
\item{\bold{Distance Formula (DIST): }}{
\code{correlation}, \code{euclidean}, \code{maximum}, \code{manhattan},
\code{canberra}, \code{binary}, or \code{minkowski}.
}
\item{\bold{Taxonomic Rank (RANK): }}{
\code{Kingdom}, \code{Phylum}, \code{Class}, \code{Order},
\code{Family}, \code{Genus}, \code{Species}, \code{Strain}, and
\code{OTU}. Multiple allowed. Supported ranks will vary by BIOM. Run
\code{taxa.ranks(biom)} to see the available options.
}
\item{\bold{Individual Taxon (TAXON): }}{
The name of a taxon from the BIOM's \code{taxonomy()}.
For instance, \code{Firmicutes} or \code{Prevotella}.
}
\item{\bold{Metadata (FACTOR or NUMERIC): }}{
The name of a column from the BIOM's \code{metadata()}.
For instance, \code{`Body Site`} or \code{Age}.
}
\item{\bold{Clustering Method (CLUST): }}{
\code{average}, \code{ward}, \code{mcquitty}, \code{single},
\code{median}, \code{complete}, or \code{centroid}. The following
aliases are also understood: \code{heatmap} = \code{complete},
\code{UPGMA} = \code{average}, \code{WPGMA} = \code{mcquitty},
\code{WPGMC} = \code{median}, and \code{UPGMC} = \code{centroid}.
}
\item{\bold{Special: }}{
\code{Rarefied}, \code{Reads}, \code{Samples}, \code{.}, \code{stacked}
}
}
}

\section{Term Combinations}{
\tabular{lll}{
   Combination \tab Chart Type \tab Example \cr
   ADIV ~ . \tab Box Plot \tab \code{plot(hmp50, Shannon ~ .)          } \cr
   ADIV ~ FACTOR \tab Box Plot \tab \code{plot(hmp50, Shannon ~ `Body Site`)} \cr
   BDIV ~ FACTOR \tab Box Plot \tab \code{plot(hmp50, Bray ~ Sex)           } \cr
   RANK ~ . \tab Box Plot \tab \code{plot(hmp50, Phylum ~ .)           } \cr
   TAXON ~ FACTOR \tab Box Plot \tab \code{plot(hmp50, Firmicutes ~ Sex)     } \cr
   ADIV ~ NUMERIC \tab Scatter Plot \tab \code{plot(hmp50, Shannon ~ Age)        } \cr
   ADIV ~ ADIV \tab Scatter Plot \tab \code{plot(hmp50, Depth ~ OTUs)         } \cr
   TAXON ~ NUMERIC \tab Scatter Plot \tab \code{plot(hmp50, Prevoltella ~ BMI)    } \cr
   BDIV ~ CLUST \tab Heatmap \tab \code{plot(hmp50, UniFrac ~ ward)       } \cr
   RANK ~ CLUST \tab Heatmap \tab \code{plot(hmp50, Genus ~ heatmap)      } \cr
   BDIV ~ ORD \tab Ordination \tab \code{plot(hmp50, Bray ~ NMDS)          } \cr
   RANK ~ stacked \tab Stacked bar \tab \code{plot(hmp50, Family ~ stacked)     } \cr
   Rarefied ~ Reads \tab Rarefaction \tab \code{plot(hmp50, Rarefied ~ Reads)     } \cr
   Rarefied ~ Samples \tab Rarefaction \tab \code{plot(hmp50, Rarefied ~ Samples)   } \cr
   Rarefied ~ ADIV \tab Rarefaction \tab \code{plot(hmp50, Rarefied ~ Shannon)   } \cr
}
}

\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    
    plot(biom, Shannon ~ `Body Site`)
    plot(biom, c(OTUs, Shannon) ~ Sex, layers="vb", color.by="Body Site")
    
    plot(biom, Simpson ~ `Body Site`, layers="p", color.by="Sex", xlab.angle=30)
    
    # Ordination
    plot(biom, bray ~ nmds, color.by="Body Site")
    
    # Dissimilarity boxplots
    plot(biom, UniFrac ~ `==Body Site`)
    
    # Dissimilarity Heatmap
    plot(biom, UniFrac ~ heatmap)
    
    # Taxa abundance boxplots
    plot(biom, c(Phylum, Genus) ~ .)
    
    # Taxa stacked abundance
    plot(biom, Phylum ~ stacked)
    
    # Taxa abundance heatmap
    plot(biom, Phylum ~ heatmap)
    

}
\seealso{
\code{\link{stats.table}}
}
