% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_multiplot.r
\name{rare_multiplot}
\alias{rare_multiplot}
\title{Combines rare_corrplot and rare_barplot into a single figure.}
\usage{
rare_multiplot(
  biom,
  adiv = "OTUs",
  depths = NULL,
  layers = "t",
  rline = TRUE,
  color.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  test = "pw_means",
  model = "log",
  level = 0.95,
  p.adj = "fdr",
  caption = TRUE,
  labels = FALSE,
  ...
)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{adiv}{Alpha diversity metric(s) to use. Options are: \code{"OTUs"},
\code{"Shannon"}, \code{"Chao1"}, \code{"Simpson"}, and/or
\code{"InvSimpson"}. Set \code{adiv=".all"} to use all metrics.
Default: \code{"Shannon"} \cr\cr
Multiple values allowed. Non-ambiguous abbreviations are allowed.}

\item{depths}{Rarefaction depths to show in the plot, or \code{NULL} to
auto-select. Default: \code{NULL}.}

\item{layers}{\code{"trend"}, \code{"scatter"}. Single letter
abbreviations are also accepted. For instance,
\code{c("trend", "scatter")} is equivalent to \code{c("t", "s")} and
\code{"ts"}. See \code{vignette("corrplots")} for examples of each.
Default: \code{"t"}.}

\item{rline}{Where to draw a horizontal line on the plot, intended to show
a particular rarefaction depth. Set to \code{TRUE} to show an
auto-selected rarefaction depth or \code{FALSE} to not show a line.
Default: \code{NULL}.}

\item{color.by, facet.by, limit.by}{Metadata columns to use for aesthetics
and partitioning. See below for details. Default: \code{NULL}}

\item{test}{The statistic of interest. An overview of options can be
found in the "Statistical Tests" section below. One of
\code{"predict"}, \code{"terms"}, \code{"fit"}, \code{"means"},
\code{"trends"}, \code{"es_means"}, \code{"es_trends"},
\code{"pw_means"}, or \code{"pw_trends"}.
Default: \code{"pw_means"}}

\item{model}{What type of trend model to fit to the data. Options are:
\code{"lm"} (linear), \code{"log"} (logarithmic), or \code{"gam"}
(generalized additive). See the "Model Options" section below for
additional details. Default: \code{"log"}}

\item{level}{The confidence level for calculating a confidence interval.
Default: \code{0.95}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}.}

\item{caption}{Add methodology caption beneath the plot.
Default: \code{TRUE}.}

\item{labels}{Show sample names under each bar. Default: \code{FALSE}.}

\item{...}{Additional parameters to pass along to ggplot2
functions. Prefix a parameter name with either \code{t.} or
\code{s.}/\code{pt.} to ensure it gets passed to (and only to)
\link[ggplot2]{geom_smooth} or \link[ggplot2]{geom_point},
respectively. For instance, \code{s.size = 2} ensures only the
scatterplot points have their size set to \code{2}.}
}
\value{
A \code{ggplot2} plot. \cr The computed data points, statistics,
ggplot command, and object history are available as \code{$data},
\code{$stats}, \code{$code}, and \code{$history}, respectively.
}
\description{
Combines rare_corrplot and rare_barplot into a single figure.
}
\section{Statistical Tests}{


\itemize{
\item{\code{"predict"} - }{
Augments original data with fitted information. See
\code{\link[broom:augment.lm]{broom::augment.lm()}}. }
\item{\code{"terms"} - }{ Summary information about the model's
terms, including p-value, r-squared, AIC, BIC, and confidence
interval at \code{level}. See \code{\link[broom:tidy.lm]{broom::tidy.lm()}}. }
\item{\code{"fit"} - }{ Goodness of fit measures, p-values, and
more for the overall model. See \code{\link[broom:glance.lm]{broom::glance.lm()}}. }
\item{\code{"means"} - }{ The estimated marginal mean (EMM) for
each \code{stat.by} group, along with confidence intervals (CI),
standard errors (SE), t-ratios, and p-values testing for mean = zero.
See \code{\link[emmeans:emmeans]{emmeans::emmeans()}} and \code{\link[emmeans:summary.emmGrid]{emmeans::summary.emmGrid()}}. }
\item{\code{"trends"} - }{ The trendline slope for each
\code{stat.by} group, along with CI and SE and p-value testing for
slope = zero. See \code{\link[emmeans:emtrends]{emmeans::emtrends()}} and
\code{\link[emmeans:summary.emmGrid]{emmeans::summary.emmGrid()}}. }
\item{\code{"pw_means"} - }{ Pairwise means. All \code{stat.by} groups are
compared to each other and the difference in means is estimated
along with SE, t-ratios, and p-values testing if the two means
are the same. See \code{\link[emmeans:contrast]{emmeans::pairs.emmGrid()}}. }
\item{\code{"pw_trends"} - }{ Pairwise trends. As above, but
comparing trendline slopes instead of means. }
\item{\code{"es_means"} - }{ Effect sizes for pairwise differences
of means, including SE and CI. See \code{\link[emmeans:eff_size]{emmeans::eff_size()}}. }
\item{\code{"es_trends"} - }{ Effect sizes for pairwise differences
of slopes, including SE and CI. See \code{\link[emmeans:eff_size]{emmeans::eff_size()}}. }
}
}

\section{Model Options}{


The predefined options are:
\itemize{
\item{\code{"lm"} - }{  Linear model: \code{stats::lm(formula = y ~ x)}.) }
\item{\code{"log"} - }{ Logarithmic model: \code{stats::lm(formula = y ~ log(x))}. }
\item{\code{"gam"} - }{ Generalized additive model: \code{mgcv::gam(formula = y ~ s(x, bs = "cs"), method = "REML")}. }
}

You can alternatively provide a list of length two where the first
element is a character vector of length 1 naming a function, and the
second element is a list of arguments to pass to that function. One
of the function's arguments must be named 'formula'.
For example, \code{model = list("stats::lm", list(formula = y ~ x))}.
}

\section{Aesthetics and Partitions}{


Metadata can be used to flexibly subset, partition, and apply aesthetics
when creating a plot. Common use cases are provided below. More thorough
documentation is available at \url{https://cmmr.github.io/rbiom}.

\preformatted{  ## Colors ----------------------------
  color.by = "Body Site"
  color.by = list('Body Site' = "bright")
  color.by = list('Body Site' = c("Stool", "Saliva"))
  color.by = list('Body Site' = list('values' = c("Stool", "Saliva"), 'colors' = "bright"))
  color.by = list('Body Site' = c('Stool' = "blue", 'Saliva' = "green"))
  
  ## Facets ----------------------------
  facet.by = "Body Site"
  facet.by = c("Body Site", "Sex")
  facet.by = list('Body Site' = c("Stool", "Saliva"), "Sex")
  
  ## Limits ----------------------------
  limit.by = list('Sex' = "Male", 'Age' = c(20,40))
  limit.by = list('Body Site' = c("Saliva", "Anterior nares"), 'Age' = c(NA,35))
}

\itemize{
\item{\code{color.by} - }{Any metadata column. (Max 1)}
\item{\code{facet.by} - }{Only categorical metadata column(s).}
\item{\code{limit.by} - }{Any metadata column(s).}
}

All built-in color palettes are colorblind-friendly.

The available categorical palette names are: \code{"okabe"}, \code{"carto"},
\code{"r4"}, \code{"polychrome"}, \code{"tol"}, \code{"bright"},
\code{"light"}, \code{"muted"}, \code{"vibrant"}, \code{"tableau"},
\code{"classic"}, \code{"alphabet"}, \code{"tableau20"}, \code{"kelly"},
and \code{"fishy"}.

The available numeric palette names are: \code{"reds"}, \code{"oranges"},
\code{"greens"}, \code{"purples"}, \code{"grays"}, \code{"acton"},
\code{"bamako"}, \code{"batlow"}, \code{"bilbao"}, \code{"buda"},
\code{"davos"}, \code{"devon"}, \code{"grayC"}, \code{"hawaii"},
\code{"imola"}, \code{"lajolla"}, \code{"lapaz"}, \code{"nuuk"},
\code{"oslo"}, \code{"tokyo"}, \code{"turku"}, \code{"bam"},
\code{"berlin"}, \code{"broc"}, \code{"cork"}, \code{"lisbon"},
\code{"roma"}, \code{"tofino"}, \code{"vanimo"}, \code{"vik"}
}

\examples{
    library(rbiom)
    
    rare_multiplot(hmp50, color.by="Body Site")
    
}
\seealso{
Other rarefaction: 
\code{\link{is_rarefied}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_depth}()},
\code{\link{rare_suggest}()},
\code{\link{sample_rarefy}()},
\code{\link{sample_sums}()}

Other visualization: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{taxa_barplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()}
}
\concept{rarefaction}
\concept{visualization}
