% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rarefy.r
\name{rare_suggest}
\alias{rare_suggest}
\title{Suggest a 'good' rarefaction depth.}
\usage{
rare_suggest(biom)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}
}
\value{
An integer.
}
\description{
Suggest a 'good' rarefaction depth.
}
\examples{
    library(rbiom)

    rare_suggest(hmp50)

}
\seealso{
Other rarefaction: 
\code{\link{is_rarefied}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_depth}()},
\code{\link{rare_multiplot}()},
\code{\link{sample_rarefy}()},
\code{\link{sample_sums}()}
}
\concept{rarefaction}
