\docType{package}
\name{rbiom-package}
\alias{rbiom-package}
\title{This is an R package for interfacing with the biom format.}
\description{
  This is an R package for interfacing with the BIOM
  format. We plan to include tools for both reading and
  writing BIOM format using R, as well as a formal
  object-oriented representation of BIOM objects that
  parallels the representation in other languages that
  support BIOM, like python.
}
\author{
  Paul J. McMurdie II \email{mcmurdie@stanford.edu}
}
\references{
  \url{http://www.biom-format.org/}
}
\keyword{package}

