% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.r
\name{repair}
\alias{repair}
\title{Run after manually editing a BIOM object's content.}
\usage{
repair(biom, prune = TRUE, fast = FALSE)
}
\arguments{
\item{biom}{The \code{BIOM} object to repair.}

\item{prune}{Remove taxa and samples with zero observations. (Default: TRUE)}

\item{fast}{Should subsetting the phylogenetic tree and sequences be 
skipped? These slow steps are often not necessary. (Default: FALSE)}
}
\value{
A \code{BIOM} object.
}
\description{
Run after manually editing a BIOM object's content.
}
