% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{sample.sums}
\alias{sample.sums}
\title{Sum the observations in each sample.}
\usage{
sample.sums(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
A named numeric vector of the number of observations in each 
        sample. The names are the sample IDs.
}
\description{
Sum the observations in each sample.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom <- read.biom(infile)
    
    sample.sums(biom)
    
    sample.sums(biom) \%>\% sort() \%>\% head()
    
    hist(sample.sums(biom))

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{info}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
