% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.r
\name{sample_metadata}
\alias{sample_metadata}
\alias{sample_metadata<-}
\title{Get or set the sample metadata.}
\usage{
sample_metadata(biom, field = NULL)

sample_metadata(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{field}{The name of a single metadata column to retrieve. When
provided, a named vector will be returned instead of a data.frame.
Default: \code{NULL}}

\item{value}{A data.frame with the metadata. Its row names or '.sample'
column must match the IDs in \code{sample_names(x)}. The returned
rbiom object will be subset to just the samples in the new metadata.
With the exception of '.sample', all character columns will be
converted to factor.}
}
\value{
A tibble (data.frame) of the metadata in \code{biom}. '.sample' will
always be the first column. If \code{field} is given, will instead
return a named vector of the field's values.
}
\description{
Get or set the sample metadata.
}
\examples{
    library(rbiom)
    
    sample_metadata(hmp50) \%>\% head(4)
    
    sample_metadata(hmp50, "Body Site") \%>\% head(4)
    
    md <- sample_metadata(hmp50)
    md <- md[,c('.sample', 'Sex', 'Body Site')]
    sample_metadata(hmp50) <- md
    sample_metadata(hmp50) \%>\% head(4)

}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{metadata_names}()},
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_names}()},
\code{\link{taxa_ranks}()}
}
\concept{setters}
