% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{sample_names}
\alias{sample_names}
\alias{sample_names<-}
\title{Get or set the sample names.}
\usage{
sample_names(biom)

sample_names(biom) <- value
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{value}{A character vector of the new sample names.}
}
\value{
A character vector of the sample names in \code{biom}.
}
\description{
Get or set the sample names.
}
\examples{
    library(rbiom) 
    
    biom <- hmp50
    
    head(sample_names(biom))
    
    sample_names(biom) <- sub('HMP', 'Sample_', sample_names(biom))
    head(sample_names(biom))
}
\seealso{
Other samples: 
\code{\link{n_samples}()},
\code{\link{sample_metadata}()},
\code{\link{sample_select}()},
\code{\link{sample_sums}()},
\code{\link{subset.rbiom}()},
\code{\link{within.rbiom}()}
}
\concept{samples}
