% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{sample_names}
\alias{sample_names}
\title{Get the sample names.}
\usage{
sample_names(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
A character vector of the sample IDs / names in \code{biom}.
}
\description{
Get the sample names.
}
\examples{
    library(rbiom)
    
    sample_names(hmp50)

}
\seealso{
Other accessors: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_sums}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
