% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{sample_sums}
\alias{sample_sums}
\title{Sum the observations in each sample.}
\usage{
sample_sums(biom, long = FALSE, md = FALSE)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read_biom}.}
}
\value{
A named numeric vector of the number of observations in each 
        sample. The names are the sample IDs.
}
\description{
Sum the observations in each sample.
}
\examples{
    library(rbiom)
    
    sample_sums(hmp50) \%>\% sort() \%>\% head()
    
    hist(sample_sums(hmp50))

}
\seealso{
Other accessors: 
\code{\link{comments}()},
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has_phylogeny}()},
\code{\link{has_sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is_rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample_names}()},
\code{\link{sequences}()},
\code{\link{taxa_names}()},
\code{\link{taxa_ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessors}
