% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{sequences}
\alias{sequences}
\title{Nucleotide sequences associated with each taxonomic identifier.}
\usage{
sequences(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
A named character vector of sequences in \code{biom}. If this data
          is not present, then returns \code{NULL}.
}
\description{
Nucleotide sequences associated with each taxonomic identifier.
}
\examples{
    library(rbiom)
    
    sequences(hmp50)[1:4]
    
    # Write to a compressed fasta file in the temporary directory:
    seqs <- sequences(hmp50)
    conn <- bzfile(file.path(tempdir(), "Sequences.fa.bz2"), "w")
    cat(sprintf(">\%s\n\%s", names(seqs), seqs), file=conn, sep="\n")
    close(conn)
    
    # You can also use the write.fasta function for this task:
    write.fasta(hmp50, file.path(tempdir(), "Sequences.fa.gz"))

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is.rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()},
\code{\link{taxonomy}()}
}
\concept{accessor functions}
