% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_table.r
\name{stats_table}
\alias{stats_table}
\title{Run non-parametric statistics on a data.frame.}
\usage{
stats_table(
  df,
  stat.by,
  resp = attr(df, "response"),
  regr = NULL,
  test = ifelse(is.null(regr), "means", "trends"),
  model = "lm",
  level = 0.95,
  split.by = NULL,
  p.adj = "fdr"
)
}
\arguments{
\item{df}{A data.frame with columns named by \code{stat.by}, \code{resp},
\code{regr}, and \code{split.by}. Required.}

\item{stat.by}{The categorical metadata field defining the statistical
groups. Required.}

\item{resp}{The response (independent) numeric variable, such as taxa
abundance or alpha diversity.  Default: \code{attr(df, 'response')}}

\item{regr}{To run a regression analysis, set \code{regr} to the numeric
metadata field with the "x-axis" values. Leaving \code{regr=NULL}
will generate boxplot-like statistics; when non-NULL, corrplot-like
statistics will be returned. Default: \code{NULL}}

\item{test}{The statistic of interest. An overview of options can be
found in the "Statistical Tests" section below. One of
\code{"predict"}, \code{"terms"}, \code{"fit"}, \code{"means"},
\code{"trends"}, \code{"es_means"}, \code{"es_trends"},
\code{"pw_means"}, or \code{"pw_trends"}.
Default: \code{ifelse(is.null(regr), "means", "trends")}}

\item{model}{What type of trend model to fit to the data. Options are:
\code{"lm"} (linear), \code{"log"} (logarithmic), or \code{"gam"}
(generalized additive). See the "Model Options" section below for
additional details. Default: \code{"lm"}}

\item{level}{The confidence level for calculating a confidence interval.
Default: \code{0.95}}

\item{split.by}{Name(s) of metadata columns that the data should be split
by prior to any calculations. Default: \code{NULL}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}.}
}
\value{
A tibble data frame with summary statistics. \cr
The R code or generating these statistics is in \code{$code}, and
the object history is in \code{$history}.
}
\description{
Run non-parametric statistics on a data.frame.
}
\section{Statistical Tests}{


\bold{When \code{regr} is NULL:}
\itemize{
\item{\code{"means"} - }{
Considers all groups at once using the Kruskal-Wallis
non-parametric test: \code{\link[stats:kruskal.test]{stats::kruskal.test()}}. }
\item{\code{"pw_means"} - }{
Pairwise comparison of \code{stat.by} groups using the Wilcox rank sum
(aka Mann-Whitney) non-parametric test: \code{\link[stats:wilcox.test]{stats::wilcox.test()}}. }
}

\bold{When \code{regr} is not NULL:}
\itemize{
\item{\code{"predict"} - }{
Augments original data with fitted information. See
\code{\link[broom:augment.lm]{broom::augment.lm()}}. }
\item{\code{"terms"} - }{ Summary information about the model's
terms, including p-value, r-squared, AIC, BIC, and confidence
interval at \code{level}. See \code{\link[broom:tidy.lm]{broom::tidy.lm()}}. }
\item{\code{"fit"} - }{ Goodness of fit measures, p-values, and
more for the overall model. See \code{\link[broom:glance.lm]{broom::glance.lm()}}. }
\item{\code{"means"} - }{ The estimated marginal mean (EMM) for
each \code{stat.by} group, along with confidence intervals (CI),
standard errors (SE), t-ratios, and p-values testing for mean = zero.
See \code{\link[emmeans:emmeans]{emmeans::emmeans()}} and \code{\link[emmeans:summary.emmGrid]{emmeans::summary.emmGrid()}}. }
\item{\code{"trends"} - }{ The trendline slope for each
\code{stat.by} group, along with CI and SE and p-value testing for
slope = zero. See \code{\link[emmeans:emtrends]{emmeans::emtrends()}} and
\code{\link[emmeans:summary.emmGrid]{emmeans::summary.emmGrid()}}. }
\item{\code{"pw_means"} - }{ Pairwise means. All \code{stat.by} groups are
compared to each other and the difference in means is estimated
along with SE, t-ratios, and p-values testing if the two means
are the same. See \code{\link[emmeans:contrast]{emmeans::pairs.emmGrid()}}. }
\item{\code{"pw_trends"} - }{ Pairwise trends. As above, but
comparing trendline slopes instead of means. }
\item{\code{"es_means"} - }{ Effect sizes for pairwise differences
of means, including SE and CI. See \code{\link[emmeans:eff_size]{emmeans::eff_size()}}. }
\item{\code{"es_trends"} - }{ Effect sizes for pairwise differences
of slopes, including SE and CI. See \code{\link[emmeans:eff_size]{emmeans::eff_size()}}. }
}
}

\section{Model Options}{


The predefined options are:
\itemize{
\item{\code{"lm"} - }{  Linear model: \code{stats::lm(formula = y ~ x)}.) }
\item{\code{"log"} - }{ Logarithmic model: \code{stats::lm(formula = y ~ log(x))}. }
\item{\code{"gam"} - }{ Generalized additive model: \code{mgcv::gam(formula = y ~ s(x, bs = "cs"), method = "REML")}. }
}

You can alternatively provide a list of length two where the first
element is a character vector of length 1 naming a function, and the
second element is a list of arguments to pass to that function. One
of the function's arguments must be named 'formula'.
For example, \code{model = list("stats::lm", list(formula = y ~ x))}.
}

\examples{
    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    
    df <- taxa_table(biom, rank = "Family")
    stats_table(df, stat.by = "Body Site")
    
    df <- adiv_table(biom)
    stats_table(df, stat.by = "Sex", test = "pw_means")
    
    stats_table(df, stat.by = "Sex", regr = "BMI")

}
\seealso{
Other stats_tables: 
\code{\link{adiv_stats}()},
\code{\link{bdiv_stats}()},
\code{\link{distmat_stats}()},
\code{\link{taxa_stats}()}
}
\concept{stats_tables}
