% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.r
\name{subset.rbiom}
\alias{subset.rbiom}
\title{Subset by sample metadata.}
\usage{
\method{subset}{rbiom}(biom, subset, select)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{subset}{Logical expression indicating elements or rows to keep.}

\item{select}{An expression, indicating columns to select from a data frame.}
}
\value{
An \code{rbiom} object with only the specified samples.
}
\description{
Subset by sample metadata.
}
\examples{
    library(rbiom)
    
    biom <- hmp50
    sample_metadata(biom) \%>\% head()
    
    biom \%<>\% subset(Sex == "Male")
    sample_metadata(biom) \%>\% head()

}
\seealso{
Other metadata: 
\code{\link{bdply}()},
\code{\link{blply}()},
\code{\link{metadata_names}()},
\code{\link{metadata_numeric}()},
\code{\link{n_metadata}()},
\code{\link{sample_metadata}()},
\code{\link{sample_select}()},
\code{\link{within.rbiom}()},
\code{\link{write_metadata}()}

Other samples: 
\code{\link{n_samples}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()},
\code{\link{sample_select}()},
\code{\link{sample_sums}()},
\code{\link{within.rbiom}()}
}
\concept{metadata}
\concept{samples}
