% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.r
\name{subset}
\alias{subset}
\alias{subset.rbiom}
\alias{[.rbiom}
\title{Subset an rbiom object by sample names or metadata.}
\usage{
\method{subset}{rbiom}(biom, ..., clone = TRUE)

\method{[}{rbiom}(biom, i)
}
\arguments{
\item{biom}{An \link[=rbiom_objects]{rbiom object}, such as from \code{\link[=as_rbiom]{as_rbiom()}}.}

\item{...}{Passed on to \code{\link[base:subset]{base::subset()}}.}

\item{clone}{Create a copy of \code{biom} before modifying. If \code{FALSE}, \code{biom}
is modified in place as a side-effect. See \link[=speed]{speed ups} for
use cases. Default: \code{TRUE}}
}
\value{
An \link[=rbiom_objects]{rbiom object}.
}
\description{
Subset an rbiom object by sample names or metadata.
}
\examples{
    library(rbiom) 
    
    # Subset to specific samples
    biom <- hmp50[c('HMP20', 'HMP42', 'HMP12')]
    biom$metadata
    
    # Subset according to metadata (using base::subset)
    biom <- subset(hmp50, `Body Site` \%in\% c('Saliva') & Age < 25)
    biom$metadata

}
\seealso{
Other transformations: 
\code{\link{mutate}()},
\code{\link{rarefy}()},
\code{\link{rarefy_cols}()},
\code{\link{slice}()},
\code{\link{with}()}
}
\concept{transformations}
