% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.r
\name{subset}
\alias{subset}
\alias{subset.BIOM}
\title{Subset samples using the BIOM object's metadata}
\usage{
\method{subset}{BIOM}(x, expr, env = parent.frame(), drop.na = TRUE, refactor = TRUE)
}
\arguments{
\item{x}{A BIOM object, as returned from \link{read.biom}.}

\item{expr}{Logical expression to run on the metadata to identify samples
to retain.}

\item{env}{The environment to search for variables used in \code{expr}.
(Default: \code{parent.frame()})}

\item{drop.na}{When \code{expr} is e.g. \code{Age > 30}, should 
\code{!is.na(Age)} be automatically applied too?}

\item{refactor}{When \code{expr} is e.g. 
\code{`Body Site` \%in\% c("Stool", "Saliva")}, should
\code{`Body Site`} be redefined as 
\code{factor(`Body Site`, levels=c('Stool', 'Saliva'))}?}
}
\value{
A \code{BIOM} object.
}
\description{
Subset samples using the BIOM object's metadata
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom <- read.biom(infile)
    
    ex1 <- subset(biom, Age > 30)
    ex2 <- subset(biom, `Body Site` \%in\% c("Saliva", "Stool"))
    ex3 <- subset(biom, a == b, list(a = as.name("Body Site"), b ="Saliva"))
    ex4 <- subset(biom, Age < 25 & BMI > 22)

}
\seealso{
\code{\link{select}}
}
