% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_boxplot.r
\name{taxa_boxplot}
\alias{taxa_boxplot}
\title{Visualize BIOM data with boxplots.}
\usage{
taxa_boxplot(
  biom,
  x = NULL,
  rank = -1,
  taxa = 6,
  layers = "x",
  unc = "singly",
  other = FALSE,
  color.by = NULL,
  pattern.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  flip = FALSE,
  stripe = NULL,
  p.top = Inf,
  p.adj = "fdr",
  p.label = 0.05,
  ci = "ci",
  level = 0.95,
  outliers = NULL,
  xlab.angle = "auto",
  trans = "percent",
  y.trans = "sqrt",
  caption = TRUE,
  ...
)
}
\arguments{
\item{biom}{An rbiom-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{x}{A categorical metadata column name to use for the x-axis. The
default, \code{NULL} puts taxa along the x-axis.}

\item{rank}{What rank(s) of taxa to display. E.g. \code{"Phylum"},
\code{"Genus"}, \code{".otu"}, etc. An integer vector can also be
given, where \code{1} is the highest rank, \code{2} is the second
highest, \code{-1} is the lowest rank, \code{-2} is the second
lowest, and \code{0} is the OTU "rank". Run \code{biom$ranks} to
see all options for a given rbiom object. Default: \code{-1}.}

\item{taxa}{Which taxa to display. An integer value will show the top n
most abundant taxa. A value 0 <= n < 1 will show any taxa with that
mean abundance or greater (e.g. \code{0.1} implies >= 10\%). A
character vector of taxa names will show only those named taxa.
Default: \code{6}.}

\item{layers}{One or more of
\code{c("bar", "box" ("x"), "violin", "dot", "strip", "crossbar", "errorbar", "linerange", "pointrange")}.
Single letter abbreviations are also accepted. For instance,
\code{c("box", "dot")} is equivalent to \code{c("x", "d")} and \code{"xd"}.
See \link[=plots]{plot types} for examples of each. Default: \code{"x"}}

\item{unc}{How to handle unclassified, uncultured, and similarly ambiguous
taxa names. Options are:
\itemize{
\item{\code{"singly"} - }{ Replaces them with the OTU name. }
\item{\code{"grouped"} - }{ Replaces them with a higher rank's name. }
\item{\code{"drop"} - }{ Excludes them from the result. }
\item{\code{"asis"} - }{ To not check/modify any taxa names. }
}
Default: \code{"singly"} \cr\cr
Abbreviations are allowed.}

\item{other}{Sum all non-itemized taxa into an "Other" taxa. When
\code{FALSE}, only returns taxa matched by the \code{taxa}
argument. Specifying \code{TRUE} adds "Other" to the returned set.
A string can also be given to imply \code{TRUE}, but with that
value as the name to use instead of "Other".
Default: \code{FALSE}}

\item{color.by, pattern.by, shape.by, facet.by, limit.by}{Metadata columns to
use for aesthetics and partitioning. Default: \code{NULL}}

\item{flip}{Transpose the axes, so that taxa are present as rows instead
of columns. Default: \code{FALSE}}

\item{stripe}{Shade every other x position. Default: \emph{same as flip}}

\item{p.top}{Only display taxa with the most significant differences in
abundance. If \code{p.top} is >= 1, then the \code{p.top} most
significant taxa are displayed. If \code{p.top} is less than one, all
taxa with an adjusted p-value <= \code{p.top} are displayed.
Recommended to be used in combination with the \code{taxa} parameter
to set a lower bound on the mean abundance of considered taxa.
Default: \code{Inf}}

\item{p.adj}{Method to use for multiple comparisons adjustment of
p-values. Run \code{p.adjust.methods} for a list of available
options. Default: \code{"fdr"}}

\item{p.label}{Minimum adjusted p-value to display on the plot with a
bracket.
\itemize{
\item{\code{p.label = 0.05} - }{ Show p-values that are <= 0.05. }
\item{\code{p.label = 0} - }{ Don't show any p-values on the plot. }
\item{\code{p.label = 1} - }{ Show all p-values on the plot. }
}
If a numeric vector with more than one value is
provided, they will be used as breaks for asterisk notation.
Default: \code{0.05}}

\item{ci}{How to calculate min/max of the \bold{crossbar},
\bold{errorbar}, \bold{linerange}, and \bold{pointrange} layers.
Options are: \code{"ci"} (confidence interval), \code{"range"},
\code{"sd"} (standard deviation), \code{"se"} (standard error), and
\code{"mad"} (median absolute deviation).
The center mark of \bold{crossbar} and \bold{pointrange} represents
the mean, except for code{"mad"} in which case it represents
the median. Default: \code{"ci"}}

\item{level}{The confidence level for calculating a confidence interval.
Default: \code{0.95}}

\item{outliers}{Show boxplot outliers? \code{TRUE} to always show.
\code{FALSE} to always hide. \code{NULL} to only hide them when
overlaying a dot or strip chart.  Default: \code{NULL}}

\item{xlab.angle}{Angle of the labels at the bottom of the plot.
Options are \code{"auto"}, \code{'0'}, \code{'30'}, and \code{'90'}.
Default: \code{"auto"}.}

\item{trans}{Transformation to apply. Options are:
\code{c("none", "rank", "log", "log1p", "sqrt")}. \code{"rank"} is useful for
correcting for non-normally distributions before applying regression
statistics. Default: \code{"none"}}

\item{y.trans}{The transformation to apply to the y-axis. Visualizing
differences of both high- and low-abundance taxa is best done with
a non-linear axis. Options are:
\itemize{
\item{\code{"sqrt"} - }{ square-root transformation }
\item{\code{"log1p"} - }{ log(y + 1) transformation }
\item{\code{NULL} - }{ no transformation }
}
These methods allow visualization of both high- and low-abundance
taxa simultaneously, without complaint about 'zero' count
observations. Default: \code{"sqrt"}}

\item{caption}{Add methodology caption beneath the plot.
Default: \code{TRUE}}

\item{...}{Parameters are matched to formal arguments of ggplot2
functions. Prefixing parameter names with a layer name ensures that
a particular parameter is passed to, and only to, that layer. For
instance, \code{dot.size = 2} or \code{d.size = 2} ensures only the
dotplot layer has its size set to \code{2}. The special prefix
\code{pt.} will control both the dot and strip layers.}
}
\value{
A \code{ggplot2} plot. \cr The computed data points, statistics,
and ggplot command are available as \verb{$data}, \verb{$stats}, and
\verb{$code}, respectively.
}
\description{
Visualize BIOM data with boxplots.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    
    taxa_boxplot(biom, color.by = "Body Site", stripe = TRUE)
    taxa_boxplot(biom, layers = "bed", rank = c("Phylum", "Genus"), flip = TRUE)
    taxa_boxplot(
      biom = biom, taxa = 3, layers = "ps", 
      color.by = list("Body Site" = c('Saliva' = "blue", 'Stool' = "red")) )
    
}
\seealso{
Other taxa_abundance: 
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_matrix}()},
\code{\link{taxa_stacked}()},
\code{\link{taxa_stats}()},
\code{\link{taxa_sums}()}

Other visualization: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_corrplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{rare_stacked}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()}
}
\concept{taxa_abundance}
\concept{visualization}
