% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_boxplot.r
\name{taxa_boxplot}
\alias{taxa_boxplot}
\title{Visualize BIOM data with boxplots.}
\usage{
taxa_boxplot(
  biom,
  x = ".taxa",
  rank = NULL,
  taxa = 5,
  layers = "rls",
  color.by = NULL,
  pattern.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  xvals = NULL,
  colors = NULL,
  patterns = NULL,
  shapes = NULL,
  facets = NULL,
  p.top = Inf,
  p.adj = "fdr",
  p.label = 0.05,
  ci = 95,
  xlab.angle = "auto",
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{x}{A categorical metadata column name to use for the x-axis. The 
default, \code{".taxa"} puts the taxa along the x-axis.}

\item{rank}{What rank of taxa to display. E.g. "Phylum", "Genus", etc. Run
\code{taxa_ranks()} to see all options for a given BIOM object. The
default, \code{NULL}, selects the lowest level.}

\item{taxa}{Which taxa to display. An integer value will show the top n
most abundant taxa. A value 0 <= n < 1 will show any taxa with that 
mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. Default: \code{5}.}

\item{layers}{\code{"box"}, \code{"bar" ("r")}, \code{"violin"}, 
\code{"dot"}, \code{"strip"}, \code{"crossbar"}, \code{"errorbar"}, 
\code{"linerange"}, and \code{"pointrange"}. Single letter 
abbreviations are also accepted. For instance, \code{c("box", "dot")} 
is equivalent to \code{c("b", "d")} and \code{"bd"}.
Default: \code{"rls"}.}

\item{color.by, pattern.by, shape.by, facet.by}{Metadata 
column to color, pattern, shape, and/or facet by. If 
that column is a \code{factor}, the ordering of levels will be 
maintained in the plot.}

\item{colors, patterns, shapes, facets, xvals}{Names of the colors, patterns,
shapes, facets, and/or x values to use in the plot. Available values 
for colors, patterns, and shapes are given by \code{colors()}, 
\code{gridpattern::names_magick}, and \code{0:25},
respectively. Use a named character vector to map them to specific
factor levels in the metadata. \code{facets} and \code{xvals} are
coerced to unnamed character vectors. If the length of these vectors
is less than the values present in their corresponding metadata 
column, then the data set will be subseted accordingly.}

\item{p.top}{Only display taxa with the most significant differences in 
abundance. If \code{p.top} is >= 1, then the \code{p.top} most 
significant taxa are displayed. If \code{p.top} is less than one, all 
taxa with an adjusted p-value <= \code{p.top} are displayed. 
Recommended to be used in combination with the \code{taxa} parameter 
to set a lower bound on the mean abundance of considered taxa. 
Default: \code{Inf}}

\item{p.adj}{Method to use for multiple comparisons adjustment of p-values.
Run \code{p.adjust.methods} for a list of available options.
Default: \code{fdr}}

\item{p.label}{Minimum adjusted p-value to display on the plot with a 
bracket. Set to \code{Inf} to display all p-values, or \code{-Inf} 
for no brackets. If a numeric vector with more than one value is 
provided, they will be used as breaks for asterisk notation. 
Default: \code{0.05}}

\item{ci}{How to calculate min/max of the \bold{crossbar}, 
\bold{errorbar}, \bold{linerange}, and \bold{pointrange} layers.
Provide a number between 75 and 100 to define a confidence interval's
confidence level, commonly 95 or 97.5. Other options are: 
\bold{range}, 
\bold{sd} (standard deviation), 
\bold{se} (standard error), and 
\bold{mad} (median absolute deviation). 
The center mark of \code{crossbar} and \code{pointrange} represents
the mean, except for \bold{mad} in which case it represents
the median. Trendlines require a confidence interval value. 
Set to \code{NULL} to disable. Default: \code{95}}

\item{xlab.angle}{How to rotate the tick labels on the x-axis. 
\bold{'auto'} (the default), automatically selects a rotation value. 
\bold{0}, \bold{30}, and \bold{90} sets the angle to horizontal, 
angled, and vertical, respectively.}

\item{...}{Parameters are matched to formal arguments of ggplot2
functions. Prefixing parameter names with a layer name ensures that
a particular parameter is passed to, and only to, that layer. For
instance, \code{dot.size = 2} or \code{d.size = 2} ensures only the 
dotplot layer has its size set to \code{2}.}

\item{other}{Should non-selected taxa be displayed as an "Other" group?
Default: \code{FALSE}}

\item{safe}{If \code{FALSE}, data.frame column names such as 
\code{".metric"} will be auto-converted to \code{"Metric"} to improve
human-readability. Conversion if aborted if a conflict is found with
a metadata column name. Default: \code{FALSE}}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
        
Shapes can also be given as their string values, defined in pch_table here:
https://github.com/tidyverse/ggplot2/blob/master/R/geom-point.r . Note that
some shapes have a colored outline given by `color`, some are filled with 
`color`, and some are outlined in `color` and filled with `fill`. See
https://blog.albertkuo.me/post/point-shape-options-in-ggplot/ for details.

To expand the low end of the y axis, you can set \code{y.trans = "sqrt"} or
\code{y.trans = "log1p"}. The former applies a square-root transformation, 
and the latter plots log(y + 1). Both of these methods work well with data
that contains zeroes.
}
\description{
Visualize BIOM data with boxplots.
}
\examples{
    library(rbiom)
    
    biom <- rarefy(hmp50)
    taxa_boxplot(biom, rank = c("Phylum", "Genus"))
    

}
\seealso{
\code{\link{stats_table}}

Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()}
}
\concept{plotting}
