% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_corrplot.r
\name{taxa_corrplot}
\alias{taxa_corrplot}
\title{Visualize taxa abundance with scatterplots and trendlines.}
\usage{
taxa_corrplot(
  biom,
  x,
  rank = NULL,
  taxa = 6,
  points = FALSE,
  color.by = NULL,
  facet.by = NULL,
  colors = NULL,
  facets = NULL,
  model = "linear",
  ci = 95,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{x}{A numeric metadata column name to use for the x-axis. Required.}

\item{rank}{What rank of taxa to display. E.g. "Phylum", "Genus", etc. Run
\code{taxa_ranks()} to see all options for a given BIOM object. The
default, \code{NULL}, selects the lowest level.}

\item{taxa}{Which taxa to display. An integer value will show the top n
most abundant taxa. A value 0 <= n < 1 will show any taxa with that 
mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. Default: \code{6}.}

\item{points}{Overlay a scatter plot. Default: \code{FALSE}.}

\item{color.by, facet.by}{Metadata column to color and/or facet by. If that
column is a \code{factor}, the ordering of levels will be maintained 
in the plot. Default: \code{color.by = NULL, facet.by = NULL}.}

\item{colors, facets}{Names of the colors and/or facets values to use in
the plot. Available values for colors are given by \code{colors()}. 
Use a named character vector to map them to specific factor levels in
the metadata. \code{facets} are coerced to unnamed character vectors.
If the length of these vectors is less than the values present in 
their corresponding metadata column, then the data set will be 
subseted accordingly. Default: \code{colors = NULL, facets = NULL}.}

\item{model}{What type of trendline to fit to the data. Options are: 
\code{"linear"}, \code{"logarithmic"}, or \code{"local"}. You can
alternatively provide \bold{method} and/or \bold{formula} arguments
which will override these preset options for 
\link[ggplot2]{stat_smooth}. Default: \code{"linear"}.}

\item{ci}{The confidence interval to display around the fitted curve. Set
to \code{FALSE} to hide the confidence interval. Default: \code{95}.}

\item{...}{Additional parameters to pass along to ggplot2
functions. Prefix a parameter name with either \code{p.} or \code{s.}
to ensure it gets passed to (and only to) 
\link[ggplot2]{geom_point} or 
\link[ggplot2]{geom_smooth}, respectively. For instance, 
\code{p.size = 2} ensures only the points have their size set to 
\code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
}
\description{
Visualize taxa abundance with scatterplots and trendlines.
}
\examples{
    library(rbiom)
    
    taxa_corrplot(rarefy(hmp50), "BMI", color.by="Body Site") 
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()},
\code{\link{tree_plot}()}
}
\concept{plotting}
