% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_corrplot.r
\name{taxa_corrplot}
\alias{taxa_corrplot}
\title{Visualize taxa abundance with scatterplots and trendlines.}
\usage{
taxa_corrplot(
  biom,
  x,
  rank = NULL,
  taxa = 6,
  layers = "t",
  color.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  model = "linear",
  stats = "emtrends",
  p.top = Inf,
  p.adj = "fdr",
  ci = 95,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{x}{A numeric metadata column name to use for the x-axis. Required.}

\item{rank}{What rank of taxa to display. E.g. "Phylum", "Genus", etc. 
Run \code{taxa_ranks()} to see all options for a given BIOM object. 
The default, \code{NULL}, selects the lowest level.}

\item{taxa}{Which taxa to display. An integer value will show the top n
most abundant taxa. A value 0 <= n < 1 will show any taxa with that 
mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. Default: \code{6}.}

\item{layers}{\code{"trend"}, \code{"scatter"}. Single letter 
abbreviations are also accepted. For instance, 
\code{c("trend", "scatter")} is equivalent to \code{c("t", "s")} and 
\code{"ts"}. See \code{vignette("corrplots")} for examples of each. 
Default: \code{"t"}.}

\item{color.by, facet.by, limit.by}{Metadata columns to use for aesthetics 
and partitioning. See below for details. Default: \code{NULL}}

\item{model}{What type of trendline to fit to the data. Options are: 
\code{"linear"}, \code{"logarithmic"}, or \code{"local"}. You can
alternatively provide a list of length two where the first element is
a character vector of length 1 naming a function, and the second 
element is a list of arguments to pass to that function. One of the 
function's arguments must be named 'formula'. For example, 
\code{model = list("stats::lm", list(formula = y ~ x))}.
Default: \code{"linear"}.}

\item{stats}{Which statistic to display on the plot. Options are: 
\describe{
  \item{\code{"fit"}}{ How well does the model fit all the data? }
  \item{\code{"emtrends"}}{ Does any trendline have a non-zero slope? }
  \item{\code{"emt_pairs"}}{ Are the slopes of any trendlines different? }
  \item{\code{"emm_pairs"}}{ Are the means of any trendlines different? }
  \item{\code{"hide"}}{ Don't show stats on the plot, but still compute them. }
  \item{\code{"none"}}{ Do not compute or show statistics. }
}
Statistical tests are run separately on each facet (except 
\code{"fit"}). P-values are adjusted for multiple comparisons by 
considering all facets together. Unless \code{stats = "none"}, all 
stats are attached to the plot as \code{attr(,'stats')}. 
Default: \code{"emtrends"}}

\item{p.top}{Only display taxa with the most significant differences in 
abundance. If \code{p.top} is >= 1, then the \code{p.top} most 
significant taxa are displayed. If \code{p.top} is less than one, all 
taxa with an adjusted p-value <= \code{p.top} are displayed. 
Recommended to be used in combination with the \code{taxa} parameter 
to set a lower bound on the mean abundance of considered taxa. 
Default: \code{Inf}}

\item{p.adj}{Method to use for multiple comparisons adjustment of p-values.
Run \code{p.adjust.methods} for a list of available options.
Default: \code{"fdr"}.}

\item{ci}{The confidence interval to display around the fitted curve. Set
to \code{FALSE} to hide the confidence interval. Default: \code{95}.}

\item{...}{Additional parameters to pass along to ggplot2
functions. Prefix a parameter name with either \code{t.} or 
\code{s.}/\code{pt.} to ensure it gets passed to (and only to) 
\link[ggplot2]{geom_smooth} or \link[ggplot2]{geom_point}, 
respectively. For instance, \code{s.size = 2} ensures only the 
scatterplot points have their size set to \code{2}.}
}
\value{
A \code{ggplot2} plot. The computed data points and statistics will 
        be attached as \code{attr(p, 'data')} and \code{attr(p, 'stats')}, 
        respectively.
}
\description{
Visualize taxa abundance with scatterplots and trendlines.
}
\section{Aesthetics and Partitions}{


Metadata can be used to flexibly subset, partition, and apply aesthetics 
when creating a plot. Common use cases are provided below. More thorough 
documentation is available at \url{https://cmmr.github.io/rbiom}.

\preformatted{  ## Colors ----------------------------
  color.by = "Body Site"
  color.by = list('Body Site' = "bright")
  color.by = list('Body Site' = c("Stool", "Saliva"))
  color.by = list('Body Site' = list('values' = c("Stool", "Saliva"), 'colors' = "bright"))
  color.by = list('Body Site' = c('Stool' = "blue", 'Saliva' = "green"))
  
  ## Facets ----------------------------
  facet.by = "Body Site"
  facet.by = c("Body Site", "Sex")
  facet.by = list('Body Site' = c("Stool", "Saliva"), "Sex")
  
  ## Limits ----------------------------
  limit.by = list('Sex' = "Male", 'Age' = c(20,40))
  limit.by = list('Body Site' = c("Saliva", "Anterior nares"), 'Age' = c(NA,35))
}

\itemize{
  \item{\code{color.by} - }{Any metadata column. (Max 1)}
  \item{\code{facet.by} - }{Only categorical metadata column(s).}
  \item{\code{limit.by} - }{Any metadata column(s).}
}

All built-in color palettes are colorblind-friendly.

The available categorical palette names are: \code{"okabe"}, \code{"carto"}, 
\code{"r4"}, \code{"polychrome"}, \code{"tol"}, \code{"bright"}, 
\code{"light"}, \code{"muted"}, \code{"vibrant"}, \code{"tableau"}, 
\code{"classic"}, \code{"alphabet"}, \code{"tableau20"}, \code{"kelly"}, 
and \code{"fishy"}.

The available numeric palette names are: \code{"reds"}, \code{"oranges"}, 
\code{"greens"}, \code{"purples"}, \code{"grays"}, \code{"acton"}, 
\code{"bamako"}, \code{"batlow"}, \code{"bilbao"}, \code{"buda"}, 
\code{"davos"}, \code{"devon"}, \code{"grayC"}, \code{"hawaii"}, 
\code{"imola"}, \code{"lajolla"}, \code{"lapaz"}, \code{"nuuk"}, 
\code{"oslo"}, \code{"tokyo"}, \code{"turku"}, \code{"bam"}, 
\code{"berlin"}, \code{"broc"}, \code{"cork"}, \code{"lisbon"}, 
\code{"roma"}, \code{"tofino"}, \code{"vanimo"}, \code{"vik"}
}

\examples{
    library(rbiom)
    
    taxa_corrplot(rarefy(hmp50), "BMI", color.by="Body Site") 
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()},
\code{\link{tree_plot}()}
}
\concept{plotting}
