% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_heatmap.r
\name{taxa_heatmap}
\alias{taxa_heatmap}
\title{Display taxa abundances as a heatmap.}
\usage{
taxa_heatmap(
  biom,
  rank = -1,
  taxa = 6,
  grid = list(label = "{rank} Abundance", colors = "bilbao"),
  color.by = NULL,
  order.by = NULL,
  limit.by = NULL,
  other = FALSE,
  unc = "singly",
  lineage = FALSE,
  label = TRUE,
  label_size = NULL,
  rescale = "none",
  trees = TRUE,
  clust = "complete",
  dist = "euclidean",
  tree_height = NULL,
  track_height = NULL,
  ratio = 1,
  legend = "right",
  xlab.angle = "auto",
  ...
)
}
\arguments{
\item{biom}{An \code{rbiom}-class object, or data coercible with
\code{\link[=as_rbiom]{as_rbiom()}}.}

\item{rank}{What rank(s) of taxa to display. E.g. \code{"Phylum"},
\code{"Genus"}, \code{".otu"}, etc. An integer vector can also be
given, where \code{1} is the highest rank, \code{2} is the second
highest, \code{-1} is the lowest rank, \code{-2} is the second
lowest, and \code{0} is the OTU "rank". Run \code{taxa_ranks()} to
see all options for a given \code{rbiom} object. Default: \code{-1}.}

\item{taxa}{Which taxa to display. An integer value will show the top n
most abundant taxa. A value 0 <= n < 1 will show any taxa with that
mean abundance or greater (e.g. \code{0.1} implies >= 10\%). A
character vector of taxa names will show only those named taxa.
Default: \code{6}.}

\item{grid}{Color palette name, or a list with entries for \code{label},
\code{colors}, \code{range}, \code{bins}, \code{na.color}, and/or
\code{guide}. See the Track Definitions section for details.
Default: \code{list(label = "{rank} Abundance", colors = "bilbao")}.}

\item{other}{Sum all non-itemized taxa into an "Other" taxa. When
\code{FALSE}, only returns taxa matched by the \code{taxa}
argument. Specifying \code{TRUE} adds "Other" to the returned set.
A string can also be given to imply \code{TRUE}, but with that
value as the name to use instead of "Other".
Default: \code{FALSE}}

\item{unc}{How to handle unclassified, uncultured, and similarly ambiguous
taxa names. Options are:
\itemize{
\item{\code{"singly"} - }{ Replaces them with the OTU name. }
\item{\code{"grouped"} - }{ Replaces them with a higher rank's name. }
\item{\code{"drop"} - }{ Excludes them from the result. }
\item{\code{"asis"} - }{ To not check/modify any taxa names. }
}
Default: \code{"singly"} \cr\cr
Abbreviations are allowed.}

\item{lineage}{Include all ranks in the name of the taxa. For instance,
setting to \code{TRUE} will produce
\code{Bacteria; Actinobacteria; Coriobacteriia; Coriobacteriales}.
Otherwise the taxa name will simply be \code{Coriobacteriales}. You
want to set this to TRUE when \code{unc = "asis"} and you have taxa
names (such as \emph{Incertae_Sedis}) that map to multiple higher
level ranks. Default: \code{FALSE}}

\item{label}{Label the matrix rows and columns. You can supply a list
or logical vector of length two to control row labels and column
labels separately, for example
\code{label = c(rows = TRUE, cols = FALSE)}, or simply
\code{label = c(T, F)}. Other valid options are \code{"rows"},
\code{"cols"}, \code{"both"}, \code{"bottom"}, \code{"right"},
and \code{"none"}.
Default: \code{TRUE}.}

\item{label_size}{The font size to use for the row and column labels. You
can supply a numeric vector of length two to control row label sizes
and column label sizes separately, for example
\code{c(rows = 20, cols = 8)}, or simply \code{c(20, 8)}.
Default: \code{NULL}, which computes:
\code{pmax(8, pmin(20, 100 / dim(mtx)))}.}

\item{rescale}{Rescale rows or columns to all have a common min/max.
Options: \code{"none"}, \code{"rows"}, or \code{"cols"}.
Default: \code{"none"}.}

\item{trees}{Draw a dendrogram for rows (left) and columns (top). You can
supply a list or logical vector of length two to control the row tree
and column tree separately, for example
\code{trees = c(rows = T, cols = F)}, or simply \code{trees = c(T, F)}.
Other valid options are \code{"rows"}, \code{"cols"}, \code{"both"},
\code{"left"}, \code{"top"}, and \code{"none"}.
Default: \code{TRUE}.}

\item{clust}{Clustering algorithm for reordering the rows and columns by
similarity. You can supply a list or character vector of length two to
control the row and column clustering separately, for example
\code{clust = c(rows = "complete", cols = NA)}, or simply
\code{clust = c("complete", NA)}. Options are:
\itemize{
\item{\code{FALSE} or \code{NA} - }{ Disable reordering. }
\item{An \code{hclust} class object}{ E.g. from \code{\link[stats:hclust]{stats::hclust()}}. }
\item{A method name - }{ \code{"ward.D"},
\code{"ward.D2"}, \code{"single"}, \code{"complete"},
\code{"average"}, \code{"mcquitty"}, \code{"median"}, or
\code{"centroid"}. }
}
Default: \code{"complete"}.}

\item{dist}{Distance algorithm to use when reordering the rows and columns
by similarity. You can supply a list or character vector of length
two to control the row and column clustering separately, for example
\code{dist = c(rows = "euclidean", cols = "maximum")}, or simply
\code{dist = c("euclidean", "maximum")}. Options are:
\itemize{
\item{A \code{dist} class object}{ E.g. from \code{\link[stats:dist]{stats::dist()}} or \code{\link[=bdiv_distmat]{bdiv_distmat()}}. }
\item{A method name - }{ \code{"euclidean"},
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"},
\code{"binary"}, or \code{"minkowski"}. }
}
Default: \code{"euclidean"}.}

\item{tree_height, track_height}{The height of the dendrogram or annotation
tracks in multiples (or fractions) of the smaller dimension of the
grid cell size. Use a numeric vector of length two to assign
\code{c(top, left)} independently.
Default: \code{NULL}, which computes:
\code{tree_height = sqrt(min(dim(mtx))), track_height = tree_height / 4}.}

\item{ratio}{Height/width ratio for entire grid.
Default: \code{1} (square).}

\item{legend}{Where to place the legend. Options are: \code{"right"} or
\code{"bottom"}. Default: \code{"right"}.}

\item{xlab.angle}{Angle of the labels at the bottom of the plot.
Options are \code{"auto"}, \code{'0'}, \code{'30'}, and \code{'90'}.
Default: \code{"auto"}.}

\item{...}{Additional arguments to pass on to ggplot2::theme().}
}
\value{
A \code{ggplot2} plot. \cr The computed data points, ggplot command,
and object history are available as \code{$data}, \code{$code}, and
\code{$history}, respectively.
}
\description{
Display taxa abundances as a heatmap.
}
\examples{
    library(rbiom)
    
    biom <- hmp50 \%>\% sample_rarefy() \%>\% sample_select(1:10)
    taxa_heatmap(biom, rank="Phylum", color.by="Body Site")
    
}
\seealso{
Other taxa_abundance: 
\code{\link{taxa_barplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_matrix}()},
\code{\link{taxa_stats}()},
\code{\link{taxa_table}()}

Other visualization: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{bdiv_ord_plot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_barplot}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_barplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()}
}
\concept{taxa_abundance}
\concept{visualization}
