% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_heatmap.r
\name{taxa_heatmap}
\alias{taxa_heatmap}
\title{Display taxa abundances as a heatmap.}
\usage{
taxa_heatmap(
  biom,
  rank = NULL,
  taxa = 6,
  color.by = NULL,
  order.by = NULL,
  colors = NULL,
  orders = NULL,
  ...
)
}
\arguments{
\item{biom}{A BIOM object, as returned from \link{read_biom}.}

\item{rank}{What rank(s) of taxa to display, for example \code{"Phylum"} 
or \code{"Genus"}. Run \code{taxa_ranks()} to see all options for a 
given BIOM object. The default, \code{NULL}, selects the lowest
level.}

\item{taxa}{Which taxa to give separate rows. An integer value will show
the top n most abundant taxa. A value 0 <= n < 1 will show all taxa 
with that mean abundance or greater (e.g. 0.1). A character vector of
taxon names will show only those taxa. Default: \code{6}.}

\item{color.by}{Add annotation tracks for these metadata column(s). 
Default: \code{NULL}.}

\item{order.by}{Which metadata columns to use for ordering the samples 
across the x-axis. The default setting, \code{order.by=NULL}, uses 
sample names as labels and arranges samples by similarity. See 
\link{plot_heatmap} for details.}

\item{colors}{A named \code{list()} with names matching \code{color.by}, 
each specifying a \bold{palette} as described in \link{plot_heatmap}.
For example: \code{colors = list(Sex = c(Male="blue", Female="red"))}.
Default: \code{NULL}.}

\item{orders}{Subset to just these values in the \code{order.by} metadata 
column. Default: \code{NULL}.}

\item{...}{Additional arguments to pass on to \link{plot_heatmap}.}
}
\value{
A \code{ggplot2} plot. The computed data points will be attached as 
        \code{attr(, 'data')}.
}
\description{
Display taxa abundances as a heatmap.
}
\examples{
    library(rbiom)
    
    biom <- hmp50 \%>\% rarefy() \%>\% select(1:10)
    taxa_heatmap(biom, rank="Phylum", color.by="Body Site")
    
}
\seealso{
Other plotting: 
\code{\link{adiv_boxplot}()},
\code{\link{adiv_corrplot}()},
\code{\link{bdiv_biplot}()},
\code{\link{bdiv_boxplot}()},
\code{\link{bdiv_heatmap}()},
\code{\link{depths_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{rare_corrplot}()},
\code{\link{rare_multiplot}()},
\code{\link{taxa_boxplot}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_stacked}()},
\code{\link{tree_plot}()}
}
\concept{plotting}
