% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.r
\name{taxa_ranks}
\alias{taxa_ranks}
\alias{taxa_ranks<-}
\title{Get or set the names of the taxonomic ranks.}
\usage{
taxa_ranks(biom)

taxa_ranks(x) <- value
}
\arguments{
\item{biom, x}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{value}{A character vector the new taxa rank names.}
}
\value{
A character vector of the taxa rank names in \code{biom}.
}
\description{
Get or set the names of the taxonomic ranks.
}
\examples{
    library(rbiom) 
    
    biom <- hmp50
    
    taxa_ranks(biom)
    
    taxa_ranks(biom) <- paste0("Level", seq_len(n_ranks(biom)))
    taxa_ranks(biom)
}
\seealso{
Other setters: 
\code{\link{biom_id}()},
\code{\link{otu_matrix}()},
\code{\link{otu_names}()},
\code{\link{otu_sequences}()},
\code{\link{otu_taxonomy}()},
\code{\link{otu_tree}()},
\code{\link{sample_metadata}()},
\code{\link{sample_names}()}
}
\concept{setters}
