% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_sums.r
\name{taxa_sums}
\alias{taxa_sums}
\title{Get the taxa abundances.}
\usage{
taxa_sums(biom, rank = NULL)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \code{\link[=read_biom]{read_biom()}}.}

\item{rank}{The taxonomic rank to return sums for. The default,
\code{NULL}, returns OTU sums in the same order as they appear in
\code{\link[=otu_matrix]{otu_matrix()}}. If not \code{NULL}, returned sums are ordered from
most abundant to least abundant.}
}
\value{
A numeric vector, named with the taxa names.
}
\description{
Get the taxa abundances.
}
\examples{
    library(rbiom)
    
    taxa_sums(hmp50, 'Genus') \%>\% head(10)

}
