% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_table.r
\name{taxa_sums}
\alias{taxa_sums}
\alias{taxa_means}
\alias{taxa_max}
\title{Get summary taxa abundances.}
\usage{
taxa_sums(biom, rank = 0)

taxa_means(biom, rank = 0)

taxa_max(biom, rank = 0)
}
\arguments{
\item{biom}{An \link[=rbiom_objects]{rbiom object}, such as from \code{\link[=as_rbiom]{as_rbiom()}}.
Any value accepted by \code{\link[=as_rbiom]{as_rbiom()}} can also be given here.}

\item{rank}{The taxonomic rank to return sums or means for. The default,
\code{0}, returns per-OTU summaries.}
}
\value{
A named, sorted numeric vector.
}
\description{
Get summary taxa abundances.
}
\examples{
    library(rbiom) 
    
    taxa_sums(hmp50) \%>\% head(4)
    
    taxa_means(hmp50, 'Family') \%>\% head(5)

}
\seealso{
Other taxa_abundance: 
\code{\link{taxa_boxplot}()},
\code{\link{taxa_clusters}()},
\code{\link{taxa_corrplot}()},
\code{\link{taxa_heatmap}()},
\code{\link{taxa_stacked}()},
\code{\link{taxa_stats}()},
\code{\link{taxa_table}()}
}
\concept{taxa_abundance}
