% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{taxonomy}
\alias{taxonomy}
\title{Get the taxonomy table.}
\usage{
taxonomy(biom)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}
}
\value{
A character matrix of the named taxonomies in \code{biom}.
}
\description{
Get the taxonomy table.
}
\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    
    taxonomy(biom)[1:4,]

}
\seealso{
\code{\link{sample.names}}, \code{\link{taxa.names}}, 
    \code{\link{taxa.ranks}}, \code{\link{counts}}, 
    \code{\link{taxonomy}}, \code{\link{phylogeny}},
    \code{\link{metadata}}, \code{\link{info}}

Other accessor functions: \code{\link{counts}},
  \code{\link{info}}, \code{\link{metadata}},
  \code{\link{phylogeny}}, \code{\link{sample.names}},
  \code{\link{taxa.names}}, \code{\link{taxa.ranks}}
}
