% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.r
\name{taxonomy}
\alias{taxonomy}
\title{Get the taxonomy table.}
\usage{
taxonomy(biom, rank = NULL, fix.names = FALSE)
}
\arguments{
\item{biom}{A \code{BIOM} object, as returned from \link{read.biom}.}

\item{rank}{Return just a named character vector of this rank. (Default: 
\code{NULL})}

\item{fix.names}{Clean up taxa names. See example below. (Default: 
\code{FALSE})}
}
\value{
A character matrix of the named taxonomies in \code{biom}.
        When \code{fix.names = TRUE}, a second matrix is attached as an
        attribute called 'fixed' and notes which taxa names were
        rewritten.
}
\description{
Get the taxonomy table.
}
\examples{
    library(rbiom)
    
    taxonomy(hmp50)[1:4,]
    
    # Sometimes taxonomic names are incomplete
    taxonomy(hmp50)[c(107,170,194), 'Genus', drop=FALSE]
    
    # rbiom can insert more descriptive placeholders
    taxonomy(hmp50, fix.names = TRUE)[c(107,170,194), 'Genus', drop=FALSE]

}
\seealso{
Other accessor functions: 
\code{\link{counts}()},
\code{\link{depth}()},
\code{\link{has.phylogeny}()},
\code{\link{has.sequences}()},
\code{\link{id}()},
\code{\link{info}()},
\code{\link{is.rarefied}()},
\code{\link{metadata}()},
\code{\link{nsamples}()},
\code{\link{ntaxa}()},
\code{\link{phylogeny}()},
\code{\link{sample.names}()},
\code{\link{sample.sums}()},
\code{\link{sequences}()},
\code{\link{taxa.names}()},
\code{\link{taxa.ranks}()}
}
\concept{accessor functions}
