% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.xlsx.r
\name{write.xlsx}
\alias{write.xlsx}
\title{Write data and summary information to a Microsoft Excel-compatible workbook.}
\usage{
write.xlsx(biom, outfile, depth = NULL, seed = 0)
}
\arguments{
\item{biom}{The BIOM object to save to the file.}

\item{outfile}{Path to the output xlsx file.}

\item{depth}{Depth to rarefy to. See \code{rarefy} function for details.
\code{depth = NULL} auto-selects a rarefaction level.
\code{depth = 0} disables rarefaction.
  Only use \code{depth} with \code{BIOM} files of type
  'OTU table' and integer count values.}

\item{seed}{Random seed to use in rarefying. See \code{rarefy} function
for details.}
}
\value{
On success, returns \code{NULL} invisibly.
}
\description{
Write data and summary information to a Microsoft Excel-compatible workbook.
}
\section{Note}{

Any \code{data frame}, \code{matrix}, or \code{dist} attributes on 
\code{biom} will be included as separate worksheets. An attribute named 
'Reads Per Step' is treated specially and merged with the usual 'Reads Per 
Sample' tab.
}

\examples{
    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom   <- read.biom(infile) \%>\% select(1:10) \%>\% rarefy()
    
    attr(biom, "Weighted UniFrac")   <- unifrac(biom)
    attr(biom, "Unweighted Jaccard") <- beta.div(biom, 'jaccard', weighted=FALSE)
    
    outfile <- tempfile(fileext = ".xlsx")
    write.xlsx(biom, outfile)

}
