% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.r
\name{write_counts}
\alias{write_counts}
\title{Write OTU counts to a tab-separated value (tsv) file.}
\usage{
write_counts(biom, file)
}
\arguments{
\item{biom}{An rbiom object}

\item{file}{Path to the output file. File names ending in
\code{.gz} or \code{.bz2} will be compressed accordingly.}
}
\value{
The normalized filepath that was written to (invisibly).
}
\description{
Write OTU counts to a tab-separated value (tsv) file.
}
\seealso{
Other otus: 
\code{\link{n_otus}()},
\code{\link{otu_matrix}()},
\code{\link{otu_names}()}

Other writers: 
\code{\link{write_biom}()},
\code{\link{write_fasta}()},
\code{\link{write_metadata}()},
\code{\link{write_taxonomy}()},
\code{\link{write_tree}()},
\code{\link{write_xlsx}()}
}
\concept{otus}
\concept{writers}
