% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.r
\name{write_fasta}
\alias{write_fasta}
\title{Write DNA sequences to a file in fasta format.}
\usage{
write_fasta(seqs, file = NULL)
}
\arguments{
\item{seqs}{A named character vector where names are sequence names and
values are the sequences. Also accepts an \code{rbiom} object which
contains sequences.}

\item{file}{Path to the output fasta file. Filenames ending in
\code{.gz} or \code{.bz2} will be compressed accordingly.}
}
\value{
The normalized filepath that was written to (invisibly).
}
\description{
Write DNA sequences to a file in fasta format.
}
