% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.r
\name{write_metadata}
\alias{write_metadata}
\title{Write BIOM metadata to a tab-separated value (tsv) file.}
\usage{
write_metadata(biom, file)
}
\arguments{
\item{biom}{A BIOM object}

\item{file}{Path to the output file. Filenames ending in
\code{.gz} or \code{.bz2} will be compressed accordingly.}
}
\value{
The normalized filepath that was written to (invisibly).
}
\description{
Write BIOM metadata to a tab-separated value (tsv) file.
}
