% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.r
\name{write_taxonomy}
\alias{write_taxonomy}
\title{Write BIOM taxonomy map to a tab-separated value (tsv) file.}
\usage{
write_taxonomy(biom, file)
}
\arguments{
\item{biom}{An rbiom object}

\item{file}{Path to the output file. Filenames ending in
\code{.gz} or \code{.bz2} will be compressed accordingly.}
}
\value{
The normalized filepath that was written to (invisibly).
}
\description{
Write BIOM taxonomy map to a tab-separated value (tsv) file.
}
\seealso{
Other taxonomy: 
\code{\link{n_ranks}()},
\code{\link{otu_taxonomy}()},
\code{\link{taxa_ranks}()}

Other writers: 
\code{\link{write_biom}()},
\code{\link{write_counts}()},
\code{\link{write_fasta}()},
\code{\link{write_metadata}()},
\code{\link{write_tree}()},
\code{\link{write_xlsx}()}
}
\concept{taxonomy}
\concept{writers}
