% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlsx.r
\name{write_xlsx}
\alias{write_xlsx}
\title{Write data and summary information to a Microsoft Excel-compatible workbook.}
\usage{
write_xlsx(biom, file = NULL, depth = NULL, seed = 0, unc = "asis")
}
\arguments{
\item{biom}{The \code{rbiom} object to save to the file.}

\item{file}{Path to the output xlsx file.}

\item{depth}{Depth to rarefy to. See \code{\link[=sample_rarefy]{sample_rarefy()}} function for details.
\code{depth = NULL} auto-selects a rarefaction level.
\code{depth = 0} disables rarefaction. Only use \code{depth} with
'OTU table' data and integer count values.}

\item{seed}{Random seed to use in rarefying. See \code{\link[=sample_rarefy]{sample_rarefy()}} function
for details.}

\item{unc}{How to handle unclassified, uncultured, and similarly ambiguous
taxa names. Options are:
\itemize{
\item{\code{"singly"} - }{ Replaces them with the OTU name. }
\item{\code{"grouped"} - }{ Replaces them with a higher rank's name. }
\item{\code{"drop"} - }{ Excludes them from the result. }
\item{\code{"asis"} - }{ To not check/modify any taxa names. }
}
Default: \code{"singly"} \cr\cr
Abbreviations are allowed.}
}
\value{
The normalized filepath that was written to (invisibly).
}
\description{
Write data and summary information to a Microsoft Excel-compatible workbook.
}
\section{Note}{

Any \code{data frame}, \code{matrix}, or \code{dist} attributes on
\code{biom} will be included as separate worksheets. An attribute named
'Reads Per Step' is treated specially and merged with the usual 'Reads Per
Sample' tab.
}

\examples{
    library(rbiom) 
    
    if (FALSE) {
      
      biom <- sample_select(hmp50, 1:10) \%>\% sample_rarefy()
      
      attr(biom, "Weighted UniFrac")   <- bdiv_distmat(biom, 'unifrac')
      attr(biom, "Unweighted Jaccard") <- bdiv_distmat(biom, 'jaccard', weighted=FALSE)
      
      outfile <- write_xlsx(biom, tempfile(fileext = ".xlsx"))
    }
}
\seealso{
Other biom: 
\code{\link{as_percent}()},
\code{\link{as_rbiom}()},
\code{\link{bdply}()},
\code{\link{biom_build}()},
\code{\link{biom_id}()},
\code{\link{biom_info}()},
\code{\link{biom_merge}()},
\code{\link{biom_repair}()},
\code{\link{blply}()},
\code{\link{read_biom}()},
\code{\link{write_biom}()}

Other writers: 
\code{\link{write_biom}()},
\code{\link{write_counts}()},
\code{\link{write_fasta}()},
\code{\link{write_metadata}()},
\code{\link{write_taxonomy}()},
\code{\link{write_tree}()}
}
\concept{biom}
\concept{writers}
